% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{ad_ref_alt_from_gds}
\alias{ad_ref_alt_from_gds}
\title{ad_ref_alt_from_gds}
\usage{
ad_ref_alt_from_gds(gds)
}
\description{
Extract REF and ALT read-depth matrices from the AD field stored in GDS.
}
\keyword{internal}
