% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcftools_helpers.R
\name{bcftools_require}
\alias{bcftools_require}
\title{Check that bcftools is available}
\usage{
bcftools_require(bcftools.path = "bcftools")
}
\arguments{
\item{bcftools.path}{(character)
Path or name of the \code{bcftools} executable.
Default: \code{bcftools.path = "bcftools"}.}
}
\description{
Internal helper that checks whether \code{bcftools} can be executed.
It runs \code{bcftools.path --version} and errors if the command fails.
}
\keyword{internal}
