% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{boxplot_stats}
\alias{boxplot_stats}
\title{boxplot_stats}
\usage{
boxplot_stats(
  data,
  title,
  subtitle = NULL,
  x.axis.title = NULL,
  y.axis.title,
  flip = FALSE,
  scale.log = FALSE,
  scientific.format = FALSE,
  facet.columns = FALSE,
  facet.rows = FALSE,
  bp.filename = NULL,
  path.folder = NULL
)
}
\description{
Generate a boxplot
}
\keyword{internal}
