% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_dots.R
\name{check_args_class}
\alias{check_args_class}
\title{check_args_class}
\usage{
check_args_class(x)
}
\description{
Check the class of the argument/parameter value and return a
  compact printable representation (value or class name).
}
\keyword{internal}
