% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_header_source}
\alias{check_header_source}
\title{Check the vcf header and detect vcf source}
\usage{
check_header_source(data)
}
\description{
Check the vcf header and detect vcf source
}
\keyword{internal}
