% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{choose_markers_subsample_prop}
\alias{choose_markers_subsample_prop}
\title{choose markers subsample prop}
\usage{
choose_markers_subsample_prop(
  n.markers,
  n.ind,
  vector.size.limit = 2^31,
  max.mem.gb = 10,
  bytes.per.cell = 4,
  n.matrices = 1,
  small.vcf.cutoff = 2e+05,
  cap.prop = 0.3,
  allowed = c(1, 0.3, 0.2, 0.1, 0.05, 0.02, 0.01),
  safety.margin = 0.9
)
}
\description{
Choose the best subsampling prop based on different metrics
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
