% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{clean_dart_colnames}
\alias{clean_dart_colnames}
\title{clean_dart_colnames}
\usage{
clean_dart_colnames(data, strata)
}
\description{
clean_dart_colnames (only the DArT columns = snakecase...)
}
\keyword{internal}
