% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_genomic_data.R
\name{clean_gl}
\alias{clean_gl}
\title{clean_gl}
\usage{
clean_gl(x, split.vec, parallel.core = parallel::detectCores() - 1)
}
\description{
Clean GL column.
GL column: cleaning Genotype Likelihood column
}
\keyword{internal}
