% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_markers_names}
\alias{clean_markers_names}
\title{Clean marker's name}
\usage{
clean_markers_names(x)
}
\arguments{
\item{x}{(character string) Markers character string.}
}
\description{
Function to clean marker's name
of weird separators
that interfere with some packages
or codes. \code{/}, \code{:}, \code{-} and \code{.} are changed to an underscore
\code{_}.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
