% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{clean_nr}
\alias{clean_nr}
\title{clean_nr}
\usage{
clean_nr(x, split.vec, parallel.core = parallel::detectCores() - 1)
}
\description{
Cleaning NR: Number of reads covering variant location in
this sample
}
\keyword{internal}
