% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{clean_nv}
\alias{clean_nv}
\title{clean_nv}
\usage{
clean_nv(x, split.vec, parallel.core = parallel::detectCores() - 1)
}
\description{
Cleaning NV: Number of reads containing variant in this sample
}
\keyword{internal}
