% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_genomic_data.R
\name{clean_pl}
\alias{clean_pl}
\title{clean_pl}
\usage{
clean_pl(x, split.vec, parallel.core = parallel::detectCores() - 1)
}
\description{
Clean PL column.
PL column (normalized, phred-scaled likelihoods for genotypes):
separating into PROB_HOM_REF, PROB_HET and PROB_HOM_ALT")
Value 1: probability that the site is homozgyous REF
Value 2: probability that the sample is heterzygous
Value 2: probability that it is homozygous ALT
}
\keyword{internal}
