% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mac.R
\name{compute_mac}
\alias{compute_mac}
\title{compute_mac}
\usage{
compute_mac(
  data,
  blacklist.markers = NULL,
  markers.meta = NULL,
  parallel.core = parallel::detectCores() - 1
)
}
\description{
Compute global MAC (Minor Allele Count)
}
\keyword{internal}
