% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_monomorphic.R
\name{count_monomorphic}
\alias{count_monomorphic}
\title{count_monomorphic}
\usage{
count_monomorphic(x, parallel.core = parallel::detectCores() - 1)
}
\description{
count monomorphe in gds
}
\keyword{internal}
