% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{dart_calibration}
\alias{dart_calibration}
\title{dart_calibration}
\usage{
dart_calibration(switch, genotypes = NULL, metadata = NULL)
}
\description{
Calibrate REF and ALT in DArT imported genotypes or metadata
}
\keyword{internal}
