% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_imputations_module.R
\name{decoding_haplotypes}
\alias{decoding_haplotypes}
\title{decoding_haplotypes}
\usage{
decoding_haplotypes(data = NULL, parallel.core = parallel::detectCores() -
  1)
}
\description{
separate snp group merged with encoding_snp
}
\keyword{internal}
