% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_imputations_module.R
\name{defactorize_gt}
\alias{defactorize_gt}
\title{defactorize_gt}
\usage{
defactorize_gt(data.to.change, data.with.info)
}
\description{
Function to "defactorize/decode" the imputed data back to original.
}
\keyword{internal}
