% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_biallelic_markers.R
\name{detect_biallelic_markers}
\alias{detect_biallelic_markers}
\title{Detect biallelic data}
\usage{
detect_biallelic_markers(
  data,
  verbose = FALSE,
  parallel.core = parallel::detectCores() - 1
)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty
during execution.
Default: \code{verbose = FALSE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
A logical character string (TRUE/FALSE). That answer the question if
the data set is biallelic or not.
}
\description{
Detect if markers in tidy dataset or radiator gds are biallelic.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
