% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_mixed_genomes.R
\name{detect_mixed_genomes}
\alias{detect_mixed_genomes}
\title{Detect mixed genomes}
\usage{
detect_mixed_genomes(data, ind.heterozygosity.threshold = NULL)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{ind.heterozygosity.threshold}{(string, double, optional)
Blacklist individuals based on observed heterozygosity (averaged across markers).


The string contains 2 thresholds values (min and max). 
The values are proportions (0 to 1), where 0 turns off the min threshold and
1 turns off the max threshold.
Individuals with mean observed heterozygosity higher (>) or lower (<) 
than the thresholds will be blacklisted. 

Default: \code{ind.heterozygosity.threshold = NULL} will turn off completely 
the filter and the function will only output the plots and table of heterozygosity.}
}
\value{
The function returns inside the global environment a list with
5 objects:

\enumerate{
\item the individual's heterozigosity (\code{$individual.heterozygosity})
a dataframe containing for each individual, the population id, the number of 
genotyped markers, the number of missing genotypes (based on the number of 
markers of the population and overall), the number of markers genotyped as heterozygote
and it's proportion based on the number of genotyped markers.
\item the heterozygosity statistics per populations and overall:\code{$heterozygosity.statistics}
\item the blacklisted individuals if \code{ind.heterozygosity.threshold} was selected: \code{$blacklist.ind.het}
\item the boxplot of individual heterozygosity:\code{$individual.heterozygosity.boxplot}
\item the manhattan plot of individual heterozygosity (\code{$individual.heterozygosity.manhattan.plot})
contrasted with missingness proportion based on the number of markers (population or overall).
The 2 facets will be identical when the dataset as common markers between
the populations. The dotted lines are the mean hetegozygosities.
}
}
\description{
Highlight outliers individual's observed heterozygosity for a quick
diagnostic of mixed samples or poor polymorphism discovery due to DNA quality,
sequencing effort, etc.
}
\details{
To help discard an individual based on his observed heterozygosity
(averaged across markers),
use the manhanttan plot to:
\enumerate{
\item contrast the individual with population and overall samples.
\item visualize the impact of missigness information (based on population or 
overall number of markers) and the individual observed heterozygosity. The
larger the point, the more missing genotypes.
}
\strong{Outlier above average:}
\itemize{
\item potentially represent two samples mixed together (action: blacklist), or...
\item a sample with more sequecing effort (point size small): did you merge your replicates fq files ? (action : keep and monitor)
\item a sample with poor sequencing effort (point size large) where the genotyped markers are
all heterozygotes, verify this with missingness (action: discard)
}
In all cases, if there is no bias in the population sequencing effort,
the size of the point will usually be "average" based on the population or
overall number of markers.


You can visualize individual observed heterozygosity, choose thresholds and
then visualize, choose thresholds and filter markers based on observed 
heterozygosity in one run with: \pkg{radiator} \code{\link{filter_het}}.

\strong{Outlier below average:}
\itemize{
\item A point with a size larger than the population or overall average (= lots of missing): 
the poor polymorphism discovery of the sample is probably the result of bad
DNA quality, a bias in sequencing effort, etc. (action: blacklist)
\item A point with a size that looks average (not much missing): 
this sample requires more attention (action: blacklist) and conduct more tests.
e.g. for biallelic data, look for coverage imbalance between ALT/REF allele.
At this point you need to distinguish between an artifact of poor polymorphism discovery 
or a biological reason (highly inbred individual, etc.).
}
}
\examples{
\dontrun{
#Step1: highlight outlier individuals, the simplest way to run:

outlier.ind.het <- radiator::detect_mixed_genomes(data = "wombat_tidy.tsv")

# Or if you don't have a tidy df:

outlier.ind.het <- radiator::tidy_genomic_data(
data = "wombat.vcf",
strata = "strata.wombat.tsv",
common.markers = FALSE,
vcf.metadata = FALSE,
verbose = TRUE
) \%>\% 
radiator::detect_mixed_genomes(.)


#This example, without threshold, will not produce a blacklist of individuals.

#To look at the table with individual's heterozygosity:

outlier.ind.het$individual.heterozygosity

# To view the manhattan plot:

outlier.ind.het$individual.heterozygosity.manhattan.plot

# To view the box plot
outlier.ind.het$individual.heterozygosity.boxplot

# To save the boxplot:

ggsave(
"individual.heterozygosity.boxplot.pdf",
width = 15, height = 10, dpi = 600, units = "cm",
useDingbats = FALSE
)

# prefer a PNG:

ggsave(
"individual.heterozygosity.boxplot.png",
width = 15, height = 10, dpi = 300, units = "cm"
)

#Step2: blacklist ind with min and max Het obs thresholds
# Based on the look of the distribution using both jitter and boxplot,
# choose a threshold to blacklist the outliers and re-run the function.

outlier.ind.het <- radiator::detect_mixed_genomes(
data = "wombat_tidy.tsv", ind.heterozygosity.threshold = c(0.02, 0.2)
)

blacklist.ind.het <- outlier.ind.het$blacklist.ind.het

# To keep the blacklist:
readr::write_tsv(
x = blacklist.ind.het,
path = "blacklist.individuals.heterozygosity.tsv", col_names = TRUE
)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com} and Peter Grewe \email{peter.grewe@csiro.au}
}
