% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_deprecated_functions.R
\name{discard_monomorphic_markers}
\alias{discard_monomorphic_markers}
\title{Discard monomorphic markers}
\usage{
discard_monomorphic_markers(data, verbose = FALSE)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty
during execution.
Default: \code{verbose = FALSE}.}
}
\value{
A list with the filtered input file and the blacklist of markers removed.
}
\description{
Discard monomorphic markers.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
