% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distance2tibble}
\alias{distance2tibble}
\title{distance2tibble}
\usage{
distance2tibble(
  x,
  remove.diag = TRUE,
  na.diag = FALSE,
  remove.lower = TRUE,
  relative = TRUE,
  pop.levels = NULL,
  distance.class.double = TRUE
)
}
\description{
melt the distance matrix into a tibble
}
\keyword{internal}
