% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duplicate_genomes.R
\name{distance_individuals}
\alias{distance_individuals}
\title{Distance individuals}
\usage{
distance_individuals(
  x,
  strata = NULL,
  distance.method = "manhattan",
  parallel.core = parallel::detectCores() - 1
)
}
\description{
distance method
}
\keyword{internal}
