% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{dots_deprecated_read_vcf}
\alias{dots_deprecated_read_vcf}
\title{Deprecated dot-arguments for read_vcf}
\usage{
dots_deprecated_read_vcf(extra = NULL, exclude = NULL)
}
\description{
Return the set of deprecated `...` arguments understood by
\code{read_vcf()}. Currently this is just the global deprecated set,
but the helper allows you to add read_vcf-specific deprecated args
later if needed.
}
\keyword{internal}
