% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_dots.R
\name{dots_keepers_core}
\alias{dots_keepers_core}
\title{Global dot-argument "keepers" for radiator}
\usage{
dots_keepers_core()
}
\description{
Internal helper returning the core set of supported `...` arguments
used across most radiator functions.
}
\keyword{internal}
