% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase_genotypes.R
\name{erase_genotypes}
\alias{erase_genotypes}
\title{Erase genotypes}
\usage{
erase_genotypes(
  tidy.vcf.file,
  haplotypes.file,
  read.depth.threshold,
  allele.depth.threshold,
  allele.imbalance.threshold,
  filename
)
}
\arguments{
\item{tidy.vcf.file}{A data frame object or file (using ".tsv") of a tidy vcf.}

\item{haplotypes.file}{(optional) The 'batch_x.haplotypes.tsv'. If you want to
erase the genotypes that don't pass the threshold.}

\item{read.depth.threshold}{(integer) Threshold for the read depth.}

\item{allele.depth.threshold}{(integer) Threhold for the min depth
of REF or ALT alleles.}

\item{allele.imbalance.threshold}{(numeric) Threshold of ratio between ALT and REF
depth of coverage. See details below.}

\item{filename}{(optional) Name of the file written to the working directory.}
}
\value{
The function returns the blacklisted individuals genotypes,
by loci, position (SNP, POS in stacks), populations and individuals.
For VCF, return the tidy vcf in the global environment only and in
the directory with \code{filename}. For haplotype file the original
filename with "_erased_geno" is appended.
}
\description{
This function uses the information in the vcf tidy to i) make a
blacklist of individual genotypes to erase based on coverage and
genotype likelihood thresholds and ii) erase those blacklisted genotypes from
 a tidy vcf file and a STACKS haplotypes file.
}
\details{
Genotypes below average quality i.e. below threshold for the
coverage of REF and/or ALT allele and genotype likelihood
are zeroed from the file. The function erase SNP in the VCF file and loci
in the haplotypes file.
Also creates a blacklist of genotypes erased based on the genotype
likelihood threshold, the REF and ALT allele coverage threshold.
The ratio is calculated :
(read depth ALT allele - read depth REF allele)/(read depth ALT allele + read
depth REF allele).
e.g. REF = 3 and ALT = 2 the ratio = -0.2. For the function to work properly,
use positive values, the function will calculate the +/- imbalance.
}
