% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_het_outliers.R
\name{estimate_m}
\alias{estimate_m}
\title{heterozygotes miscall rate}
\usage{
estimate_m(
  data,
  nreps = 200,
  m_init = stats::runif(1),
  a0 = 0.5,
  a1 = 0.5,
  sm = 0.005
)
}
\arguments{
\item{data}{A tidy data frame (biallelic only).}

\item{nreps}{number of MCMC sweeps to do.
Default: \code{nreps = 200}.}

\item{m_init}{initial starting value for m must be between 0 and 1}

\item{a0}{beta parameter for reference alleles}

\item{a1}{beta parameter for alternate alleles}

\item{sm}{standard devation of proposal distribution for m}
}
\description{
estimate the heterozygotes miscall rate from a simple model
}
\author{
Eric Anderson \email{eric.anderson@noaa.gov}
}
\keyword{internal}
