% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_seqarray.R
\name{extract_coverage}
\alias{extract_coverage}
\title{extract_coverage}
\usage{
extract_coverage(data = NULL, variant.locus.id = NULL)
}
\description{
Extract coverage information from a GDS file
}
\keyword{internal}
