% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{extract_dart_markers_metadata}
\alias{extract_dart_markers_metadata}
\title{extract_dart_markers_metadata}
\usage{
extract_dart_markers_metadata(
  data,
  whitelist.markers = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE
)
}
\description{
Extract DArT markers/locus metadata.
}
\keyword{internal}
