% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{extract_genotypes_metadata}
\alias{extract_genotypes_metadata}
\title{extract_genotypes_metadata}
\usage{
extract_genotypes_metadata(gds, genotypes.meta.select = NULL,
  radiator.node = TRUE, index.only = FALSE,
  sync.markers.individuals = TRUE, whitelist = FALSE,
  blacklist = FALSE, verbose = FALSE)
}
\description{
Import gds or radiator genotypes meta node
}
\keyword{internal}
