% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{extract_individuals}
\alias{extract_individuals}
\title{extract_individuals}
\usage{
extract_individuals(gds, ind.field.select = NULL, radiator.node = TRUE,
  whitelist = FALSE, blacklist = FALSE, verbose = FALSE)
}
\description{
Import gds or radiator individuals node
}
\keyword{internal}
