% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{extract_individuals_metadata}
\alias{extract_individuals_metadata}
\title{extract_individuals_metadata}
\usage{
extract_individuals_metadata(gds, ind.field.select = NULL,
  radiator.node = TRUE, whitelist = FALSE, blacklist = FALSE,
  verbose = FALSE)
}
\description{
Import gds or radiator individuals node
}
\keyword{internal}
