% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{extract_individuals_vcf}
\alias{extract_individuals_vcf}
\title{Extract individuals from vcf file}
\usage{
extract_individuals_vcf(data)
}
\arguments{
\item{data}{(character) The path to the vcf file.}
}
\value{
A tibble with a column: \code{INDIVIDUALS}.
}
\description{
Function that returns the individuals present in a vcf file.
Useful to create a strata file or
to make sure you have the right individuals in your VCF.
}
\seealso{
\pkg{radiator} \code{\link{read_strata}}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
