% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vcf_individuals.R
\name{extract_vcf_individuals}
\alias{extract_vcf_individuals}
\title{extract_vcf_individuals}
\usage{
extract_vcf_individuals(data)
}
\arguments{
\item{data}{The vcf file.}
}
\value{
A tibble with a column: \code{INDIVIDUALS}.
}
\description{
Function that returns the individuals presente in a vcf file.
then estimate allele frequencies, and the genotyping error rate.
Useful to create a strata file to work with function using vcf and that require
a strata file. This will ensure that you have the same individuals in your vcf file and strata...
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
