% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_blacklist_genotypes.R
\name{filter_blacklist_genotypes}
\alias{filter_blacklist_genotypes}
\title{Filter dataset with blacklist of genotypes}
\usage{
filter_blacklist_genotypes(data, blacklist.genotypes, verbose = TRUE, ...)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{blacklist.genotypes}{(path or object)
The blacklist is an object in your
global environment or a file in the working directory (e.g. "blacklist.geno.tsv").
The dataframe contains at least these 2 columns: \code{MARKERS, INDIVIDUALS}.
Additional columns are allowed: \code{CHROM, LOCUS, POS}.

Useful to erase genotypes with bad QC, e.g. genotype with more than 2 alleles
in diploid likely
sequencing errors or genotypes with poor genotype likelihood or coverage.

Columns are cleaned of separators that interfere with some packages or codes, detailed in
\code{\link{clean_markers_names}} and \code{\link{clean_ind_names}}
Default \code{blacklist.genotypes = NULL}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function. Also used for legacy arguments (see details or
special section)}
}
\description{
Filter dataset with blacklist of genotypes.

This function allows to blacklist/erase/mask genotypes.

Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\section{Life cycle}{


This function arguments will be subject to changes. Currently the function uses
erase.genotypes, but using the \code{dots-dots-dots ...} arguments allows to
pass \code{erase.genotypes and masked.genotypes}. These arguments do exactly
the same thing and only one can be used.
}

\examples{
\dontrun{
data <- radiator::filter_blacklist_genotypes(
    data = data, blacklist.geno = "blacklist.geno.tsv"
    )
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
