% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dart.R
\name{filter_dart}
\alias{filter_dart}
\title{filter DArT file}
\usage{
filter_dart(...)
}
\description{
filter DArT file
}
\section{Life cycle}{


As of radiator v.0.0.22, \code{filter_dart} is now replaced by
\code{\link{filter_rad}} in an effort to have more meaningful functions names.
This is THE function to rule them all.
}

