% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dart_reproducibility.R
\name{filter_dart_reproducibility}
\alias{filter_dart_reproducibility}
\title{Filter data based on DArT reproducibility statistics}
\usage{
filter_dart_reproducibility(interactive.filter = TRUE, data,
  filter.reproducibility = NULL,
  parallel.core = parallel::detectCores() - 1, verbose = TRUE, ...)
}
\arguments{
\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. Figures of distribution are shown before asking for filtering
thresholds.
Default: \code{interactive.filter = TRUE}.}

\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{write_seqarray}} or
\code{\link{tidy_vcf}}.}

\item{filter.reproducibility}{(double, character) This is best decided after viewing the figures.
Usually values higher than 0.95 are not uncommon.
The value can also be character: \code{filter.reproducibility = "outliers"}.
Using this, will remove outlier markers using the lower outlier statistics.
Default: \code{filter.reproducibility = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function. Also used for legacy arguments (see details or
special section)}
}
\value{
A list in the global environment with 6 objects:
\enumerate{
\item $whitelist.markers
\item $blacklist.markers
\item $filters.parameters
}

The object can be isolated in separate object outside the list by
following the example below.
}
\description{
This filter removes markers below a certain threshold.
Based on the repoducibility column found in DArT files.

\strong{Filter targets}: Markers

\strong{Statistics}: Reproducibility (established by DArT)
}
\details{
\strong{Interactive version}

There are 2 steps in the interactive version to visualize and filter
the data based on the reproducibility value:

Step 1. Visualization using a box plot

Step 2. Choose the filtering threshold
}
\examples{
\dontrun{
spotted.cod <- radiator::tidy_dart(
    data = "Combined_1514and1614_SNP_80Callrate.csv",
    strata = "strata.dart.spotted.cod.tsv"
)
turtle.filtered <- radiator::filter_dart_reproducibility(
data = spotted.cod
filter.reproducibility = 0.97
)
}
}
