% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fis.R
\name{filter_fis}
\alias{filter_fis}
\title{Fis filter}
\usage{
filter_fis(data, approach = "haplotype", fis.min.threshold, fis.max.threshold,
  fis.diff.threshold, pop.threshold, percent, filename)
}
\arguments{
\item{data}{A data frame object or file (using ".tsv")
of class sumstats or a tidy VCF summarised.}

\item{approach}{Character. By \code{"SNP"} or by \code{"haplotype"}. 
The function will consider the SNP or haplotype MAF statistics to filter the marker. 
Default by \code{"haplotype"}.}

\item{fis.min.threshold}{Number}

\item{fis.max.threshold}{Number.}

\item{fis.diff.threshold}{Number (0 - 1)}

\item{pop.threshold}{Fixed number of pop required to keep the locus.}

\item{percent}{Is the threshold a percentage ? TRUE or FALSE.}

\item{filename}{Name of the file written to the working directory (optional).}
}
\description{
or a tidy VCF file.
}
