% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genotype_likelihood.R
\name{filter_genotype_likelihood}
\alias{filter_genotype_likelihood}
\title{Genotype likelihood interactive filter}
\usage{
filter_genotype_likelihood(data, interactive.filter = TRUE, strata = NULL,
  gl.approach = "haplotype", gl.ind.threshold = NULL,
  gl.mean.threshold = NULL, gl.min.threshold = NULL,
  gl.diff.threshold = NULL, pop.threshold = NULL, percent = NULL,
  pop.levels = NULL, pop.labels = NULL, pop.select = NULL,
  blacklist.id = NULL, blacklist.genotype = NULL,
  whitelist.markers = NULL, monomorphic.out = TRUE, max.marker = NULL,
  snp.ld = NULL, common.markers = FALSE, filename = NULL)
}
\arguments{
\item{data}{11 options: VCF (SNPs or Haplotypes,
to make the vcf population ready, see details below),
plink, stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop,
and a data frame in long/tidy or wide format.
\emph{See details} of \code{\link{tidy_genomic_data}}.}

\item{interactive.filter}{(optional, logical) Do you want the filtering session to 
be interactive. With default: \code{interactive.filter == TRUE}, the user is 
asked to see figures of distribution before making decisions for filtering 
with the genotype likelihood.}

\item{strata}{(optional/required) Required for VCF and haplotypes files,
optional for the other formats supported.

The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}.
If a \code{strata} file is specified with all file formats that don't
require it, the strata argument will have precedence on the population
groupings used internally in those file formats.
The \code{STRATA} column can be any hierarchical grouping.
To create a strata file see \code{\link[radiator]{individuals2strata}}.
If you have already run
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data,
the strata file is similar to a stacks \emph{population map file},
make sure you
have the required column names (\code{INDIVIDUALS} and \code{STRATA}).
Default: \code{strata = NULL}.}

\item{gl.approach}{Character. By \code{gl.approach = "SNP"} or 
by \code{gl.approach = "haplotype"}. 
The function will consider the SNP or haplotype GL statistics to filter the marker. 
Default: \code{gl.approach = "haplotype"}.}

\item{gl.ind.threshold}{(Integer, optional if interactive session) 
Threshold number of individual's genotype likelihood.}

\item{gl.mean.threshold}{(Integer, optional if interactive session) 
Threshold number of mean genotype likelihood.}

\item{gl.min.threshold}{(Integer, optional if interactive session) 
Threshold number of min genotype likelihood.}

\item{gl.diff.threshold}{(Integer, optional if interactive session)
Threshold number of diff genotype likelihood,
the difference between the max and min GL over a loci/read/haplotype.}

\item{pop.threshold}{(optional if interactive session)
A threshold number: proportion, percentage
or fixed number e.g. 0.50, 50 or 5.}

\item{percent}{(logical, optional if interactive session). 
Is the pop.threshold argument a percentage? TRUE/FALSE.
This argument is necessary to distinguish percentage from integer 
for population threshold, (e.g. 5 percent or 5 populations).}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
White spaces in population names are replaced by underscore.
Default: \code{pop.levels = NULL}.}

\item{pop.labels}{(optional, string) Use this argument to rename/relabel
your pop or combine your pop. e.g. To combine \code{"QUE"} and \code{"ONT"}
into a new pop called \code{"NEW"}:
(1) First, define the levels for your pop with \code{pop.levels} argument:
\code{pop.levels = c("QUE", "ONT", "ALB")}.
(2) then, use \code{pop.labels} argument:
\code{pop.levels = c("NEW", "NEW", "ALB")}.
To rename \code{"QUE"} to \code{"TAS"}:
\code{pop.labels = c("TAS", "ONT", "ALB")}.
Default: \code{pop.labels = NULL}. If you find this too complicated,
there is also the \code{strata} argument that can do the same thing,
see below.
White spaces in population names are replaced by underscore.}

\item{pop.select}{(string, optional) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops).
Default: \code{pop.select = NULL}}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is an object in your
global environment or a file in the working directory
(e.g. "blacklist.txt").
Default: \code{blacklist.id = NULL}.}

\item{blacklist.genotype}{(optional) Useful to erase genotype with below
average quality, e.g. genotype with more than 2 alleles in diploid likely
sequencing errors or genotypes with poor genotype likelihood or coverage.
The blacklist has a minimum of 2 column headers (markers and individuals).
Markers can be 1 column (CHROM or LOCUS or POS),
a combination of 2 (e.g. CHROM and POS or CHROM and LOCUS or LOCUS and POS) or
all 3 (CHROM, LOCUS, POS). The markers columns must be designated: CHROM (character
or integer) and/or LOCUS (integer) and/or POS (integer). The id column designated
INDIVIDUALS (character) columns header. The blacklist must be in the working
directory (e.g. "blacklist.genotype.txt"). For de novo VCF, CHROM column
with 'un' need to be changed to 1.
Default: \code{blacklist.genotype = NULL} for no blacklist of
genotypes to erase.}

\item{whitelist.markers}{(optional) A whitelist containing CHROM (character
or integer) and/or LOCUS (integer) and/or
POS (integer) columns header. To filter by chromosome and/or locus and/or by snp.
The whitelist is in the working directory (e.g. "whitelist.txt").
de novo CHROM column with 'un' need to be changed to 1.
In the VCF, the column ID is the LOCUS identification.
Default \code{whitelist.markers = NULL} for no whitelist of markers.}

\item{monomorphic.out}{(optional) Should the monomorphic
markers present in the dataset be filtered out ?
Default: \code{monomorphic.out = TRUE}.}

\item{max.marker}{(integer, optional) For large PLINK files,
useful to subsample marker number. e.g. if the data set
contains 200 000 markers and \code{max.marker = 10000}, 10000 markers are
subsampled randomly from the 200000 markers. If you need specific markers,
use \code{whitelist.markers} argument.
Default: \code{max.marker = NULL}.}

\item{snp.ld}{(optional) \strong{For VCF file only}.
SNP short distance linkage disequilibrium pruning. With anonymous markers from
RADseq/GBS de novo discovery, you can minimize linkage disequilibrium (LD) by
choosing among these 3 options: \code{"random"} selection, \code{"first"} or
\code{"last"} SNP on the same short read/haplotype. For long distance linkage
disequilibrium pruning, see details below.
Default: \code{snp.ld = NULL}.}

\item{common.markers}{(optional) Logical. Default: \code{common.markers = TRUE},
will only keep markers in common (genotyped) between all the populations.}

\item{filename}{(optional) Name of the filtered data set, 
written to the working directory.}
}
\value{
With \code{interactive.filter = FALSE}, the function returns a 
filtered tidy vcf data frame inside the global environment. 
In the working directory 5 files are written:
1. filters_parameters.tsv is updated or created.
The filter parameters, values and returned results 
(markers numbers or blacklist genotypes) are inside that file. 
2. blacklist.genotypes.gl.tsv, 3. blacklist.markers.gl.tsv, 
4. whitelist.markers.gl.tsv and 
5. vcf.tidy.paralogs.id.gl.tsv (the same filtered tidy vcf data frame).

With \code{interactive.filter = TRUE}, a list with 20 objects is created. 
The information range from summary tables to plots. The objects names are 
found by using \code{names(list.name)}. The object can be isolated in separate
object outside the list by following the example below.
}
\description{
Filter genotypes and markers based on genotype likelihood using
a tidy data set (long format) and any VCF file both with GL information.
The function uses 
\code{\link[radiator]{tidy_genomic_data}} and 
\code{\link[radiator]{tidy_wide}} to load the file.
}
\details{
There is 4 steps in the interactive version:  

1. gl_individuals_populations: the user is asked to inspect 
the genotype likelihood at the individuals and populations levels"), 

2. blacklist_genotypes: the option is given to blacklist individual genotypes
based on low quality GL, 

3. gl_markers: the user is asked to inspecting the genotype likelihood 
at the marker level and 

4. blacklist_markers: the option is given to blacklist markers genotypes 
based on low quality GL found at the marker level.

Using the haplotype approach: The summary statistics are averaged
and nested: SNP -> individuals -> population -> loci. e.g. the mean GL is the average
genotype likelihood for all individuals of pop x for loci x.
The gl.diff.threshold is the difference between the max and min GL found for 
a locus. This argument is given to help your spot big difference in GL for 
a loci. Markers with small differences have more stability in the estimate.
}
\examples{
\dontrun{
If I have 10 populations, the 3 examples below will give the same output results:

beluga.vcf.tidy.gl <- filter_genotype_likelihood(
data = "beluga.vcf", 
interactive.filter = TRUE,
gl.approach = "haplotype", 
gl.ind.threshold = 0,
gl.mean.threshold = 10, 
gl.min.threshold = 0, 
gl.diff.threshold = 200, 
pop.threshold = 50, 
percent = TRUE)

beluga.vcf.tidy.gl <- filter_genotype_likelihood(
data = "beluga.vcf", 
interactive.filter = FALSE,
gl.approach = "haplotype",
gl.ind.threshold = 0,
gl.mean.threshold = 10, 
gl.min.threshold = 0, 
gl.diff.threshold = 200, 
pop.threshold = 0.5, 
percent = FALSE)

beluga.vcf.tidy.gl <- filter_genotype_likelihood(
data = "beluga.vcf", 
gl.approach = "haplotype", 
gl.ind.threshold = 0,
gl.mean.threshold = 10, 
gl.min.threshold = 0, 
gl.diff.threshold = 100, 
pop.threshold = 5, 
percent = FALSE)

If interactive.filter = TRUE, a list is created and to view the filtered tidy vcf:
tidy.data <- beluga.vcf.tidy.gl$tidy.vcf.filtered.gl

Inside the same list, to isolate the blacklist.genotypes:
bg <- beluga.vcf.tidy.gl$blacklist.genotypes
}
}
