% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_het.R
\name{filter_het}
\alias{filter_het}
\title{Heterozygosity filter}
\usage{
filter_het(data, interactive.filter = TRUE,
  ind.heterozygosity.threshold = NULL, het.approach = c("SNP", "overall"),
  het.threshold = 1, het.dif.threshold = 1, outlier.pop.threshold = 1,
  helper.tables = FALSE, coverage.info = FALSE, filename = NULL,
  vcf.metadata = TRUE, blacklist.id = NULL, blacklist.genotype = NULL,
  whitelist.markers = NULL, monomorphic.out = TRUE, max.marker = NULL,
  snp.ld = NULL, common.markers = FALSE, strata = NULL,
  pop.levels = NULL, pop.labels = NULL, pop.select = NULL,
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{data}{12 options: VCF (SNPs or Haplotypes,
to make the vcf population ready, see details below),
plink, stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop, DArT,
and a data frame in long/tidy or wide format.
\emph{See details} of \code{\link[radiator]{tidy_genomic_data}}.}

\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. With the default, the user is asked to see figures of
distribution before making decisions for filtering with heterozygosity statistics.
Default: \code{interactive.filter == TRUE}.}

\item{ind.heterozygosity.threshold}{(string, double, optional)
Blacklist individuals based on observed heterozygosity (averaged across markers).


The string contains 2 thresholds values (min and max).
The values are proportions (0 to 1), where 0 turns off the min threshold and
1 turns off the max threshold.
Individuals with mean observed heterozygosity higher (>) or lower (<)
than the thresholds will be blacklisted.

Default: \code{ind.heterozygosity.threshold = NULL} will turn off completely
the filter and the function will only output the plots and table of heterozygosity.}

\item{het.approach}{(Character string). First value, \code{"SNP"} or
\code{"haplotype"}. The haplotype approach considers the statistic
consistency on the read (locus/haplotype). The major difference:
the haplotype approach results in blacklisting the entire locus/haplotype
with all the SNPs on the read.
With the SNP approach, SNPs are independently analyzed and blacklisted.
The second value, will use the statistics by population \code{"pop"} or
will consider the data overall, as 1 large group \code{"overall"}.
Default: \code{het.approach = c("SNP", "overall")}.}

\item{het.threshold}{Number Biallelic markers usually max 0.5. But departure
from this value is common. The higher the proportion threshold, the more relaxed
the filter is. With default there is no filtering.
Default: \code{het.threshold = 1}.}

\item{het.dif.threshold}{Number (0 - 1). For \code{het.approach = "haplotype"} only.
You can set a threshold for the difference in het along your read.
Set the number your willing to tolerate on the same read/haplotype.
e.g. if you have 2 SNP on a read/haplotype and on as a het of 0.9 and the other
0.1 and you set \code{het.dif.threshold = 0.3}, this markers will be blacklisted.
You should strive to have similar statistics along short read like RAD.
The higher the proportion threshold, the more relaxed the filter is.
With default, there is no filtering and all the range of differences are allowed.
Default: \code{het.dif.threshold = 1}.}

\item{outlier.pop.threshold}{(integer, optional) Useful to incorporate problematic
populations dragging down polymorphism discovery, but still wanted for analysis.
Use this threshold to allow variance in the number of populations passing
the thresholds described above.
e.g. with \code{outlier.pop.threshold = 2},
you tolerate a maximum of 2 populations failing the
\code{het.threshold} and/or \code{het.dif.threshold}.
Manage outlier markers, individuals and populations
downstream with blacklists and whitelists produced by the function.
Default: \code{outlier.pop.threshold = 1}. See details for more info.}

\item{helper.tables}{(logical) Output tables that show
the number of markers blacklisted or whitelisted based on a series of
automatic thresholds to guide decisions. When \code{interactive.filter == TRUE},
helper tables are written to the directory.
Default: \code{helper.tables = FALSE}.}

\item{coverage.info}{(optional, logical) Use
\code{coverage.info = TRUE}, if you want to visualize
the relationshio between locus coverage and locus observed heterozygosity
statistics. Coverage information is required (e.g. in a vcf file...).
Default: {coverage.info = FALSE}.}

\item{filename}{(optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in
the folder created in the working directory. The file extension appended to
the \code{filename} provided is \code{.rad}.
Default: \code{filename = NULL}.}

\item{vcf.metadata}{(optional, logical or string) For VCF files only.
With \code{logical, TRUE/FALSE}, \code{vcf.metadata = FALSE}, only the genotype
information is kept (GT field). With \code{vcf.metadata = TRUE},
all the metadata contained in the \code{FORMAT} field will be kept in
the tidy data file. radiator is currently keeping/cleaning only these metadata:
\code{"DP", "AD", "GL", "PL", "GQ", "GOF", "NR", "NV"}.
If you need different one, submit a request.
With \code{string}, explicitely ask for a particular FORMAT field you want
to keep (except the GT field that is always imported).
e.g. \code{vcf.metadata = "PL"} or \code{vcf.metadata = c("DP", "GL")}.
Default: \code{vcf.metadata = FALSE}.}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is an object in your
global environment or a file in the working directory
(e.g. "blacklist.txt"). \code{_} and \code{:} in individual's names
are changed to a dash \code{-}.
Default: \code{blacklist.id = NULL}.}

\item{blacklist.genotype}{(optional) Useful to erase genotype with below
average quality, e.g. genotype with more than 2 alleles in diploid likely
sequencing errors or genotypes with poor genotype likelihood or coverage.
The blacklist has a minimum of 2 column headers (markers and individuals).
Markers can be 1 column (CHROM or LOCUS or POS),
a combination of 2 (e.g. CHROM and POS or CHROM and LOCUS or LOCUS and POS) or
all 3 (CHROM, LOCUS, POS). The markers columns must be designated: CHROM (character
or integer) and/or LOCUS (integer) and/or POS (integer). The id column designated
INDIVIDUALS (character) columns header.
The blacklist is an object in your global environment or
a file in the working directory (e.g. "blacklist.genotype.txt").
For de novo VCF, CHROM column
with 'un' need to be changed to 1.
Marker names are cleaned of
separators that interfere with some packages or codes:
\code{/}, \code{:}, \code{-} and \code{.} are changed to an underscore
\code{_}.
Ids are also cleaned of separators that interfere with some packages or codes:
\code{_} and \code{:} are changed to a dash \code{-}.
Default: \code{blacklist.genotype = NULL} for no blacklist of
genotypes to erase.}

\item{whitelist.markers}{(optional) A whitelist containing CHROM (character
or integer) and/or LOCUS (integer) and/or
POS (integer) columns header. To filter by chromosome and/or locus and/or by snp.
The whitelist is an object in your
global environment or a file in the working directory (e.g. "whitelist.txt").
Note that \emph{de novo} CHROM column with 'un' need to be changed to 1.
In the VCF, the column ID is the LOCUS identification (VCF generated from
stacks have the SNP position on the read embedded in the ID,
so the ID = no longer represent the LOCUS). Marker names are cleaned of
separators that interfere with some packages or codes:
\code{/}, \code{:}, \code{-} and \code{.} are changed to an underscore
\code{_}.
Default \code{whitelist.markers = NULL} for no whitelist of markers.}

\item{monomorphic.out}{(optional) Should the monomorphic
markers present in the dataset be filtered out ?
Default: \code{monomorphic.out = TRUE}.}

\item{max.marker}{(integer, optional) For large PLINK files,
useful to subsample marker number. e.g. if the data set
contains 200 000 markers and \code{max.marker = 10000}, 10000 markers are
subsampled randomly from the 200000 markers. If you need specific markers,
use \code{whitelist.markers} argument.
Default: \code{max.marker = NULL}.}

\item{snp.ld}{(optional) \strong{For data with locus and SNP info, like VCF and DArT file}.
SNP short distance linkage disequilibrium pruning. With anonymous markers from
RADseq/GBS de novo discovery, you can minimize linkage disequilibrium (LD) by
choosing among these 5 options:
\enumerate{
\item \code{snp.ld = "random"} for a random selection of 1 SNP on the read,
\item \code{snp.ld = "first"} for the first one on the read...,
\item \code{snp.ld = "last"} for the last SNP on the read and
\item \code{snp.ld = "middle"} for locus with > 2 SNPs/read the option to select at random
one SNP between the first and the last SNP on the read. If the locus as <= 2
SNPs on the read, the first one is selected. Note that for that last option,
the numbers are reported.
\item \code{snp.ld = "maf"} will select the SNP on the locus with the maximum global
Minor Allele Frequency (MAF).
}
For long distance linkage disequilibrium pruning, see details below.
Default: \code{snp.ld = NULL}, for no pruning.}

\item{common.markers}{(optional) Logical. Default: \code{common.markers = TRUE},
will only keep markers in common (genotyped) between all the populations.}

\item{strata}{(optional)
The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}.
If a \code{strata} file is specified with all file formats that don't
require it, the strata argument will have precedence on the population
groupings used internally in those file formats. For file formats without
population/strata groupings (e.g. vcf, haplotype files) if no strata file is
provided, 1 pop/strata grouping will automatically be created.
For vcf and haplotypes file, the strata can also be used as a whitelist of id.
Samples not in the strata file will be discarded from the data set.
The \code{STRATA} column can be any hierarchical grouping.
To create a strata file see \code{\link[radiator]{individuals2strata}}.
If you have already run
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data,
the strata file is similar to a stacks \emph{population map file},
make sure you
have the required column names (\code{INDIVIDUALS} and \code{STRATA}).
The strata column is cleaned of a white spaces that interfere with some
packages or codes: space is changed to an underscore \code{_}.
Default: \code{strata = NULL}.}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
White spaces in population names are replaced by underscore.
Default: \code{pop.levels = NULL}.}

\item{pop.labels}{(optional, string) Use this argument to rename/relabel
your pop or combine your pop. e.g. To combine \code{"QUE"} and \code{"ONT"}
into a new pop called \code{"NEW"}:
(1) First, define the levels for your pop with \code{pop.levels} argument:
\code{pop.levels = c("QUE", "ONT", "ALB")}.
(2) then, use \code{pop.labels} argument:
\code{pop.labels = c("NEW", "NEW", "ALB")}.
To rename \code{"QUE"} to \code{"TAS"}:
\code{pop.labels = c("TAS", "ONT", "ALB")}.
Default: \code{pop.labels = NULL}. If you find this too complicated,
there is also the \code{strata} argument that can do the same thing,
see below.
White spaces in population names are replaced by underscore.}

\item{pop.select}{(string, optional) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops).
Default: \code{pop.select = NULL}}

\item{parallel.core}{(optional) The number of core used for parallel
execution during vcf import.
Default: \code{parallel::detectCores() - 1}.}
}
\value{
The function returns inside the global environment a list with
15 objects, the objects names are found by using \code{names(your.list.name)}:

\enumerate{
\item filtered tidy data frame: \code{$tidy.filtered.het}
\item whitelist of markers:\code{$whitelist.markers}
\item the strata:\code{$strata}
\item the filters parameters used:\code{$filters.parameters}
\item the individual's heterozigosity:\code{$individual.heterozigosity}
\item the heterozygosity statistics per populations and overall:\code{$heterozygosity.statistics}
\item the blacklisted individuals based on the individual's heterozigosity:\code{$blacklist.ind.het}
\item a list containing the helper tables:\code{$helper.table.het}
\item the boxplot of individual observed heterozygosity:\code{$individual.heterozygosity.boxplot}
\item the manhattan plot of individual heterozygosity:\code{$individual.heterozygosity.manhattan.plot}
\item the boxplot of observed heterozygosity averaged across markers and pop:\code{$markers.pop.heterozygosity.boxplot}
\item the density plot of observed heterozygosity averaged across markers and pop:\code{$markers.pop.heterozygosity.density.plot}
\item the manhattan plot of observed heterozygosity averaged across markers and pop:\code{$markers.pop.heterozygosity.manhattan.plot}
\item the relationship between the number of SNPs on a locus and locus observed heterozygosity statistics:\code{$snp.per.locus.het.plot}
\item the relationship between locus coverage (read depth) and locus observed heterozygosity statistics:\code{$het.cov.plot}
\item whitelist of markers:\code{$whitelist.markers}
\item whitelist of markers:\code{$whitelist.markers}
}


In the working directory, output is conditional to \code{interactive.filter} argument
}
\description{
Observed Heterozygosity based filtering.
The filter arguments of \code{filter_het} allows you to test rapidly
if departure from realistic expectations of heterozygosity statistics
are a problem in downstream analysis.
\enumerate{
\item Highlight outliers individual's observed heterozygosity for a quick
diagnostic of mixed samples or poor polymorphism discovery. The statistic is
also contrasted with missing data to help differentiate problems.
(also computed in \link{detect_mixed_genomes})

\item The observed heterozygosity in the dataset: an assembly artefact,
a genotyping problem, a problem of population groupings
or a reliable signal of biological polymorphism?
Detect assembly artifact or genotyping problem (e.g. under/over-splitting loci)
by looking at marker's observed heterozygosity statistics by population
or overall.

\item Use haplotype or snp level statistics. When the haplotype approach
is selected, consistencies of heterozygosity statistics along the read are
highlighted.

\item Interactive approach help by visualizing the data before making
a decision on thresholds.
}
}
\details{
\strong{Interactive version}

There are 4 steps in the interactive version to visualize and filter
the data based heterozygosity statistics:

Step 1. Individual's observed heterozygosity: outliers that might represent mixed samples
Step 2. Blacklist outliers based on a proportion threshold of mean observed heterozygosity
Step 3. Observed heterozygosity statistics per populations and overall
Step 4: Blacklist markers based on observed heterozygosity



\strong{outlier.pop.threshold}

If your a regular radiator user, you've seen the \code{pop.num.threshold}.
\code{outlier.pop.threshold}, is different and requires more thinking,
because the number of populations genotyped potentially vary across markers,
which makes the use of \code{pop.num.threshold} less optimal.
e.g. If only 4 populations out of 10 are genotyped, for a marker you want
to keep, using \code{pop.num.threshold = 0.6} will lead to unwanted results
and inconsistensis.
It's easier to tolerate outliers with this new approach:
\code{outlier.pop.threshold = 2}.


\strong{Individual observed heterozygosity (averaged across markers):}
To help discard an individual based on his observed heterozygosity
(averaged across markers),
use the manhanttan plot to:
\enumerate{
\item contrast the individual with population and overall samples.
\item visualize the impact of missigness information (based on population or
overall number of markers) and the individual observed heterozygosity. The
larger the point, the more missing genotypes.
}
\strong{Outlier above average:}
\itemize{
\item potentially represent two samples mixed together (action: blacklist), or...
\item a sample with more sequecing effort (point size small): did you merge your replicates fq files ? (action : keep and monitor)
\item a sample with poor sequencing effort (point size large) where the genotyped markers are
all heterozygotes, verify this with missingness (action: discard)
}
In all cases, if there is no bias in the population sequencing effort,
the size of the point will usually be "average" based on the population or
overall number of markers.


You can visualize individual observed heterozygosity, choose thresholds and
then visualize, choose thresholds and filter markers based on observed
heterozygosity in one run with: \pkg{radiator} \code{\link{filter_het}}.

\strong{Outlier below average:}
\itemize{
\item A point with a size larger than the population or overall average (= lots of missing):
the poor polymorphism discovery of the sample is probably the result of bad
DNA quality, a bias in sequencing effort, etc. (action: blacklist)
\item A point with a size that looks average (not much missing):
this sample requires more attention (action: blacklist) and conduct more tests.
e.g. for biallelic data, look for coverage imbalance between ALT/REF allele.
At this point you need to distinguish between an artifact of poor polymorphism discovery
or a biological reason (highly inbred individual, etc.).
}
}
\seealso{
\link{plot_density_distribution_het}
}
