% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_hwe.R
\name{filter_hwe}
\alias{filter_hwe}
\title{Filter markers based on Hardy-Weinberg Equilibrium}
\usage{
filter_hwe(interactive.filter = TRUE, data, strata = NULL,
  hw.pop.threshold = NULL, midp.threshold = "****", filename = NULL,
  blacklist.id = NULL, whitelist.markers = NULL, pop.levels = NULL,
  pop.labels = NULL, pop.select = NULL,
  parallel.core = parallel::detectCores() - 1, verbose = TRUE)
}
\arguments{
\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. With default: \code{interactive.filter == TRUE}, the user is
asked to see figures of distribution before making decisions for filtering.}

\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{strata}{(optional)
The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}.
If a \code{strata} file is specified with all file formats that don't
require it, the strata argument will have precedence on the population
groupings used internally in those file formats. For file formats without
population/strata groupings (e.g. vcf, haplotype files) if no strata file is
provided, 1 pop/strata grouping will automatically be created.
For vcf and haplotypes file, the strata can also be used as a whitelist of id.
Samples not in the strata file will be discarded from the data set.
The \code{STRATA} column can be any hierarchical grouping.
To create a strata file see \code{\link[radiator]{individuals2strata}}.
If you have already run
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data,
the strata file is similar to a stacks \emph{population map file},
make sure you
have the required column names (\code{INDIVIDUALS} and \code{STRATA}).
The strata column is cleaned of a white spaces that interfere with some
packages or codes: space is changed to an underscore \code{_}.
Default: \code{strata = NULL}.}

\item{hw.pop.threshold}{(integer, optional)
Remove markers that have a certain number of pops in Hardy-Weinberg
disequilibrium.
With default, all populations in dataset need to be in HWD before discarding
the marker.
Default: \code{hw.pop.threshold = NULL}.}

\item{midp.threshold}{(character, optional)
By default the function generates blacklists/whitelists of markers and
filtered tidy datasets for the 5 mid p-value.
However, to get a final filtered object associated with the output of the
function, user need to choose one
of the 5 mid p-value \code{"*", "**", "***", "****", "*****"}.
With default, a very conservative mid p-value threshold (= 0.0001) is selected.
Default: \code{midp.threshold = "****"}.}

\item{filename}{(optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in
the folder created in the working directory. The file extension appended to
the \code{filename} provided is \code{.rad}.
Default: \code{filename = NULL}.}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is an object in your
global environment or a file in the working directory
(e.g. "blacklist.txt"). \code{_} and \code{:} in individual's names
are changed to a dash \code{-}.
Default: \code{blacklist.id = NULL}.}

\item{whitelist.markers}{(optional) A whitelist containing CHROM (character
or integer) and/or LOCUS (integer) and/or
POS (integer) columns header. To filter by chromosome and/or locus and/or by snp.
The whitelist is an object in your
global environment or a file in the working directory (e.g. "whitelist.txt").
Note that \emph{de novo} CHROM column with 'un' need to be changed to 1.
In the VCF, the column ID is the LOCUS identification (VCF generated from
stacks have the SNP position on the read embedded in the ID,
so the ID = no longer represent the LOCUS). Marker names are cleaned of
separators that interfere with some packages or codes:
\code{/}, \code{:}, \code{-} and \code{.} are changed to an underscore
\code{_}.
Default \code{whitelist.markers = NULL} for no whitelist of markers.}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
White spaces in population names are replaced by underscore.
Default: \code{pop.levels = NULL}.}

\item{pop.labels}{(optional, string) Use this argument to rename/relabel
your pop or combine your pop. e.g. To combine \code{"QUE"} and \code{"ONT"}
into a new pop called \code{"NEW"}:
(1) First, define the levels for your pop with \code{pop.levels} argument:
\code{pop.levels = c("QUE", "ONT", "ALB")}.
(2) then, use \code{pop.labels} argument:
\code{pop.labels = c("NEW", "NEW", "ALB")}.
To rename \code{"QUE"} to \code{"TAS"}:
\code{pop.labels = c("TAS", "ONT", "ALB")}.
Default: \code{pop.labels = NULL}. If you find this too complicated,
there is also the \code{strata} argument that can do the same thing,
see below.
White spaces in population names are replaced by underscore.}

\item{pop.select}{(string, optional) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops).
Default: \code{pop.select = NULL}}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\value{
A list in the global environment objects:
\enumerate{
\item $path.folder: the path to the folder generated.
\item $hw.pop.threshold: the number of populations tolerated to be in HWD before
blacklisting the markers.
\item $plot.hwd.thresholds: useful figure that highlight the number of markers
blacklisted based on the number of populations in HWD and mid p-value thresholds.
\item $plot.tern: ternary plot of markers.
\item $hw.manhattan: manhattan plot of markers in Hardy-Weinberg disequilibrium.
\item $hwe.pop.sum: a summary tibble with populations, number of markers in total,
number of markers monomorphic for the populations,
number of markers in Hardy-Weinberg Equilibrium (HWE),
number of markers in Hardy-Weinberg Dquilibrium (HWD) with all the different
mid p-values observed on the data.
\item $midp.threshold: the mid p-value threshold chosen for the final dataset (next)
\item $tidy.hw.filtered: the final filtered dataset (oter datasets \code{.rad}
are generated automatically by the function, check the folder)
}

Written in the folder:
\enumerate{
\item genotypes.summary.tsv: A tibble with these columns:
\code{MARKERS, POP_ID, HET, HOM_ALT, HOM_REF, MISSING, N,
FREQ_ALT, FREQ_REF, FREQ_HET, FREQ_HOM_REF_O, FREQ_HET_O, FREQ_HOM_ALT_O,
FREQ_HOM_REF_E, FREQ_HET_E, FREQ_HOM_ALT_E, N_HOM_REF_EXP,
N_HET_EXP, N_HOM_ALT_EXP, HOM_REF_Z_SCORE, HOM_HET_Z_SCORE,
HOM_ALT_Z_SCORE, READ_DEPTH}
\item hw.pop.sum.tsv: a summary tibble with populations, number of markers in total,
number of markers monomorphic for the populations,
number of markers in Hardy-Weinberg Equilibrium (HWE),
number of markers in Hardy-Weinberg Dquilibrium (HWD) with all the different
mid p-values observed on the data.
\item hwd.helper.table.tsv: useful tibble that highlight the number of markers
blacklisted based on the number of populations in HWD and mid p-value thresholds.
\item hwd.plot.blacklist.markers.pdf: useful figure that highlight the number of markers
blacklisted based on the number of populations in HWD and mid p-value thresholds.
\item hwe.manhattan.plot.pdf: manhattan plot of markers in Hardy-Weinberg disequilibrium.
\item hwe.ternary.plots.missing.data.pdf: ternary plot of markers.
\item tidy.filtered.hwe.xxx.mid.p.value.xxx.hw.pop.threshold.rad: several
tidy dataset filtered with different mid p value and populations in HWD thresholds
\item whitelist.markers.hwe.xxx.mid.p.value.xxx.hw.pop.threshold.tsv: several
whitelist of  markers with different mid p value and populations in HWD thresholds
\item blacklist.markers.hwd.xxx.mid.p.value.xxx.hw.pop.threshold.tsv: several
blacklist of markers with different mid p value and populations in HWD thresholds
}
}
\description{
Testing markers for Hardy-Weinberg proportions is a valuable tool
for the analysis and quality control of RADseq datasets.
HWE can highligh genotyping errors, presence of null alleles,
sequence duplication, copy number variation and other sequencing problems
related to read depth.
This function is designed for
\strong{bi-allelic markers} and uses the exact test from the package
HardyWeinberg. The function is speedy because it uses the
C++ code developed by Christopher Chang and also available in PLINK.
The p-value generated by the function is the mid p-value. It's computed
as half the probability of the current sample + the probabilities of all
samples that are more extreme (see references below). Several output are
generated to help users filter the data (see details).

Prior to HW filtering, I highly recommend removing outlier individuals,
filtering coverage and genotype likelihood (see details).
}
\details{
\strong{Interactive version}

The user is asked to look at figures before choosing filter thresholds.


\strong{HWE threshold}

I recommend starting with a low threshold.
Serious genotyping errors will generate extreme p-values (e.g. 1e-50),
which are detected by any reasonable configuration of this test,
while various life-history caracteristics will deflate/inflate
Hardy-Weinberg equilibrium.
Consequently, it's dangerous to choose a threshold that filters out too many
markers.

\strong{strategies:}
Disk space is cheap! Consequently, the function will automatically generate
several blacklists/whitelists of markers and
filtered tidy data (in the directory)
based on the \code{hw.pop.threshold} for 4 groups of mid p-values:
\itemize{
\item MID_P_VALUE <= 0.00001: *****
\item MID_P_VALUE <= 0.0001: ****
\item MID_P_VALUE <= 0.001: ***
\item MID_P_VALUE <= 0.01: **
\item MID_P_VALUE <= 0.05: *
}

Test the sensitivity of downstream analysis and delete unwanted datasets.

\strong{missing data}

The mid-p adjustment tends to bring the null rejection rate in line with the
nominal p-value, and also reduces the filter's tendency to favor retention
of SNPs with missing data (Graffelman and Moreno, 2013, Purcell et al., 2007).

If pattern of missing data is present in the dataset, or when missing data
accross markers vary by more than 0.10, you should not apply a single mid-p-value
threshold accross markers.

\strong{Read depth, pooling lanes/chips and weird pattern of individual heterozygosity}

Because of read depth, heterozygote deficiency is usually observed in RADseq data,
but if sequencing lanes/chips were combined to generate individuals with more
coverage, the situation will likely be the reverse: heterozygote excess.
If lanes/chips pooling was used or if highly variable sequencing coverage is
observed between individuals and/or markers, there's a couple qc and filtering
steps you should do before conducting HWE filtering.

\itemize{
\item run \pkg{radiator} \code{\link{detect_mixed_genomes}} and
\code{\link{detect_het_outliers}} to highlight
outlier individuals with potential heterozygosity problems and get an idea of
the genotyping and heterozygote miscall rate
\item I recommend normalizing the data before \emph{de novo} assembly, if this
is not possible...
\item use radiator::filter_rad or a more chirurgical approach to
coverage and genotype likelihood with radiator::filter_rad.
}


\strong{Permutation test}
Hardy-Weinberg equilibrium refers to the statistical independence of alleles
within individuals. This independence can also be assessed by permutation test
inside the HardyWeinberg package of Jan Graffelman. To filter out markers with
genotyping problems the approach provided in this function is enough.

\strong{Power test for HWE}
Based on allele count/frequency and sample size. This function is longer to
generate for each markers and is on my todo list to include it in this filter.


\strong{Markers under selection and genome scans:}

Scared of deleting those precious markers or that the filter might interfere
with genome scan analysis/detection ? Don't be. Your markers or analysis is no
good if it's done on bad data... Test the sensitivity of your downstream
analysis with the datasets generated with the different thresholds.
}
\note{
\strong{Hardy-Weinberg assumptions (refresh):}
\enumerate{
\item Diploid organisms
\item Only reproduction sexual occurs
\item Generations are non-overlapping
\item Mating is random
\item Population size is infinitely large
\item Allele frequencies are equal in the sexes
\item Migration, Mutation and Selection are negligible
}
}
\examples{
\dontrun{
require(HardyWeinberg)
require(ggtern)
library(radiator)
# for the interactive version (recommended)
turtle.pop <- radiator::filter_hwe(
   data = "turtle.vcf",
   strata = "turtle.strata.tsv",
   filename = "hwe.turtle"
)
}
}
\references{
Weir, B.S. (1996) Genetic data analysis II. Sinauer Associates,
Massachusetts. See Chapter3.

Wigginton, J.E., Cutler, D.J. and Abecasis, G.R. (2005)
A note on exact tests of Hardy-Weinberg equilibrium,
American Journal of Human Genetics (76) pp. 887-893.

Purcell et al. (2007) PLINK: A Toolset for Whole-Genome Association
and Population-Based Linkage Analysis.
American Journal of Human Genetics 81(3) pp. 559-575.

Graffelman, J. and Moreno, V. (2013) The mid p-value in exact
tests for Hardy-Weinberg equilibrium, Statistical Applications in Genetics
and Molecular Biology 12(4) pp. 433-448.

Graffelman J, Jain D, Weir B (2017)
A genome-wide study of Hardy-Weinberg equilibrium with next generation
sequence data. Human Genetics, 136, 727-741.
}
