% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_deprecated_functions.R
\name{filter_individual}
\alias{filter_individual}
\title{Individual filter}
\usage{
filter_individual(
  interactive.filter = TRUE,
  data,
  strata = NULL,
  ind.approach = "pop",
  ind.threshold = 0.5,
  prob.pop.threshold = 0,
  percent = FALSE,
  filename = NULL,
  parallel.core = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. With default: \code{interactive.filter == TRUE}, the user is
asked to see figures of distribution before making decisions for filtering.}

\item{data}{14 options for input (\strong{diploid data only}): VCFs (SNPs or Haplotypes,
to make the vcf population ready),
plink (tped, bed), stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop, DArT,
and a data frame in long/tidy or wide format. To verify that radiator detect
your file format use \code{\link{detect_genomic_format}} (see example below).
Documented in \strong{Input genomic datasets} of \code{\link{tidy_genomic_data}}.

\strong{DArT and VCF data}: \pkg{radiator} was not meant to generate alleles
and genotypes if you are using a VCF file with no genotype
(only genotype likelihood: GL or PL).
Neither is \pkg{radiator} able to magically generate a genind object
from a SilicoDArT dataset. Please look at the first few lines of your dataset
to understand it's limit before asking raditor to convert or filter your dataset.}

\item{strata}{(optional)
The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}. Documented in \code{\link{read_strata}}.
DArT data: a third column \code{TARGET_ID} is required.
Documented on \code{\link{read_dart}}. Also use the strata read function to
blacklist individuals.
Default: \code{strata = NULL}.}

\item{ind.approach}{(character).
The approach to filter a marker, is it based on the overall number of
genotyped individuals (no concept of populations here),
\code{ind.approach = "overall"}, or
the number of genotyped individuals per population \code{ind.approach = "pop"}.
Default: \code{ind.approach = "pop"}.}

\item{ind.threshold}{The individual threshold, proportion, percentage or
number e.g. 0.70, 70, 15.
Default: \code{ind.threshold = 0.5}.}

\item{prob.pop.threshold}{(integer, optional) Useful to incorporate problematic
populations dragging down polymorphism discovery, but still wanted for analysis.
If individuals with missing genotypes are not managed upstream,
use this threshold to allow variance in the number of populations passing
the \code{ind.threshold} argument. e.g. with \code{prob.pop.threshold = 2},
you tolerate a maximum of 2 populations failing to pass the
\code{ind.threshold}. Manage after the problematic populations
downstream with blacklist of individuals and/or missing data imputations. This
argument is not necessary with \code{ind.approach = "overall"}.
Default: \code{prob.pop.threshold = 0}. See details for more info.}

\item{percent}{Is the threshold a percentage? TRUE or FALSE.
This argument is necessary to distinguish percentage from integer individual
threshold (e.g. 70 percent or 70 individuals).
Default: \code{percent = FALSE}.}

\item{filename}{(optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in
the folder created in the working directory. The file extension appended to
the \code{filename} provided is \code{.rad}.
Default: \code{filename = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{...}{(optional) To pass further arguments for fine-tuning the function.}
}
\value{
With \code{interactive.filter = FALSE}, a list in the global environment,
with 7 objects:
\enumerate{
\item $tidy.filtered.ind
\item $whitelist.markers
\item $blacklist.markers
\item $strata
\item $filters.parameters
}

With \code{interactive.filter = TRUE}, a list with 2 additionnal objects is created.
\enumerate{
\item $plot.ind.threshold
\item $ind.threshold.helper.table
}

The object can be isolated in separate object outside the list by
following the example below.
}
\description{
Filter individuals genotyped at a marker from a tidy data
set (long format) of any of these file format:
vcf, plink (tped/tfam), stacks haplotype file, genind,
genepop, data frame in wide format. The function uses
\code{\link[radiator]{tidy_genomic_data}} and
\code{\link[radiator]{tidy_wide}} to load the file. For filtering
you can consider the overall number of individuals (no concept of populations here),
or the number of genotyped individuals per pop. The threshold can be a fixed
number of individuals, a proportion or a percentage.
}
\details{
\strong{Interactive version}

There is 2 steps in the interactive version to visualize and filter
the data based on the number of genotyped individuals:

Step 1. Impact of individual threshold on marker discovery

Step 2. Choose the filtering approach and thresholds



\strong{prob.pop.threshold}

If your a regular radiator user, you've seen the \code{pop.num.threshold}.
\code{prob.pop.threshold}, is a bit different and requires more thinking,
because the number of populations genotyped potentially vary across markers,
which make difficult, sometimes, the use of \code{pop.num.threshold}.
e.g. If only 2 populations out of 10 are genotyped for a marker and you use
\code{pop.num.threshold = 0.6} this could lead to inconsistensis. Thus,
it's easier to use the number of population you are willing to accept
failling to conform to the threshold.
}
\examples{
\dontrun{
turtle.ind <- filter_individual(
data = "turtle.vcf",
strata = "turtle.strata.tsv",
ind.approach = "pop",
ind.threshold = 50,
percent = TRUE,
prob.pop.threshold = 2,
filename = "tidy.data.turtle"
)

#If interactive.filter = TRUE, a list is created and to view the filtered tidy data:
tidy.data <- turtle.ind$tidy.filtered.ind

#Inside the same list, to isolate the blacklist.genotypes:
bg <- turtle.ind$blacklist.genotypes

# The remaining argument are used in tidy_genomic_data during import and allow
# the user to apply some filtering or selection before doing the filtering.
}
}
