% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_individuals.R
\name{filter_individuals}
\alias{filter_individuals}
\title{Filter individuals based on genotyping/missingness rate,
heterozygosity and total coverage}
\usage{
filter_individuals(data, interactive.filter = TRUE,
  filter.individuals.missing = NULL,
  filter.individuals.heterozygosity = NULL,
  filter.individuals.coverage.total = NULL,
  parallel.core = parallel::detectCores() - 1, verbose = TRUE, ...)
}
\arguments{
\item{data}{(2 options) A Genomic Data Structure (GDS)
file or object generated by radiator.

\emph{How to get GDS?}
Look into:
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. Figures of distribution are shown before asking for filtering
thresholds.
Default: \code{interactive.filter = TRUE}.}

\item{filter.individuals.missing}{(optional, double) A proportion above which the individuals are
blacklisted and removed from the dataset.
Default: \code{filter.individuals.missing = NULL}.}

\item{filter.individuals.heterozygosity}{(optional, string of doubles) A proportion below and
above which the individuals are blacklisted and removed from the dataset.
Default: \code{filter.individuals.heterozygosity = NULL}.}

\item{filter.individuals.coverage.total}{(optional, string of doubles) A proportion below and
above which the individuals are blacklisted and removed from the dataset.
Default: \code{filter.individuals.coverage.total = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function. Also used for legacy arguments (see details or
special section)}
}
\value{
A list with the filtered input and blacklist of individuals.
}
\description{
Remove individuals with bad QC based on:
\itemize{
\item missingness (genotyping rate)
\item heterozygosity
\item coverage (total)
}

\strong{Filter targets}: Individuals

\strong{Statistics}: Missingness, heterozygosity and total coverage

Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users who wants to blacklist individuals.
}
\examples{
\dontrun{
require(SeqVarTools)

# blacklisting outliers individuals:
id.qc <- radiator::filter_individuals(
    data = "my.radiator.gds.rad",
    filter.individuals.missing = "outliers",
    filter.individuals.heterozygosity = "outliers",
    filter.individuals.coverage.total = "outliers")

# using values to blacklist individuals:
id.qc <- radiator::filter_individuals(
    data = "my.radiator.gds.rad",
    filter.individuals.missing = 0.5,
    filter.individuals.heterozygosity = c(0.02, 0.03),
    filter.individuals.coverage.total = c(900000, 5000000))

}
}
\seealso{
\code{\link{filter_rad}}
\code{\link{tidy_genomic_data}}, \code{\link{read_vcf}},
\code{\link{tidy_vcf}}.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
