% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ma.R
\name{filter_ma}
\alias{filter_ma}
\title{MAC, MAF and MAD filter}
\usage{
filter_ma(
  data,
  interactive.filter = TRUE,
  ma.stats = "mac",
  filter.ma = NULL,
  calibrate.alleles = "depth",
  keep.biallelic = TRUE,
  filename = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. Figures of distribution are shown before asking for filtering
thresholds.
Default: \code{interactive.filter = TRUE}.}

\item{ma.stats}{(optional, character) The statistic of the alternate (minor)
allele to keep a SNP (see details). Options are: \code{"mac", "maf", "mad"}.
Default: \code{ma.stats = "mac"}.}

\item{filter.ma}{(optional, integer or double) The threshold to
blacklist Minor Allele (see details). e.g. with \code{ma.stats = "mac"} and
\code{filter.ma = 3} will blacklist markers based on minor allele count
less or equal to 3.
This argument as no concept of locus, local or strata or pop.
It's applied globally, by SNPs.
Default: \code{filter.ma = NULL}.}

\item{calibrate.alleles}{(character) When ancestral allele
information is not available, REF / ALT alleles can be calibrated using
alleles count or depth (sequencing coverage for REF and ALT alleles).
Removing individuals will impact REF/ALT alleles. Ideally,
it should be based on observed read depth, when this information is available
for both alleles. By selecting \code{calibrate.alleles = "depth"}, radiator will
check if the information is available, if not, it will use
\code{calibrate.alleles = "count"}.
Using \code{calibrate.alleles = "ancestral"} will turn off calibration and
use the information preset available in the data.
Default: \code{calibrate.alleles = "depth"}.}

\item{keep.biallelic}{(logical) Keep only biallelic variants (REF + 1 ALT)
when computing allele counts from GDS.
Default: \code{keep.biallelic = TRUE}.}

\item{filename}{(optional, character) Write to folder the MAF filtered
tidy dataset. The name will be appended \code{.rad}. With default, the filtered
data is only in the global environment.
Default: \code{filename = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function. Also used for legacy arguments (see details or
special section)}
}
\value{
With \code{interactive.filter = FALSE}, a list in the global environment,
with 7 objects:
\enumerate{
\item $tidy.filtered.mac
\item $whitelist.markers
\item $blacklist.markers
\item $mac.data
\item $filters.parameters
}

With \code{interactive.filter = TRUE}, a list with 4 additionnal objects are generated.
\enumerate{
\item $distribution.mac.global
\item $distribution.mac.local
\item $mac.global.summary
\item $mac.helper.table
}

\strong{mac.helper.table:}

First and second variables (in columns) represents POP_ID and sample size (n).

The last 2 rows are the local MAF (suggested based on the lowest pop value)
and the TOTAL/GLOBAL observations.

Columns starting with ALT are the variable corresponding
to the number of alternative (ALT) allele (ranging from 1 to 20).
The observations in the ALT allele variable columns are the local (for the pop)
and global (last row) MAF of your dataset.
e.g. ALT_3 can potentially represent 3 heterozygote individuals with
the ALT allele or 1 homozygote individuals for the ALT allele and
1 heterozygote individual. And so on...
}
\description{
The Allele Frequency is the relative frequency of an allele for a markers/SNP.
Calculations usually involve classifying the alleles into Major/Minor,
Reference (REF)/Alternate (ALT) and/or Ancestral or derived allele.
radiator provides a filter based of Minor Allele Count, frequency and depth function.
Remove/blacklist markers based on Minor/Alternate Allele Count (MAC), Frequency (MAF)
or Depth of coverage (MAD).
Use it to remove noise, sequencing errors or low marker polymorphism.
Some analysis performs better with the full spectrum of allele frequency,
consequently it's important to understand the limite of using the
thresholds to avoid generating biases.

\strong{Filter targets}: Marker's alternate allele(s)

\strong{Statistics}: count, frequency or depth of the allele per markers.
This filter as NO concept of strata/populations. It is computed globally.
}
\note{
Thanks to Charles Perrier and Jeremy Gaudin for very useful comments
on previous version of this function.
}
\section{MAC, MAF or MAD ?}{

Using count or frequency to remove a SNPs ? The preferred choice in radiator
as changed from frequency to count, because we think the filtering should not
alter the spectrum and this is only achieved if the same criteria is applied
for each SNP.


Even small differences in missing data between RADseq markers
generates differences in MAF frequency thresholds applied.

Example with a datset consisting of N = 36 individuals and 3 SNPs
with varying level of missing genotypes:

\itemize{
\item \code{SNP number : number samples genotypes : REF/ALT counts}
\item SNP1 : 36 : 69/3
\item SNP2 : 30 : 65/3
\item SNP3 : 24 : 45/3
}

Each SNPs have the same alternate allele count, corresponding to
2 individuals with the polymorphism: 1 homozygote + 1 heterozygote.
Applying a MAF threshold of
0.05 would mean that SNP3 would be blacklisted
(\code{24 * 2 * 0.05 = 2.4 alt alleles required to pass}).


\strong{Using count instead of frequency allows each RADseq markers,
with varying missing data, to be treated equally.}
}

\section{Advance mode}{


\emph{dots-dots-dots ...} allows to pass several arguments for fine-tuning the function:
\enumerate{
\item \code{filter.common.markers} (optional, logical).
Default: \code{filter.common.markers = FALSE},
Documented in \code{\link{filter_common_markers}}.
\item \code{filter.monomorphic} (logical, optional) Should the monomorphic
markers present in the dataset be filtered out ?
Default: \code{filter.monomorphic = TRUE}.
Documented in \code{\link{filter_monomorphic}}.
\item \code{path.folder}: to write ouput in a specific path
(used internally in radiator).
Default: \code{path.folder = getwd()}.
If the supplied directory doesn't exist, it's created.
}
}

\section{Interactive version}{


To help choose a threshold for the local and global MAF
use the interactive version.

2 steps in the interactive version:

Step 1. Visualization and helper table.

Step 2. Minor Allele statistic and threshold selection

Step 3. Filtering markers based on MAC, MAF or MAD
}

\examples{
\dontrun{
# The minumum
mac <- radiator::filter_ma(data = turtle.tidy.data)
# This will use the default: interactive version,
# a list is created and to view the filtered tidy data:
mac.tidy.data <- mac$tidy.filtered.mac

# No user interaction

# Using filter.ma = 4, is a good practice to remove mostly sequencing errors
# and assembly artifacts, because it requires the markers to be genotyped in
# 4 heterozygote individuals or 2 homozygote individuals for the alternate
# allele or 2 heterozygote individuals and 1 homozygote individual for the
# alternate allele. Overall, never less than 2 indiduals are required.

mac <- radiator::filter_ma(
        data = turtle.gds, # using gds object
        ma.stats = "mac",
        filter.ma = 4,
        filename = "turtle.mac")

# This will remove monomorphic markers (by default filter.monomorphic = TRUE)
# The filtered data will be written in the directory under the name:
# turtle.maf.rad
}
}
\references{
Good reading: Linck, E., Battey, C. (2019).
Minor allele frequency thresholds strongly affect population structure
inference with genomic data sets.
Molecular Ecology Resources 19(3), 639-647.
https://dx.doi.org/10.1111/1755-0998.12995
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
