% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_maf.R
\name{filter_maf}
\alias{filter_maf}
\title{MAF filter}
\usage{
filter_maf(data, strata = NULL, interactive.filter = TRUE,
  maf.thresholds = NULL, monomorphic.out = TRUE, common.markers = FALSE,
  snp.ld = NULL, pop.select = NULL, filename = NULL,
  parallel.core = parallel::detectCores() - 1, verbose = TRUE, ...)
}
\arguments{
\item{data}{12 options: VCF (SNPs or Haplotypes,
to make the vcf population ready, see details below),
plink, stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop, DArT,
and a data frame in long/tidy or wide format.
\emph{See details} of \code{\link[radiator]{tidy_genomic_data}}.}

\item{strata}{(optional)
The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}.
If a \code{strata} file is specified with all file formats that don't
require it, the strata argument will have precedence on the population
groupings used internally in those file formats. For file formats without
population/strata groupings (e.g. vcf, haplotype files) if no strata file is
provided, 1 pop/strata grouping will automatically be created.
For vcf and haplotypes file, the strata can also be used as a whitelist of id.
Samples not in the strata file will be discarded from the data set.
The \code{STRATA} column can be any hierarchical grouping.
To create a strata file see \code{\link[radiator]{individuals2strata}}.
If you have already run
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data,
the strata file is similar to a stacks \emph{population map file},
make sure you
have the required column names (\code{INDIVIDUALS} and \code{STRATA}).
The strata column is cleaned of a white spaces that interfere with some
packages or codes: space is changed to an underscore \code{_}.
Default: \code{strata = NULL}.}

\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. With default: \code{interactive.filter == TRUE}, the user is
shown figures and helper tables before making decisions for filtering.}

\item{maf.thresholds}{(optional) Default: \code{maf.thresholds = NULL}. For no
MAF filtering.

String with 5 values
(e.g. \code{maf.thresholds = c("maf approach", local maf, "OR", global maf, number pop threshold)},
\code{maf.thresholds = c("SNP", 0.001, "OR", 0.001, 1)}):
\enumerate{
\item maf approach (character: "SNP"/"locus"):
MAF filtering is conducted by SNPs or locus.
\code{"SNP"}: will consider all the SNPs on the same locus/read as independent.
\code{"locus"}: looks at the minimum MAF found on the
read/locus. Using this option will discard all the markers/snp on
that read based on the thresholds chosen. For the locus approach to work, your dataset
requires SNP and Locus info (e.g. from a VCF file).

\item local maf (double): The maf threshold at the population level.
Not sure about the threshold to use, choose the interactive mode argument.

\item maf operator (character: "OR" / "AND"):
When filtering over LOCUS or SNP, 2 options available to consider the local/global MAF:
The local \code{"AND"} the global MAF. The local \code{"OR"}
global MAF.

\item global maf (double): The maf threshold taken at the dataset level.
Not sure about the threshold to use, choose the interactive mode argument.

\item maf pop num threshold (integer)
The number of pop required to pass the thresholds
to keep the marker. Usually, I always use \code{1}, for 1 pop is required to
pass thresholds and keep the marker.
}}

\item{monomorphic.out}{(optional) Should the monomorphic
markers present in the dataset be filtered out ?
Default: \code{monomorphic.out = TRUE}.}

\item{common.markers}{(optional) Logical. Default: \code{common.markers = TRUE},
will only keep markers in common (genotyped) between all the populations.}

\item{snp.ld}{(optional) \strong{For data with locus and SNP info, like VCF and DArT file}.
SNP short distance linkage disequilibrium pruning. With anonymous markers from
RADseq/GBS de novo discovery, you can minimize linkage disequilibrium (LD) by
choosing among these 4 options:
\code{snp.ld = "random"} for a random selection of 1 SNP on the read,
\code{snp.ld = "first"} for the first one on the read...,
\code{snp.ld = "last"} for the last SNP on the read and
\code{snp.ld = "middle"} for locus with > 2 SNPs/read the option to select at random
one SNP between the first and the last SNP on the read. If the locus as <= 2
SNPs on the read, the first one is selected. Note that for that last option,
the numbers are reported. For long distance linkage
disequilibrium pruning, see details below.
Default: \code{snp.ld = NULL}.}

\item{pop.select}{(string, optional) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops).
Default: \code{pop.select = NULL}}

\item{filename}{(optional, character) Write to folder the MAF filtered
tidy dataset. The name will be appended \code{.rad}. With default, the filtered
data is only in the global environment.
Default: \code{filename = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during vcf import.
Default: \code{parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) To pass further arguments for fine-tuning the filter with
whitelist and blacklists (see details below).}
}
\value{
With \code{interactive.filter = FALSE}, a list in the global environment,
with 7 objects:
\enumerate{
\item $tidy.filtered.maf
\item $whitelist.markers
\item $blacklist.markers
\item $maf.data
\item $filters.parameters
}

With \code{interactive.filter = TRUE}, a list with 4 additionnal objects are generated.
\enumerate{
\item $distribution.maf.global
\item $distribution.maf.local
\item $maf.global.summary
\item $maf.helper.table
}

\strong{maf.helper.table:}

First and second variables (in columns) represents POP_ID and sample size (n).

The last 2 rows are the local MAF (suggested based on the lowest pop value)
and the TOTAL/GLOBAL observations.

Columns starting with ALT are the variable corresponding
to the number of alternative (ALT) allele (ranging from 1 to 20).
The observations in the ALT allele variable columns are the local (for the pop)
and global (last row) MAF of your dataset.
e.g. ALT_3 can potentially represent 3 heterozygote individuals with
the ALT allele or 1 homozygote individuals for the ALT allele and
1 heterozygote individual. And so on...
}
\description{
Minor Allele Frequency filter. Most arguments are inherited from
\code{\link[radiator]{tidy_genomic_data}}.
}
\details{
To help choose a threshold for the local and global MAF
use the interactive version.

2 steps in the interactive version:

Step 1. Global and Local MAF visualization and helper table.

Step 2. Filtering markers based on the different MAF arguments


\strong{... :dot dot dot arguments}

The arguments listed here are available to tailor your dataset prior to MAF filtering.
The description of these arguments are found in
\code{\link[radiator]{tidy_genomic_data}}: \code{whitelist.markers,
blacklist.id, blacklist.genotype, pop.levels, pop.labels, vcf.metadata, module}.
}
\examples{
\dontrun{
# The minumum
turtle.maf <- filter_maf(
data = "turtle.vcf",
strata = "turtle.strata.tsv")
#This will use the default: interactive version,


#With default interactive.filter = TRUE, a list is created and to view the filtered tidy data:
tidy.data <- turtle.maf$tidy.filtered.maf

# The remaining argument are used in tidy_genomic_data during import and allow
# the user to apply some filtering or selection before doing the MAF filtering.
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
