% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_snp_position_read.R
\name{filter_snp_position_read}
\alias{filter_snp_position_read}
\title{Filter markers/SNP based on their position on the read}
\usage{
filter_snp_position_read(data, strata = NULL,
  interactive.filter = TRUE, filter.snp.position.read = NULL,
  filename = NULL, parallel.core = parallel::detectCores() - 1,
  verbose = TRUE, ...)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{strata}{(path or object) The strata file or object.
Additional documentation is available in \code{\link{read_strata}}.
Use that function to whitelist/blacklist populations/individuals.
Option to set \code{pop.levels/pop.labels} is also available.}

\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. Figures of distribution are shown before asking for filtering
thresholds.
Default: \code{interactive.filter = TRUE}.}

\item{filter.snp.position.read}{(character)
Options are: \code{"outliers", "q75", "iqr", c(min value,max value)}.
For a safe and conservative
value, use \code{"outliers"}, this will remove SNPs with outlier position on
the reads.
Default: \code{filter.snp.read.position = NULL}.}

\item{filename}{(optional) Name of the filtered tidy data frame file
written to the working directory (ending with \code{.tsv})
Default: \code{filename = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function. Also used for legacy arguments (see details or
special section)}
}
\value{
A list in the global environment with 6 objects:
\enumerate{
\item $snp.number.markers
\item $number.snp.reads.plot
\item $whitelist.markers
\item $tidy.filtered.snp.number
\item $blacklist.markers
\item $filters.parameters
}

The object can be isolated in separate object outside the list by
following the example below.
}
\description{
This filter removes markers/SNPs based on their position on the read.
The data requires snp, locus and col information (e.g. from a VCF file).

The impact of assembly artifacts can be tested in downstream analysis with
the whitelist and blacklist generated by this function.


\strong{Filter targets}: Markers

\strong{Statistics}: The position of the SNPs on the read.
}
\details{
\strong{Interactive version}

There are 2 steps in the interactive version to visualize and filter
the data based on the number of SNP on the read/locus:

Step 1. SNP number per read/locus visualization

Step 2. Choose the filtering thresholds
}
\examples{
\dontrun{
turtle <- radiator::filter_snp_position_read(
data = "turtle.vcf",
strata = "turtle.strata.tsv",
filter.snp.position.read = "outliers",
filename = "tidy.data.turtle.tsv"
)
}
}
