% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_whitelist.R
\name{filter_whitelist}
\alias{filter_whitelist}
\title{Filter dataset with whitelist of markers}
\usage{
filter_whitelist(data, whitelist.markers = NULL, verbose = TRUE, ...)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{write_seqarray}} or
\code{\link{tidy_vcf}}.}

\item{whitelist.markers}{(path or object)
The whitelist is an object in your
global environment or a file in the working directory (e.g. "whitelist.txt").
The dataframe contains one, a combination
or all of these columns: \code{MARKERS, CHROM, LOCUS, POS}.
Columns are cleaned of separators that interfere with some packages or codes, detailed in
\code{\link{clean_markers_names}}.
Default \code{whitelist.markers = NULL}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function. Also used for legacy arguments (see details or
special section)}
}
\description{
Filter dataset with whitelist of markers
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\examples{
\dontrun{
data <- radiator::filter_whitelist(data = data, whitelist.markers = "mywhitelist.tsv")
}
}
\seealso{
\code{\link{read_whitelist}}.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
