% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_individuals_id.R
\name{filter_id_whitelist}
\alias{filter_id_whitelist}
\title{Filter individuals with whitelist}
\usage{
filter_id_whitelist(data, whitelist.id, pop.id.start, pop.id.end, pop.levels,
  filename)
}
\arguments{
\item{data}{A data frame object or file (using ".tsv") with an 'INDIVIDUALS'
column.}

\item{whitelist.id}{The whitelist of individuals with no header column,
the first column is assumed to be the indivuals ID.}

\item{pop.id.start}{The start of your population id 
in the name of your individual sample.}

\item{pop.id.end}{The end of your population id 
in the name of your individual sample.}

\item{pop.levels}{A character string with your populations ordered.}

\item{filename}{The name of the file written to the directory.}
}
\description{
This function use a whitelist of individuals to filter
a dataset.
}
