% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{fis_summary}
\alias{fis_summary}
\title{Fis summary}
\usage{
fis_summary(data, vcf.metadata = TRUE, whitelist.markers = NULL,
  monomorphic.out = TRUE, snp.ld = NULL, common.markers = TRUE,
  maf.thresholds = NULL, blacklist.id = NULL, pop.levels = NULL,
  pop.labels = NULL, strata = NULL, pop.select = NULL,
  filename = NULL, verbose = FALSE)
}
\arguments{
\item{data}{12 options for input: VCFs (SNPs or Haplotypes,
to make the vcf population ready),
plink, stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop, DArT,
and a data frame in long/tidy or wide format. To verify that radiator detect
your file format use \code{\link{detect_genomic_format}} (see example below).
Documented in \strong{Input genomic datasets} of \code{\link{tidy_genomic_data}}.}

\item{strata}{(optional)
The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}. Documented in \code{\link{read_strata}}.
Default: \code{strata = NULL}.}

\item{filename}{(optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in
the working directory. The file extension appended to
the \code{filename} provided is \code{.rad}.
With default: \code{filename = NULL}, the tidy data frame is
in the global environment only (i.e. not written in the working directory...).}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\value{
A list with 2 tables: the long format of loci and populations
coverage statistics and the short format by populations.
}
\description{
Fis summary (per markers) by populations and overall.
}
\details{
The tables contains summary statistics (mean, median, min, max).
}
\keyword{internal}
