% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{folder_prefix}
\alias{folder_prefix}
\title{folder_prefix}
\usage{
folder_prefix(prefix.number = NULL, prefix.name = NULL,
  path.folder = NULL)
}
\description{
Generate a seq and folder prefix
}
\keyword{internal}
