% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator.R
\name{folder_short}
\alias{folder_short}
\title{folder_short}
\usage{
folder_short(f)
}
\arguments{
\item{f}{Folder name}
}
\description{
Extract the short name of the folder generated by radiator
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
