% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{gds2tidy}
\alias{gds2tidy}
\title{gds2tidy}
\usage{
gds2tidy(gds, markers.meta = NULL, individuals = NULL,
  calibrate.alleles = TRUE, parallel.core = parallel::detectCores() -
  1, ...)
}
\description{
GDS to tidy...
}
\keyword{internal}
