% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{generate_geno}
\alias{generate_geno}
\title{generate_geno}
\usage{
generate_geno(x, source, gt = FALSE, gt.vcf.nuc = FALSE,
  gt.vcf = FALSE)
}
\description{
Generate the genotypes and prep for GDS
}
\keyword{internal}
