% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_het_outliers.R
\name{generate_geno_freq_boundaries}
\alias{generate_geno_freq_boundaries}
\title{generate_geno_freq_boundaries}
\usage{
generate_geno_freq_boundaries()
}
\description{
Function that returns a tibble with the min/max values of
genotype freqs possible.

These mins and maxes occur because the genotypes are used to estimate
the allele frequencie.
}
\author{
Eric Anderson \email{eric.anderson@noaa.gov}
}
\keyword{internal}
