% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{generate_geno_v1}
\alias{generate_geno_v1}
\title{generate_geno_v1------------------------------------------------------}
\usage{
generate_geno_v1(x, data.source, gt = FALSE, gt.vcf.nuc = FALSE,
  gt.vcf = FALSE)
}
\description{
Generate the genotypes and prep for GDS
}
\keyword{internal}
