% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{generate_id_stats}
\alias{generate_id_stats}
\title{generate_id_stats}
\usage{
generate_id_stats(gds, missing = TRUE, heterozygosity = TRUE,
  coverage = TRUE, subsample = NULL, depth.info = FALSE,
  path.folder = NULL, plot = TRUE, file.date = NULL,
  parallel.core = parallel::detectCores() - 1, verbose = TRUE)
}
\description{
Generate id/sample statistics
}
\keyword{internal}
