% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{generate_stats}
\alias{generate_stats}
\title{generate_stats}
\usage{
generate_stats(
  gds,
  individuals = TRUE,
  markers = TRUE,
  missing = TRUE,
  heterozygosity = TRUE,
  coverage = TRUE,
  allele.coverage = TRUE,
  mac = TRUE,
  snp.position.read = TRUE,
  snp.per.locus = TRUE,
  subsample = NULL,
  exhaustive = TRUE,
  force.stats = TRUE,
  path.folder = NULL,
  plot = TRUE,
  digits = 6,
  file.date = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE
)
}
\description{
Generate individuals and markers statistics
}
\keyword{internal}
