% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{generate_strata}
\alias{generate_strata}
\title{Generate strata object from the data}
\usage{
generate_strata(data, pop.id = FALSE)
}
\arguments{
\item{data}{A tidy dataset object.
Documented in \code{\link[radiator]{tidy_genomic_data}}.}

\item{pop.id}{(logical) When \code{pop.id = TRUE}, the strata returns
the stratification colname \code{POP_ID}.
Default: \code{pop.id = FALSE}, returns \code{STRATA}.}
}
\description{
Generate a strata object from the data.
The function uses the \code{POP_ID} or \code{STRATA} columns along the
\code{INDIVIDUALS}.
}
\seealso{
\code{\link{read_strata}}, \code{\link{summary_strata}},
\code{\link{individuals2strata}}, \code{\link{change_pop_names}},
\code{\link{join_strata}}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
