% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_alleles.R
\name{generate_vcf_nuc}
\alias{generate_vcf_nuc}
\title{generate_vcf_nuc}
\usage{
generate_vcf_nuc(x, parallel.core = parallel::detectCores() - 1)
}
\description{
Generate GT_VCF_NUC field from REF/ALT and GT_BIN columns
}
\keyword{internal}
