% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_whitelist.R
\name{generate_whitelist}
\alias{generate_whitelist}
\title{generate_whitelist}
\usage{
generate_whitelist(x, t, path.folder = NULL)
}
\description{
Generate Whitelist of markers automatically for different threshold
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
