% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_genomic_data.R
\name{gt_haplo2gt_vcf_nuc}
\alias{gt_haplo2gt_vcf_nuc}
\title{gt_haplo2gt_vcf_nuc}
\usage{
gt_haplo2gt_vcf_nuc(x)
}
\description{
to apply the conversion of GT haplotype coding to gt_vcf_nuc in parallel
}
\keyword{internal}
