% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gt_recoding}
\alias{gt_recoding}
\title{gt_recoding}
\usage{
gt_recoding(
  x,
  gt = TRUE,
  gt.bin = TRUE,
  gt.vcf = TRUE,
  gt.vcf.nuc = TRUE,
  arrange = TRUE
)
}
\description{
Detect the genotype format used in the data set. Sample one if than one is present.
}
\keyword{internal}
