% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_reconstruction.R
\name{haplotype_reconstruction}
\alias{haplotype_reconstruction}
\title{haplotype_reconstruction}
\usage{
haplotype_reconstruction(data, parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\description{
Reconstruct haplotypes
}
\keyword{internal}
