% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_hwe.R
\name{hwe_analysis}
\alias{hwe_analysis}
\title{hwe_analysis}
\usage{
hwe_analysis(x, parallel.core = parallel::detectCores() - 1)
}
\description{
main hwe function
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
