% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdg_fh.R
\name{ibdg_fh}
\alias{ibdg_fh}
\title{FH measure of IBDg}
\usage{
ibdg_fh(data, path.folder = NULL, verbose = TRUE, ...)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{path.folder}{(path, optional) By default will print results in the working directory.
Default: \code{path.folder = NULL}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) To pass further arguments for fine-tuning the function.}
}
\value{
A list is created with 6 objects (function call, tables, manhattan,
boxplot and distribution plot).
FH measure is on average negative when the parents are less related than
expected by random mating. The distribution \code{fh.distribution.plot}
should be centered around 0 in samples of non-inbred individuals.
The first table, \code{$fh}, gives the individual's value
while the second table, \code{$fh.stats}, show the population and overall averaged.
}
\description{
FH is a proxy mesure of IBDg based on the excess in the observed
number of homozygous genotypes within an individual,
relative to the mean number of homozygous genotypes expected under random mating
(Keller et al., 2011; Kardos et al., 2015; Hedrick & Garcia-Dorado, 2016).

\strong{IBDg} is the realized proportion of the individual genome
that is identical by descent by reference to the current population
under hypothetical random mating
(Keller et al., 2011; Kardos et al., 2015; Hedrick & Garcia-Dorado, 2016).

This function is using a modified version of the FH measure
(constructed using \href{https://www.cog-genomics.org/plink2}{PLINK}
\code{-het} option) described in (Keller et al., 2011; Kardos et al., 2015).

The novelties are:

\itemize{
\item \strong{population-wise:} the individual's observed homozygosity is
contrasted against the expected homozygosity.
Two estimates of the expected homozygosity are provided based
on the population and/or the overall expected homozygosity
averaged across markers.
\item \strong{tailored for RADseq:} instead of using the overall number
of markers, the population and the overall expected homozygosity
are averaged with the same markers the individual's are genotyped for.
This reduces the bias potentially introduced by comparing the individual's
observed homozygosity (computed from non-missing genotypes) with
an estimate computed with more markers found at the population or at the
overall level.
}

The FH measure is also computed in
\href{https://github.com/thierrygosselin/stackr}{stackr} \emph{summary_haplotypes}
function and \href{https://github.com/thierrygosselin/grur}{grur}
\emph{missing_visualization} functions.
See theory below for the equations.
}
\section{Theory}{


\strong{Modified FH:}
\deqn{F_{h_i} = \frac{\overline{Het}_{obs_{ij}} - \overline{Het}_{exp_j}}{\sum_{i}snp_{ij} - \overline{Het}_{exp_j}}}

\strong{Individual Observed Heterozygosity averaged across markers:}
\deqn{\overline{Het}_{obs_i} = \frac{\sum_iHet_{obs_i}}{\sum_i{snp_i}}}

\strong{Population expected Heterozygosity (under Hardy-Weinberg) and
tailored by averaging for each individual using his genotyped markers:}
\deqn{\overline{Het}_{exp_j} = \frac{\sum_jHet_{exp_j}}{\sum_j{snp_j}}}
}

\section{Advance mode}{


\emph{dots-dots-dots ...} allows to pass several arguments for fine-tuning the function.
These arguments are described in \code{\link{tidy_genomic_data}}.
\itemize{
\item filter.monomorphic = TRUE (default)
\item filter.common.markers = FALSE. The argument for common markers
between populations is set by default to maximize genome coverage of
individuals and populations.
}
}

\examples{
\dontrun{
# Using a  VCF file, the simplest for of the function:
fh <- ibdg_fh(data = "sturgeon.gds")

# To see what's inside the list
names(fh)

# To view the boxplot:
fh$fh.boxplot

# To view the distribution of FH values:
fh$fh.distribution.plot
}
}
\references{
Keller MC, Visscher PM, Goddard ME (2011)
Quantification of inbreeding due to distant ancestors and its detection
 using dense single nucleotide polymorphism data. Genetics, 189, 237–249.

Kardos M, Luikart G, Allendorf FW (2015)
Measuring individual inbreeding in the age of genomics: marker-based
measures are better than pedigrees. Heredity, 115, 63–72.

Hedrick PW, Garcia-Dorado A. (2016)
Understanding Inbreeding Depression, Purging, and Genetic Rescue.
Trends in Ecology and Evolution. 2016; 31: 940-952.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
