% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{import_dart_geno}
\alias{import_dart_geno}
\title{import_dart_geno}
\usage{
import_dart_geno(strata.split, variant.id, data, dart.check, parallel.core)
}
\description{
Import DArt genotypes
}
\keyword{internal}
