% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_alleles.R
\name{integrate_ref}
\alias{integrate_ref}
\title{integrate_ref}
\usage{
integrate_ref(x, conversion.data = NULL, biallelic = TRUE,
  parallel.core = parallel::detectCores() - 1)
}
\description{
Integrate ref allele in original dataset
}
\keyword{internal}
