% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interactive_question}
\alias{interactive_question}
\title{interactive_question}
\usage{
interactive_question(x, answer.opt = NULL, minmax = NULL)
}
\description{
Ask to enter a word or number
}
\keyword{internal}
