% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_rad}
\alias{join_rad}
\title{join_rad}
\usage{
join_rad(x, s, m, g, env.arg = NULL)
}
\arguments{
\item{x}{The data}

\item{s}{The strata metadata.}

\item{m}{The markers metadata.}

\item{g}{The genotypes metadata.}

\item{env.arg}{You want to redirect \code{rlang::current_env()} to this argument.}
}
\description{
Join  back the parts stripped in strip_rad. Used internally.
}
