% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{join_strata}
\alias{join_strata}
\title{Join the strata with the data}
\usage{
join_strata(data, strata = NULL, pop.id = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{A tidy dataset object.
Documented in \code{\link[radiator]{tidy_genomic_data}}.}

\item{strata}{(path or object) The strata file or object.
Additional documentation is available in \code{\link{read_strata}}.
Use that function to whitelist/blacklist populations/individuals.
Option to set \code{pop.levels/pop.labels} is also available.}

\item{pop.id}{(logical) When \code{pop.id = TRUE}, the strata returns
the stratification colname \code{POP_ID}.
Default: \code{pop.id = FALSE}, returns \code{STRATA}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\value{
The data filtered by the strata by individuals.
}
\description{
The function first filters individuals in data then include the
strata.
}
\examples{
\dontrun{
data <- radiator::join_strata(
    data = my_tidy_dataset_object,
    strata = my_strata_object)
}
}
\seealso{
\code{\link{read_strata}}, \code{\link{summary_strata}},
\code{\link{individuals2strata}}, \code{\link{change_pop_names}},
\code{\link{generate_strata}}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
