% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ld.R
\name{ld_boxplot}
\alias{ld_boxplot}
\title{ld_boxplot}
\usage{
ld_boxplot(
  gds,
  subsample.markers.stats = 0.2,
  ld.method = "r2",
  path.folder = NULL,
  parallel.core = parallel::detectCores() - 2,
  verbose = TRUE
)
}
\description{
Generate a boxplot of LD using subsampled markers.
Return the matrix of LD.
Write to directory the stats and boxplot.
}
\keyword{internal}
