% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_long_distance.R
\name{ld_long_distance}
\alias{ld_long_distance}
\title{Generate SNP long distance Linkage Disequilibrium statistics}
\usage{
ld_long_distance(data, ld.threshold = NULL, ldna = FALSE,
  filename = NULL, parallel.core = parallel::detectCores() - 1, ...)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.
\strong{The genotypes are biallelic.}}

\item{ld.threshold}{(optional, double) The threshold to prune SNP based on
Long Distance Threshold.
Default: \code{ld.threshold = NULL}.}

\item{ldna}{(optional, logical) Output an LDna lower matrix object.
Default: \code{ldna = FALSE}.}

\item{filename}{(optional) TTO DO
he file name of the LDna lower matrix file.
Radiator will append \code{.ldna.rds} to the filename.
If filename chosen is already present in the
working directory, the default \code{radiator_datetime.ldna.rds} is chosen.
With default, \code{filename = NULL}, no file is generated, only an object in
the Global Environment.
To read the data back into R, use readRDS("filename.ldna.rds").}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel::detectCores() - 1}.}

\item{...}{(optional) To pass further argument for fine-tuning the
function (see details).}
}
\description{
Generate SNP long distance Linkage Disequilibrium statistics
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
The function requires \href{https://github.com/zhengxwen/SNPRelate}{SNPRelate}

To install SNPRelate:
source("https://bioconductor.org/biocLite.R")
biocLite("SNPRelate").
#'

\strong{Further arguments passed via the \emph{dots-dots-dots}:}
\itemize{
\item keep.gds Default \code{keep.gds = FALSE}, the SNPRelate GDS object
generated is removed after completion of the LDna object.
}
}
\references{
Zheng X, Levine D, Shen J, Gogarten SM, Laurie C, Weir BS.
(2012) A high-performance computing toolset for relatedness and principal component
analysis of SNP data. Bioinformatics. 28: 3326-3328.
doi:10.1093/bioinformatics/bts606
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
