% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ld.R
\name{ld_pruning}
\alias{ld_pruning}
\title{Prune dataset based on LD.}
\usage{
ld_pruning(ld.tibble = NULL, stats = NULL, ld.threshold = 0.8,
  verbose = TRUE)
}
\arguments{
\item{ld.tibble}{(path) The markers LD pairwise data.
Default: \code{ld.tibble = NULL}.}

\item{stats}{(path) The markers missingness info statistics.
Default: \code{stats = NULL}.}

\item{ld.threshold}{(double) The threshold to prune SNPs in LD.
Default: \code{ld.threshold = 0.8}.}

\item{verbose}{(logical, optional) Default: \code{verbose = TRUE}.}
}
\value{
A list with blacklisted SNPs.  Write the blacklist in the working
directory.
}
\description{
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
Prune dataset based on LD.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
