% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{list_filters}
\alias{list_filters}
\title{List current active filters (individuals and markers) in radiator GDS object.}
\usage{
list_filters(gds)
}
\arguments{
\item{gds}{(2 options) A Genomic Data Structure (GDS) file or object

\emph{How to get GDS ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}},
\code{\link{tidy_vcf}} or \code{\link{write_gds}}.}

\item{list.filters}{filters (logical, optional) List current active filters for individuals and markers.
Default: \code{list.filters = TRUE}.}
}
\description{
List current active filters (individuals and markers) in radiator GDS object.
}
\examples{
\dontrun{
# List active filters for individuals and markers
list_filters(gds)
}
}
\seealso{
\code{\link{sync_gds}}, \code{\link{list_filters}}.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
