% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{markers_genotyped_helper}
\alias{markers_genotyped_helper}
\title{markers_genotyped_helper}
\usage{
markers_genotyped_helper(x, y, overall.only = FALSE)
}
\description{
Help individual's genotyped threshold
}
\keyword{internal}
