% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_parallel.R
\name{mclapply_win}
\alias{mclapply_win}
\title{hack to switch function for parallel computation based on OS}
\usage{
mclapply_win(
  X,
  FUN,
  ...,
  mc.cores = getOption("mc.cores", 2L),
  mc.preschedule = TRUE,
  mc.set.seed = TRUE,
  mc.cleanup = TRUE,
  mc.allow.recursive = TRUE
)
}
\description{
\code{parallel::mclapply} doesn't work on Windows,
because forking is not supported.
This function defines a socket version of mclapply for windows computer
An implementation that switch automatically the parallel process when detecting
the os.
The code below was inspired from
\pkg{parallel} \code{\link{mclapply}},
\href{https://github.com/nathanvan}{Nathan VanHoudnos},
\href{https://github.com/kvnkuang/pbmcapply}{Kevin Kuang},
\href{https://github.com/psolymos/pbapply}{Peter Solymos} and
\href{https://github.com/EricArcher/}{Eric Archer}.
}
\keyword{internal}
