% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_common_markers.R
\name{not_common_markers}
\alias{not_common_markers}
\title{not_common_markers}
\usage{
not_common_markers(x, strata, parallel.core = parallel::detectCores() -
  2)
}
\description{
Generate a blacklist of markers not in common
}
\keyword{internal}
