% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_parallel.R
\name{parallel_core_opt}
\alias{parallel_core_opt}
\title{parallel_core_opt}
\usage{
parallel_core_opt(parallel.core = NULL, max.core = NULL)
}
\description{
Optimization of parallel core argument for radiator
}
\keyword{internal}
