% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{pi}
\alias{pi}
\title{Nucleotide diversity}
\usage{
pi(data, strata, read.length, parallel.core = parallel::detectCores() -
  1, ...)
}
\arguments{
\item{data}{12 options for input: VCFs (SNPs or Haplotypes,
to make the vcf population ready),
plink, stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop, DArT,
and a data frame in long/tidy or wide format. To verify that radiator detect
your file format use \code{\link{detect_genomic_format}} (see example below).
Documented in \strong{Input genomic datasets} of \code{\link{tidy_genomic_data}}.}

\item{strata}{(optional)
The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}. Documented in \code{\link{read_strata}}.
Default: \code{strata = NULL}.}

\item{read.length}{(number) The length in nucleotide of your reads
(e.g. \code{read.length = 100}).}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{...}{(optional) To pass further arguments for fine-tuning the function.}
}
\value{
The function returns a list with the function call and:
\enumerate{
\item $pi.individuals: the pi estimated for each individual
\item $pi.populations: the pi statistics estimated per populations and overall.
\item $boxplot.pi: showing the boxplot of Pi for each populations and overall.
}
use $ to access each #' objects in the list.
}
\description{
Calculates the nucleotide diversity (Nei & Li, 1979).

To get an estimate with the consensus reads, use the
function \emph{summary_haplotypes} found in the package
[stackr](https://github.com/thierrygosselin/stackr). The estimate in
\emph{summary_haplotypes} integrates the consensus markers found in
[STACKS](http://catchenlab.life.illinois.edu/stacks/)
populations.haplotypes.tsv file.
Both radiator and stackr functions requires \code{stringdist} package.

The \code{read.length} argument below is used directly in the calculations.
To be correctly estimated, the reads obviously need to be of identical size...
}
\examples{
\dontrun{
require(stringdist)
# The simplest way to run the function:
sum <- radiator::pi(
data = "batch_1.vcf",
strata = "strata_brook_charr.tsv",
read.length = 90)
}
}
\references{
Nei M, Li WH (1979)
Mathematical model for studying genetic variation in terms of
restriction endonucleases.
Proceedings of the National Academy of Sciences of
the United States of America, 76, 5269–5273.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
