% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{pi_pop}
\alias{pi_pop}
\title{pi_pop}
\usage{
pi_pop(data, read.length, parallel.core = parallel::detectCores() - 1)
}
\description{
pi_pop function
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
