% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_boxplot_coverage}
\alias{plot_boxplot_coverage}
\title{Figure box plot of coverage summary statistics}
\usage{
plot_boxplot_coverage(data)
}
\arguments{
\item{data}{Coverage summary file.}
}
\description{
Create box plots of coverage summary statistics.
Use the coverage summary file created with coverage_summary function.
}
\keyword{internal}
