% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_density_distribution_het}
\alias{plot_density_distribution_het}
\title{Figure density distribution of the observed heterozygosity
summary statistics.}
\usage{
plot_density_distribution_het(
  data,
  pop.levels,
  het.group,
  aes.colour,
  adjust.bin,
  x.title
)
}
\arguments{
\item{data}{sumstats or tidy vcf files.}

\item{pop.levels}{Character string defining your ordered populations.}

\item{het.group}{= aes(x = HET_MAX, na.rm = F)}

\item{aes.colour}{GGPLOT2 aesthetics colour,
e.g. aes(y = ..scaled.., color = GROUP).}

\item{adjust.bin}{Adjust GGPLOT2 bin size (0 to 1).}

\item{x.title}{Title of the x-axis.}
}
\description{
Create density distribution of the observed heterozygosity
summary statistics.
}
\keyword{internal}
