% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_distribution_diversity}
\alias{plot_distribution_diversity}
\title{Density distribution of diversity (Gene and Haplotypes)}
\usage{
plot_distribution_diversity(data, aes.x, aes.colour, x.title, y.title)
}
\arguments{
\item{data}{The hapstats summary file or object.}

\item{aes.x}{GGPLOT2 aesthetics,
e.g. aes.x = aes(x = GENE_DIVERSITY, na.rm = T).}

\item{aes.colour}{GGPLOT2 aesthetics colour,
e.g. aes.colour = aes(y = ..scaled.., colour = POP_ID).}

\item{x.title}{Title of the x-axis.}

\item{y.title}{Title of the y-axis.}
}
\description{
GGPLOT2 Density distribution of diversity (Gene and Haplotypes).
}
\keyword{internal}
