% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_het_outliers.R
\name{plot_het_outliers}
\alias{plot_het_outliers}
\title{plot_het_outliers}
\usage{
plot_het_outliers(data, path.folder = NULL)
}
\description{
Function that calculates genotypes observed and expected frequencies
returns a tibble with all the summary info and a plot with boundaries.
}
\author{
Eric Anderson \email{eric.anderson@noaa.gov} and
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
