% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_snp_number_loci}
\alias{plot_snp_number_loci}
\title{Figure of the distribution of SNP per locus before and after filters}
\usage{
plot_snp_number_loci(before.filter.data, after.filter.data)
}
\arguments{
\item{before.filter.data}{Data set before filter.}

\item{after.filter.data}{Data set after filter.}
}
\description{
Create density distribution of SNP per locus before and after filters.
}
\keyword{internal}
