% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_alleles.R
\name{private_alleles}
\alias{private_alleles}
\title{Find private alleles}
\usage{
private_alleles(data, strata = NULL, verbose = TRUE)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{write_seqarray}} or
\code{\link{tidy_vcf}}.}

\item{strata}{(path or object) The strata file or object.
Additional documentation is available in \code{\link{read_strata}}.
Use that function to whitelist/blacklist populations/individuals.
Option to set \code{pop.levels/pop.labels} is also available.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\value{
A list with an object highlighting private alleles by markers and strata and
a second object with private alleles summarized by strata.
}
\description{
The function highlight private alleles by strata, using a GDS or tidy file or object.
}
\examples{
\dontrun{
corals.private.alleles.by.pop <- radiator::private_alleles(data = tidy, strata = strata.pop)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
