% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_haplotypes.R
\name{private_haplotypes}
\alias{private_haplotypes}
\title{private haplotypes}
\usage{
private_haplotypes(data, strata = NULL, verbose = TRUE)
}
\arguments{
\item{data}{A tidy genomic dataframe
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{strata}{(optional) A strata file or object in the global environment.
The strata is a tab-separated data frame with 2 columns: \code{INDIVIDUALS} and
\code{STRATA}. If used, the strata will replace the current STRATA or POP_ID in the
dataset. Use this argument if you want to find private haplotypes on another
hierarchical level, other than POP_ID.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\value{
A list with private haplotypes per markers and strata and a summary of
overall number of private haplotypes per strata.
}
\description{
Minor Allele Frequency filter.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
