% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_imputations_module.R
\name{rad_decoding_haplotypes}
\alias{rad_decoding_haplotypes}
\title{rad_decoding_haplotypes}
\usage{
rad_decoding_haplotypes(data = NULL, parallel.core = parallel::detectCores()
  - 1)
}
\description{
separate snp group merged with rad_encoding_snp
}
\keyword{internal}
