% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_imputations_module.R
\name{rad_encoding_snp}
\alias{rad_encoding_snp}
\title{rad_encoding_snp}
\usage{
rad_encoding_snp(locus.list = NULL, data = NULL)
}
\description{
bind snp found on the same locus
}
\keyword{internal}
