% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_imputations_module.R
\name{rad_factorize_gt}
\alias{rad_factorize_gt}
\title{rad_factorize_gt}
\usage{
rad_factorize_gt(x)
}
\description{
Necessary to factorize by markers in tidy format.
XGBoost needs numbering to start at , hence the codes below.
}
\keyword{internal}
