% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rad_folder}
\alias{rad_folder}
\title{rad_folder}
\usage{
rad_folder(f, path.folder = NULL)
}
\arguments{
\item{f}{Folder name}

\item{path.folder}{path of the folder}
}
\description{
Generate the rad folders
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
