% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_imputations_module.R
\name{rad_impute_genotypes}
\alias{rad_impute_genotypes}
\title{rad_impute_genotypes}
\usage{
rad_impute_genotypes(marker.list, data, data.na, data.gl = NULL, data.imp,
  num.tree = 100, pred.mean.matching = 0, random.seed = NULL,
  parallel.core = parallel::detectCores() - 1,
  hierarchical.levels = "strata", pred.error = pred.error)
}
\description{
imputations using Ranger package and predictive mean matching of missRanger
}
\keyword{internal}
