% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rad_long}
\alias{rad_long}
\title{rad_long}
\usage{
rad_long(
  x,
  cols = NULL,
  measure_vars = NULL,
  names_to = NULL,
  values_to = NULL,
  variable_factor = TRUE,
  keep_rownames = FALSE,
  tidy = FALSE
)
}
\description{
Gather, melt and pivot_longer
}
\keyword{internal}
