% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rad_wide}
\alias{rad_wide}
\title{rad_wide}
\usage{
rad_wide(
  x,
  formula = NULL,
  names_from = NULL,
  values_from = NULL,
  values_fill = NULL,
  sep = "_",
  tidy = FALSE
)
}
\description{
Spread, dcast and pivot_wider
}
\keyword{internal}
