% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_imputations_module.R
\name{radiator_boost_imputer}
\alias{radiator_boost_imputer}
\title{radiator_boost_imputer}
\usage{
radiator_boost_imputer(marker.list = NULL, input.wide = NULL,
  gl.wide = NULL, eta = 0.2, gamma = 0, max_depth = 6,
  min_child_weight = 1, subsample = 0.8, colsample_bytree = 1,
  num_parallel_tree = 1, parallel.core = 1, nrounds = 200,
  early_stopping_rounds = 20, save_name = "imputation.model.temp",
  hierarchical.levels = "strata")
}
\description{
imputations using Ranger package and predictive mean matching
}
\keyword{internal}
