% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator.R
\name{radiator_common_arguments}
\alias{radiator_common_arguments}
\title{radiator common arguments}
\usage{
radiator_common_arguments(
  interactive.filter = TRUE,
  gds,
  data,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE,
  random.seed = NULL,
  ...
)
}
\arguments{
\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. Figures of distribution are shown before asking for filtering
thresholds.
Default: \code{interactive.filter = TRUE}.}

\item{gds}{(2 options) A Genomic Data Structure (GDS) file or object

\emph{How to get GDS ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}},
\code{\link{tidy_vcf}} or \code{\link{write_gds}}.}

\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{random.seed}{(integer, optional) For reproducibility, set an integer
that will be used inside the function that requires randomness. With default,
a random number is generated and printed in the appropriate output.
Default: \code{random.seed = NULL}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function. Also used for legacy arguments (see details or
special section)}
}
\description{
radiator common arguments
}
\keyword{internal}
