% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator.R
\name{radiator_folder}
\alias{radiator_folder}
\title{radiator_folder}
\usage{
radiator_folder(rad.folder, path.folder = NULL, prefix.int = TRUE)
}
\arguments{
\item{path.folder}{path of the folder}

\item{prefix.int}{Use an integer prefix padded left with 0.
Default: \code{prefix.int = TRUE}.}
}
\description{
Generate the rad folders
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
