% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator.R
\name{radiator_function_header}
\alias{radiator_function_header}
\title{radiator_function_header}
\usage{
radiator_function_header(f.name = NULL, start = TRUE, verbose = TRUE)
}
\description{
Generate function header
}
\keyword{internal}
