% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_future.R
\name{radiator_future}
\alias{radiator_future}
\title{radiator parallel function}
\usage{
radiator_future(
  X,
  FUN,
  parallel.core = parallel::detectCores() - 1,
  split.tibble = NULL,
  bind.rows = FALSE,
  flatten.int = FALSE,
  ...
)
}
\arguments{
\item{X}{an array, including a matrix.}

\item{FUN}{A function taking at least one argument.}
}
\description{
Updating radiator to use future
}
\keyword{internal}
