% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{radiator_gds}
\alias{radiator_gds}
\title{radiator gds constructor}
\usage{
radiator_gds(genotypes.df, strata = NULL, biallelic = TRUE,
  markers.meta = NULL, genotypes.meta = NULL, dp = NULL, ad = NULL,
  filename = NULL, source = NULL, open = FALSE, verbose = TRUE)
}
\description{
helper function to construct a radiator gds
}
\keyword{internal}
