% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_seqarray.R
\name{radiator_gds_markers_meta}
\alias{radiator_gds_markers_meta}
\title{radiator_gds_markers_meta}
\usage{
radiator_gds_markers_meta(gds, path = "radiator/markers.meta/MARKERS")
}
\description{
Get the markers meta node in radiator GDS
}
\keyword{internal}
