% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_imputations_module.R
\name{radiator_imputer}
\alias{radiator_imputer}
\title{radiator_imputer}
\usage{
radiator_imputer(data, num.tree = 100, pred.mean.matching = 0,
  random.seed = NULL, parallel.core = parallel::detectCores() - 1,
  hierarchical.levels = "strata", marker.list = marker.list,
  verbose = verbose)
}
\description{
imputations using Ranger package and predictive mean matching
}
\keyword{internal}
