% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator.R
\name{radiator_packages_dep}
\alias{radiator_packages_dep}
\title{radiator_packages_dep}
\usage{
radiator_packages_dep(package, cran = TRUE, bioc = FALSE)
}
\description{
Verify required packages
}
\keyword{internal}
