% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_parallel.R
\name{radiator_parallel}
\alias{radiator_parallel}
\title{Enable parallel execution with pbmcapply}
\usage{
radiator_parallel(
  X,
  FUN,
  ...,
  mc.style = "ETA",
  mc.substyle = NA,
  mc.cores = getOption("mc.cores", 2L),
  ignore.interactive = getOption("ignore.interactive", F),
  mc.preschedule = TRUE,
  mc.set.seed = TRUE,
  mc.cleanup = TRUE,
  mc.allow.recursive = TRUE
)
}
\description{
Internal hack to enable parallel execution
on all OS with pbmclapply and it's progress bar.
}
\keyword{internal}
