% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_parallel.R
\name{radiator_parallel_mc}
\alias{radiator_parallel_mc}
\title{functions on Windows.}
\usage{
radiator_parallel_mc(
  X,
  FUN,
  ...,
  mc.preschedule = TRUE,
  mc.set.seed = TRUE,
  mc.silent = FALSE,
  mc.cores = getOption("mc.cores", 2L),
  mc.cleanup = TRUE,
  mc.allow.recursive = TRUE,
  affinity.list = NULL
)
}
\description{
functions on Windows.
}
\keyword{internal}
