% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_pkg_install.R
\name{radiator_pkg_install}
\alias{radiator_pkg_install}
\title{radiator packages install helper}
\usage{
radiator_pkg_install(check = TRUE, minimal.install = FALSE)
}
\arguments{
\item{check}{(logical, optional) Check the packages installed and system info.
Default: \code{check = TRUE}}

\item{minimal.install}{(logical, optional) The minimal install does nothing,
and the function just check what's installed.
Default: \code{check = FALSE}, does the full installation of packages you might
need using radiator.}
}
\description{
This function help to install packages for the full suits of
radiator functions or the minimal install to lower packages dependencies installation.
It also re-install the latest version of radiator when available.
}
