% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator.R
\name{radiator_results_message}
\alias{radiator_results_message}
\title{radiator_results_message}
\usage{
radiator_results_message(rad.message = NULL, filters.parameters,
  internal = FALSE, verbose = TRUE)
}
\description{
Message printed at the end of most radiator functions
}
\keyword{internal}
