% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{radiator_split_tibble}
\alias{radiator_split_tibble}
\title{radiator_split_tibble}
\usage{
radiator_split_tibble(x, parallel.core = parallel::detectCores() - 1)
}
\description{
radiator alternative to data.table::tstrsplit
}
\keyword{internal}
