% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator.R
\name{radiator_toc}
\alias{radiator_toc}
\title{radiator_toc}
\usage{
radiator_toc(
  timing,
  end.message = "Computation time, overall:",
  verbose = TRUE
)
}
\description{
radiator tictoc function
}
\keyword{internal}
