% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{read_blacklist_id}
\alias{read_blacklist_id}
\title{read_blacklist_id}
\usage{
read_blacklist_id(blacklist.id = NULL, verbose = TRUE)
}
\arguments{
\item{blacklist.id}{(optional, path or object) A blacklist file in the working directory
or object in the global environment. The data frame
as 1 column (named \code{INDIVIDUALS}) and is filled with the individual IDs
The ids are cleaned with \code{\link{clean_ind_names}} for separators,
only \code{-} are tolerated. Duplicates are removed automatically.
Default: \code{blacklist.id = NULL}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\value{
A tibble with column \code{INDIVIDUALS}.
}
\description{
Read a file or object with blacklisted individuals.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\examples{
\dontrun{
bl <- radiator::read_blacklist_id("blacklist.tsv")
}
}
