% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{read_dart_markers_metadata}
\alias{read_dart_markers_metadata}
\title{read_dart_markers_metadata}
\usage{
read_dart_markers_metadata(
  data,
  whitelist.markers = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE
)
}
\description{
Read DArT markers/locus metadata.
}
\keyword{internal}
