% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{read_plink}
\alias{read_plink}
\title{Reads PLINK tped and bed files}
\usage{
read_plink(
  data,
  filename = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{The PLINK file.
\itemize{
\item {bi-allelic data only}. For haplotypes use VCF.
\item \code{tped} file format: the corresponding \code{tfam} file must be in the directory.
\item \code{bed} file format: IS THE PREFERRED format, the corresponding
\code{fam} and \code{bim} files must be in the directory.
}}

\item{filename}{(optional) The file name of the Genomic Data Structure (GDS) file.
radiator will append \code{.gds.rad} to the filename.
If the filename chosen exists in the working directory,
the default \code{radiator_datetime.gds} is chosen.
Default: \code{filename = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) To pass further arguments for fine-tuning the function.}
}
\value{
For \code{tped} the function returns a list object with the non-modified \code{tped} and
the strata corresponding to the \code{tfam}.
With \code{bed}, the function returns a GDS object.
}
\description{
The function reads PLINK tped and bed files.
radiator prefers the use of BED file. These files are converted to
a connection SeqArray \href{https://github.com/zhengxwen/SeqArray}{SeqArray}
GDS object/file of class \code{SeqVarGDSClass} (Zheng et al. 2017).
The Genomic Data Structure (GDS) file format is detailed in
\href{https://github.com/zhengxwen/gdsfmt}{gdsfmt}.

Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
Large PLINK files will require the use of BED plink format. Look below
in the example for conversion with PLINK.

Large PLINK bed files will take longer to import and transform in GDS, but
after the file is generated, you can close your computer and
come back to it a month later and it's now a matter of sec to open a connection!
}
\examples{
\dontrun{
data <- radiator::read_plink(data = "my_plink_file.bed")
# when conversion is required from TPED to BED, in Terminal:
# plink --tfile my_plink_file --make-bed --allow-no-sex --allow-extra-chr --chr-set 95
}
}
\references{
Zheng X, Gogarten S, Lawrence M, Stilp A, Conomos M, Weir BS,
Laurie C, Levine D (2017). SeqArray -- A storage-efficient high-performance
data format for WGS variant calls.
Bioinformatics.

PLINK: a tool set for whole-genome association and population-based linkage
analyses.
American Journal of Human Genetics. 2007: 81: 559–575. doi:10.1086/519795
}
\seealso{
\href{https://www.cog-genomics.org/plink/1.9/}{PLINK}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
