% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rad.R
\name{read_rad}
\alias{read_rad}
\title{Read tidy genomic data file ending .rad}
\usage{
read_rad(data)
}
\arguments{
\item{data}{A file in the working directory ending with .rad,
a tidy genomic data produced by radiator, assigner or grur.}
}
\value{
A tidy data frame in the global environment.
}
\description{
Fast read .rad file.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
