% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_whitelist.R
\name{read_whitelist}
\alias{read_whitelist}
\title{read whitelist of markers}
\usage{
read_whitelist(whitelist.markers = NULL, verbose = FALSE)
}
\arguments{
\item{whitelist.markers}{(path or object)
The whitelist is an object in your
global environment or a file in the working directory (e.g. "whitelist.txt").
The dataframe contains one, a combination
or all of these columns: \code{MARKERS, CHROM, LOCUS, POS}.
Columns are cleaned of separators that interfere with some packages or codes, detailed in
\code{\link{clean_markers_names}}.
Default \code{whitelist.markers = NULL}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\description{
Read a whitelist object or file.


Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
\strong{multi allelic datasets:}
Example: VCF with haplotypes containing \code{CHROM, LOCUS, POS} columns:
If the whitelist was not created from the same dataset,
the filtering could result in losing all the markers.
The POS column is different in biallelic and multiallelic file...
}
\examples{
\dontrun{
wl <- radiator::read_whitelist(data = data, whitelist.markers = "mywhitelist.tsv")
}
}
\seealso{
\code{\link{filter_whitelist}}.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
