% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_alt_alleles.R
\name{ref_alt_alleles}
\alias{ref_alt_alleles}
\title{**deprecated** use \code{\link[radiator]{change_alleles}}}
\usage{
ref_alt_alleles(data, monomorphic.out = TRUE,
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{data}{A biallelic genomic data set in the working directory or
object in the global environment in wide or long (tidy) formats.
See details for more info.}

\item{monomorphic.out}{(optional) Should the monomorphic
markers present in the dataset be filtered out ?
Default: \code{monomorphic.out = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during vcf import.
Default: \code{parallel::detectCores() - 1}.}
}
\value{
A tidy data frame with 6 columns:
\code{MARKERS, INDIVIDUALS, REF, ALT, GT_VCF, GT_BIN}.
\code{GT_VCF}: the genotype in VCF format
\code{GT_BIN}: coding used internally to easily convert to genlight,
the coding \code{0, 1, 2, NA} stands for the number of ALT allele in the
genotype and \code{NA} for missing genotype.
}
\description{
compute the REF and ALT alleles from a biallelic
genomic tidy data frame.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
\strong{Input data:}

To discriminate the long from the wide format,
the function \pkg{radiator} \code{\link[radiator]{tidy_wide}} searches
for \code{MARKERS or LOCUS} in column names (TRUE = long format).
The data frame is tab delimitted.
\strong{Wide format:}
The wide format cannot store metadata info.
The wide format starts with these 2 id columns:
\code{INDIVIDUALS}, \code{POP_ID} (that refers to any grouping of individuals),
the remaining columns are the markers in separate columns storing genotypes.

\strong{Long/Tidy format:}
The long format is considered to be a tidy data frame and can store metadata info.
(e.g. from a VCF see \pkg{radiator} \code{\link{tidy_genomic_data}}). A minimum of 4 columns
are required in the long format: \code{INDIVIDUALS}, \code{POP_ID},
\code{MARKERS or LOCUS} and \code{GENOTYPE or GT}. The rest are considered metata info.

\strong{2 genotypes formats are available:}
6 characters no separator: e.g. \code{001002 of 111333} (for heterozygote individual).
6 characters WITH separator: e.g. \code{001/002 of 111/333} (for heterozygote individual).
The separator can be any of these: \code{"/", ":", "_", "-", "."}.

\emph{How to get a tidy data frame ?}
\pkg{radiator} \code{\link{tidy_genomic_data}} can transform 6 genomic data formats
in a tidy data frame.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
