% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_alleles.R
\name{ref_dictionary}
\alias{ref_dictionary}
\title{ref_dictionary}
\usage{
ref_dictionary(x, parallel.core = parallel::detectCores() - 1)
}
\description{
Generate REF/ALT allele dictionary from data with/without nucleotides.
}
\keyword{internal}
