% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duplicate_genomes.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Read tidy genomic data file ending .rad}
\usage{
remove_duplicates(data = "individuals.pairwise.dist.tsv",
  stats = "genotyped.statistics.tsv", dup.threshold = 0.25,
  diff.pop.remove = TRUE, path.folder = NULL)
}
\arguments{
\item{data}{(path) The individual's pairwise data.
Default: \code{data = "individuals.pairwise.dist.tsv"}.}

\item{stats}{(path) The genotype statistics
Default: \code{stats = "genotyped.statistics.tsv"}.}

\item{dup.threshold}{(double) The threshold to filter out duplicates
Default: \code{dup.threshold = 0.25}.}

\item{diff.pop.remove}{Remove all individuals in pairs from different pop.
Both samples are potentially problems. With defautl, the function will not keep
one sample in the duplicate pair.
Default: \code{diff.pop.remove = TRUE}.}
}
\value{
A list with blacklisted duplicates. Write the blacklist in the working
directory.
}
\description{
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
To remove duplicate individuals based on threshold established from the
visualization figures.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
