% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{separate_gt}
\alias{separate_gt}
\title{separate_gt}
\usage{
separate_gt(
  x,
  gt = "GT_VCF_NUC",
  gather = TRUE,
  haplotypes = FALSE,
  exclude = c("LOCUS", "INDIVIDUALS", "POP_ID"),
  alleles.naming = c("A1", "A2"),
  remove = TRUE,
  filter.missing = FALSE,
  split.chunks = 3,
  parallel.core = parallel::detectCores() - 1
)
}
\description{
Separate genotype field
}
\keyword{internal}
