% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_markers.R
\name{separate_markers}
\alias{separate_markers}
\title{Separate markers column into chrom, locus and pos}
\usage{
separate_markers(data, sep = "__")
}
\arguments{
\item{data}{An object with a column named \code{MARKERS}.
If \code{CHROM}, \code{LOCUS}, \code{POS} are already present,
the columns are removed before separating the \code{MARKERS} column.
The data can be whitelists and blacklists of markers or datasets.}

\item{sep}{(optional, character) Separator used to identify the different
field in the \code{MARKERS} column.
Default: \code{sep = "__"}.}
}
\value{
The same data in the global environment, with 3 new columns:
\code{CHROM}, \code{LOCUS}, \code{POS}
}
\description{
Radiator uses unique marker names by combining
\code{CHROM}, \code{LOCUS}, \code{POS} columns, with double underscore
separators, into \code{MARKERS = CHROM__LOCUS__POS}. All the info is kept most
of the time, but in case user need to get back to the original metadata the
function provides an easy wau to do it.
}
\examples{
\dontrun{
whitelist <- radiator::separate_markers(data = whitelist.markers)
tidy.data <- radiator::separate_markers(data = bluefintuna.data)
}#' @export
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
