% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_markers.R
\name{sexy_markers}
\alias{sexy_markers}
\title{This function identifies sex-linked SNPs}
\usage{
sexy_markers(data, silicodata = NULL, strata = NULL,
  coverage.thresholds = 1, residual.threshold = NULL, filters = TRUE,
  interactive.filter = TRUE, parallel.core = parallel::detectCores() -
  1, ...)
}
\arguments{
\item{data}{(object or file) DArT file \code{.csv or .tsv}, VCF file \code{.vcf},
GDS file or object (\code{.gds} or \code{.gds.rad}).
See Input data section data for more details.}

\item{silicodata}{(optional, file) A silico DArT file (.csv or .tsv). This can be
count or genotyped data.
Default: \code{silicodata = NULL}}

\item{strata}{(file) A tab delimited file with a minimum of
2 columns \code{INDIVIDUALS, STRATA} for VCF files and 3 columns for DArT files
\code{TARGET_ID, INDIVIDUALS, STRATA}.
\itemize{
\item \code{TARGET_ID:} it's the column header of the DArT file.
\item \code{STRATA:} is the grouping column, here the sex info.
3 values: \code{M} for male, \code{F} for female and \code{U} for unknown.
Anything else is converted to \code{U}.
\item \code{INDIVIDUALS:} Is how you want your samples to be named.
}
Default: \code{strata = NULL}, the function will look for sex info in the
tidy data or GDS data (individuals.meta section).
You can easily build the strata file by starting with the output of these
functions: \code{\link{extract_dart_target_id}} and \code{\link{extract_individuals_vcf}}}

\item{coverage.thresholds}{(optional, integer) The minimum coverage required
to account for the genotype.
Default: \code{coverage.thresholds = 1}.}

\item{residual.threshold}{(optional, double)
Default: \code{residual.threshold = NULL}.}

\item{filters}{(optional, logical) When \code{filters = TRUE}, the data will
be filtered for
monomorphic loci, missingness of individuals and heterozygosity of individuals.
CAUTION: we advice to use these filter, since not filtering or filtering too
stingently will results in
false positve or false negative detections.
Default: \code{filters = TRUE}.}

\item{interactive.filter}{(optional, logical)
Default: \code{interactive.filter = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function. Also used for legacy arguments (see details or
special section)}
}
\value{
A list with ...
}
\description{
This function identifies sex-linked SNPs putatively located on
homogametic or heterogametic chromosomes. Datasets accepted: RADseq and
DArT's silico dart.
}
\details{
This function takes DArT and RAD data to find markers that have a specific
 pattern that is linked to sex.
The function hypothesizes the presence of sex-chromosomes in you
species/population. The tests are designed to identify markers that are
located on putative heterogametic (Y or W) or homogametic (X or Z) chromosomes.
\emph{Note:} If no sex-linked markers are detected, this could be
explained by the current data not meeting the assumptions or prerequisites
(see below), instead of the absence of sex chromosomes in the species/data.
}
\section{Assumptions}{

\enumerate{
\item \strong{Genetic Sex Determination System} over a e.g. environmental-sex-determination system.
\item \strong{Genome coverage:} restriction sites randomly spread throughout the whole genome.
\item \strong{Mutations:} Processes such as sex-specific mutations
in the restriction sites could lead to false positive results.
\item \strong{Deletions/duplications:} Processes such as
sex-specific deletions or duplications could lead to false positive results.
\item \strong{Homology:} The existence of homologous sequences between the
homogametic and heterogametic chromosomes could lead to false negative
results.
\item \strong{Absence of population signal}
}
}

\section{Prerequisites}{

\enumerate{
\item Ideally, the data must have enough individuals (n > 100).
\item Start with a dataset with equal ratio.
\item DArT data: if the minimum call rate is 0.5 ask DArT to lower their filtering threshold.
\item RADseq data: lower markers missingness thresholds during filtering (e.g. stacks \code{r} and \code{p}).
\item Low genotyping error rate (see \code{\link{detect_het_outliers}} and \href{https://github.com/eriqande/whoa}{whoa}).
\item Low heterozygosity miscall rate (see \code{\link{detect_het_outliers}} and \href{https://github.com/eriqande/whoa}{whoa}).
\item Absence of pattern of heterozygosity driven by missingness \code{\link{detect_mixed_genomes}}.
\item Absence of paralogous sequences in the data.
}
}

\section{Sex methods}{

\strong{Heterogametic sex-markers:}
\itemize{
\item \strong{Presence/Absence method}: To identify markers on Y or W chromosomes,
we look at the presence or
absence of a marker between females and males. More specifically, if a marker
is always present in males but never in females, they are putatively located
in the Y-chromosome; and vice versa for the W-linked markers.
}
\strong{Homogametic sex-markers:}
We have two different methods to identify markers on X or Z chromosomes:
\itemize{
\item \strong{Heterozygosity method:} By looking at the heterozygosity of a marker between sexes, we can
identify markers that are always homozygous in one sex (e.g. males for an
XY system), while exhibiting an intermediate range of heterozygosity in the
other sex (0.1 - 0.5).
\item \strong{Coverage method:} If the data includes count information, this function will look for
markers that have double the number of counts for either of the sexes.
For example if an XY/XX system is present, females are expected to have
double the number of counts for markers on the X chromosome.
}
}

\examples{
\dontrun{
# The minumum
sex.markers <- radiator::sexy_markers(
    data = "shark.dart.data.csv",
    strata = "shark.strata.tsv")
# This will use the default: interactive version and a list is created and to view the sex markers
}
}
\seealso{
Eric Anderson's \href{https://github.com/eriqande/whoa}{whoa} package.
}
\author{
Floriaan Devloo-Delva \email{Floriaan.Devloo-Delva@csiro.au} and with help from
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
