% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_het_outliers.R
\name{simulate_genos_from_posterior}
\alias{simulate_genos_from_posterior}
\title{simulate_genos_from_posterior}
\usage{
simulate_genos_from_posterior(D, p, m)
}
\arguments{
\item{D}{an 012,-1 matrix of observed genotypes}

\item{p}{the estimated allele freqs}

\item{m}{the genotyping error rate (must be a scalar)}
}
\description{
simulate values for the genotypes given the observed genotype,
then estimate allele frequencies, and the genotyping error rate.

This is a helper function for the estimate_m function
}
\author{
Eric Anderson \email{eric.anderson@noaa.gov}
}
\keyword{internal}
