% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_ld.R
\name{snp_ld}
\alias{snp_ld}
\title{GBS/RADseq short distance linkage disequilibrium pruning}
\usage{
snp_ld(data, snp.ld = "first")
}
\arguments{
\item{data}{A data set with ID (LOCUS) and POS (SNP) information.
Usually a VCF file in the working directory or
object in the global environment in wide or long (tidy) formats.}

\item{snp.ld}{(character) \strong{For VCF file only}. 3 options:
\code{snp.ld = "random"} selection, \code{snp.ld = "first"} or
\code{snp.ld = "last"} for SNP on the same short read/haplotype.
Default: \code{snp.ld = "first"}.}
}
\description{
\strong{For VCF file only}.
SNP short distance linkage disequilibrium pruning. With anonymous markers from
RADseq/GBS de novo discovery, you can minimize linkage disequilibrium (LD) by
choosing among these 3 options (see argument below).

For long distance linkage disequilibrium pruning, see details below.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
