% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_parallel.R
\name{split_tibble_rows}
\alias{split_tibble_rows}
\title{split_tibble_rows}
\usage{
split_tibble_rows(
  x,
  lines.cpu = 1000,
  parallel.core = parallel::detectCores() - 1,
  group.split = TRUE
)
}
\description{
Split rows of tibble for parallel processing
}
\keyword{internal}
