% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{split_vcf}
\alias{split_vcf}
\title{Split a VCF file}
\usage{
split_vcf(data, strata, parallel.core = parallel::detectCores() - 1, ...)
}
\arguments{
\item{data}{14 options for input (\strong{diploid data only}): VCFs (SNPs or Haplotypes,
to make the vcf population ready),
plink (tped, bed), stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop, DArT,
and a data frame in long/tidy or wide format. To verify that radiator detect
your file format use \code{\link{detect_genomic_format}} (see example below).
Documented in \strong{Input genomic datasets} of \code{\link{tidy_genomic_data}}.

\strong{DArT and VCF data}: \pkg{radiator} was not meant to generate alleles
and genotypes if you are using a VCF file with no genotype
(only genotype likelihood: GL or PL).
Neither is \pkg{radiator} able to magically generate a genind object
from a SilicoDArT dataset. Please look at the first few lines of your dataset
to understand it's limit before asking raditor to convert or filter your dataset.}

\item{strata}{A file identical to the strata file usually used in radiator,
with an additional column named: \code{SPLIT}.
This new column contains numerical values
(e.g. 1, 1, 1, ..., 2, 2, 2, 2, ..., 3, 3, ...),
that indicate for each INDIVIDUALS/STRATA, how to split.
The number of VCF to split to is based on the max value found in the column
\code{SPLIT}, above this would result in 3 VCF files created).}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{...}{(optional) To pass further arguments for fine-tuning the function.}
}
\value{
The function returns in the global environment a list with
the different tidy dataset from the split vcf. In the working directory,
the splitted VCF files with \code{"_1", "_2"} in the name.
}
\description{
This function allows to split a VCF file in several VCFs,
based on individuals or populations.
}
\examples{
\dontrun{
split.data <- radiator::split_vcf(
data = "batch_1.vcf",
strata = "strata.split.tsv",
blacklist.id = "blacklisted.id.txt",
whitelist.markers = "whitelist.loci.txt")
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
