% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_vec}
\alias{split_vec}
\title{split_vec}
\usage{
split_vec(x, chunks = NULL, tibble.split = FALSE, strata.split = FALSE)
}
\description{
Split input into chunk for parallel processing
}
\keyword{internal}
