% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_genomic_data.R
\name{split_vec_row}
\alias{split_vec_row}
\title{split_vec_row}
\usage{
split_vec_row(x, cpu.rounds, parallel.core = parallel::detectCores() - 1)
}
\description{
Split input into chunk for parallel processing
}
\keyword{internal}
