% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_rad}
\alias{strip_rad}
\title{strip_rad}
\usage{
strip_rad(
  x,
  m = c("VARIANT_ID", "MARKERS", "CHROM", "LOCUS", "POS", "COL", "REF", "ALT"),
  env.arg = NULL,
  keep.strata = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{The data}

\item{m}{(character, string) The variables part of the markers metadata.}

\item{env.arg}{You want to redirect \code{rlang::current_env()} to this argument.}

\item{keep.strata}{(logical) Keep the strata in the dataset or remove the info
and keep only the sample ids.}

\item{verbose}{(logical) The function will chat more when allowed.}
}
\description{
Strip a tidy data set of it's strata and markers meta. Used internally.
}
