% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_het_outliers.R
\name{summarise_genotypes}
\alias{summarise_genotypes}
\title{summarise genotypes}
\usage{
summarise_genotypes(data, path.folder = NULL)
}
\description{
Function that summarise genotypes info
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
