% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_markers.R
\name{summarize_sex}
\alias{summarize_sex}
\title{summarize_sex}
\usage{
summarize_sex(data, silicodata, data.source, coverage.thresholds = NULL,
  tau = 0.3)
}
\description{
Function to generate the different figures required.
}
\keyword{internal}
