% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_genotype_likelihood}
\alias{summary_genotype_likelihood}
\title{Genotype likelihood summary}
\usage{
summary_genotype_likelihood(data, pop.levels = NULL,
  gl.approach = "haplotype", folder = NULL)
}
\arguments{
\item{data}{A tidy vcf object or file (using ".tsv"),
created with tidy_genomic_data.}

\item{pop.levels}{(optional) Character string defining your ordered populations.
Default: \code{pop.levels = NULL}}

\item{gl.approach}{Character. By \code{"SNP"} or by \code{"haplotype"}.
The function will consider the SNP or haplotype GL statistics to filter the marker.
Default: \code{gl.approach = "haplotype"}.}

\item{folder}{(optional) Path to folder where results will be saved. With the
default, \code{folder = NULL}, the working directory is used.}
}
\value{
A list with 3 object (data frames): the statistics at the
individual level ($gl.individuals,
saved as: "genotype.likelihood.individual.summary.tsv"), a long format with information by
marker and population ($gl.summary.marker.pop) and the overall summary by
population ($gl.summary.pop). Those 3 data frames are also written in your
working directory.
}
\description{
This function create 3 summary tables with
genotype likelihood statistics. Individuals, populations and markers information is provided.
The input data is a tidy vcf created.
with \link{tidy_genomic_data}.
}
\details{
The table contains summary statistics: mean, median, min, max and
diff (max-min), of genotype likelihood for individuals, markers and populations. To access
the two tables, use $. The table that summarise by populations was created
using average nested: loci -> individuals -> populations.
Using the haplotype approach: The summary statistics are averaged
and nested SNP -> individuals -> population -> loci. e.g. the mean GL is the average
genotype likelihood for all individuals of pop x for loci x.
}
\keyword{internal}
