% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{switch_allele_count}
\alias{switch_allele_count}
\title{switch_allele_count}
\usage{
switch_allele_count(x, dart.check = FALSE, ref = TRUE)
}
\description{
Switch from dart genotyping 2 presence absence
}
\keyword{internal}
