% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy2wide}
\alias{tidy2wide}
\title{rad_write}
\usage{
tidy2wide(x = NULL, gds = NULL, individuals = NULL, markers = NULL,
  tidy = TRUE, wide = TRUE, wide.markers = TRUE)
}
\description{
Generate the rad folders

tidy2wide
}
\keyword{internal}
