% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dart.R
\name{tidy_dart}
\alias{tidy_dart}
\title{Swiss Army knife tool to prepare \href{http://www.diversityarrays.com}{DArT}
output file for population genetics analysis.}
\usage{
tidy_dart(data, strata, output, pop.levels = NULL, blacklist.id = NULL,
  pop.select = NULL, monomorphic.out = TRUE, common.markers = TRUE,
  mixed.genomes.analysis = TRUE, ind.heterozygosity.threshold = NULL,
  duplicate.genomes.analysis = TRUE, missing.analysis = TRUE,
  filter.reproducibility = NULL, plot.reproducibility = FALSE,
  filter.coverage.high = NULL, filter.coverage.low = NULL,
  plot.coverage = FALSE, filter.call.rate = NULL, plot.call.rate = FALSE,
  filter.ind.missing.geno = NULL, plot.ind.missing.geno = FALSE,
  filter.markers.missing.ind = NULL, plot.markers.missing.ind = FALSE,
  snp.ld = NULL, plot.number.snp.reads = FALSE, maf.thresholds = NULL,
  maf.pop.num.threshold = 1, maf.approach = "SNP", maf.operator = "OR",
  filename = NULL, imputation.method = NULL,
  hierarchical.levels = "populations", verbose = FALSE,
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{data}{DArT output file in \emph{wide} or \emph{binary} format
used in CSIRO genomic projects.}

\item{strata}{A tab delimited file with columns header:
\code{INDIVIDUALS} and \code{POP_ID}.
Note: the column \code{POP_ID} refers to any grouping of individuals. If a third column
named \code{NEW_ID} is used, this column will be used to replace the
\code{INDIVIDUALS} column in the main data file.}

\item{output}{16 genomic data formats can be exported: tidy (by default),
genind, genlight, vcf (for file format version, see details below), plink, genepop,
structure, arlequin, hierfstat, gtypes (strataG), betadiv, snprelate (for SNPRelate's GDS).
Use a character string,
e.g. \code{output = c("genind", "genepop", "structure")}, to have preferred
output formats generated. The tidy format is generated automatically.
Default: \code{output = NULL}.}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
White spaces in population names are replaced by underscore.
Default: \code{pop.levels = NULL}.}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is an object in your
global environment or a file in the working directory
(e.g. "blacklist.txt").
Default: \code{blacklist.id = NULL}.}

\item{pop.select}{(string, optional) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops).
Default: \code{pop.select = NULL}}

\item{monomorphic.out}{(optional) Should the monomorphic
markers present in the dataset be filtered out ?
Default: \code{monomorphic.out = TRUE}.}

\item{common.markers}{(optional) Logical. Default: \code{common.markers = TRUE},
will only keep markers in common (genotyped) between all the populations.}

\item{mixed.genomes.analysis}{(optional) Detect mixed genomes before and
after filters.
Default: \code{detect_mixed_genomes = TRUE}.}

\item{ind.heterozygosity.threshold}{(string, double, optional)
Blacklist individuals based on observed heterozygosity (averaged across markers).


The string contains 2 thresholds values (min and max). 
The values are proportions (0 to 1), where 0 turns off the min threshold and
1 turns off the max threshold.
Individuals with mean observed heterozygosity higher (>) or lower (<) 
than the thresholds will be blacklisted. 

Default: \code{ind.heterozygosity.threshold = NULL} will turn off completely 
the filter and the function will only output the plots and table of heterozygosity.}

\item{duplicate.genomes.analysis}{(optional) Detect duplicate individuals
before and after filters.
Default: \code{detect_duplicate_genomes = TRUE}.}

\item{missing.analysis}{(optional) Visualize pattern of missing data before
and after filters.
Default: \code{missing.analysis = TRUE}}

\item{filter.reproducibility}{(optional, numerical) Filter the \code{RepAvg}
column in the data set. Default: \code{filter.reproducibility = NULL}.
e.g to keep markers with reproducibility >= 99%,
use: \code{filter.reproducibility = 0.99}.}

\item{plot.reproducibility}{(optional, logical) Plot the distribution
(violin plot) of reproducibility. Default: \code{plot.reproducibility = FALSE}.}

\item{filter.coverage.high}{(optional, numerical) Filter the upper bound of the
\code{AvgCountSnp} column in the data set. Default: \code{filter.coverage.high = NULL}.
e.g to keep markers with coverage <= 150 (depth of coverage),
use : \code{filter.coverage.high = 150}.}

\item{filter.coverage.low}{(optional, numerical) Filter the lower bound of the
\code{AvgCountSnp} column in the data set. Default: \code{filter.coverage.low = NULL}.
e.g to keep markers with coverage >= 10 (depth of coverage),
use : \code{filter.coverage.low = 10}.}

\item{plot.coverage}{(optional, logical) Plot the coverage distribution
(violin plot). Default: \code{plot.coverage = FALSE}.}

\item{filter.call.rate}{(optional, numerical) Filter the \code{CallRate}
column in the data set. Default: \code{filter.call.rate = NULL}. e.g to keep
markers genotyped in more than 95% of the individuals use :
\code{filter.call.rate = 0.95}}

\item{plot.call.rate}{(optional, logical) Plot the distribution
(violin plot) of call rate. Default: \code{plot.call.rate = FALSE}.}

\item{filter.ind.missing.geno}{(optional, numerical) Filter with the individual's
proportion of missing genotype. Similar to call rate.
e.g to keep individuals genotyped at >= 0.90 of the markers, use:
\code{filter.ind.missing.geno = 0.90}.
Default: \code{filter.ind.missing.geno = NULL}.}

\item{plot.ind.missing.geno}{(optional, logical) Plot the distribution
(violin plot) of an individual's genotype proportion.
Default: \code{plot.ind.missing.geno = FALSE}.}

\item{filter.markers.missing.ind}{(optional, numerical) Filter markers based
on the proportion of individuals genotyped.
e.g to keep markers with >= 0.95 of the genotyped individuals, use:
\code{filter.markers.missing.ind = 0.95}.
Default: \code{filter.markers.missing.ind = NULL}.}

\item{plot.markers.missing.ind}{(optional, logical) Plot the distribution
(violin plot) of markers genotype individuals proportion.
Default: \code{plot.markers.missing.ind = FALSE}.}

\item{snp.ld}{(optional) \strong{For VCF file only}.
SNP short distance linkage disequilibrium pruning. With anonymous markers from
RADseq/GBS de novo discovery, you can minimize linkage disequilibrium (LD) by
choosing among these 3 options: \code{"random"} selection, \code{"first"} or
\code{"last"} SNP on the same short read/haplotype. For long distance linkage
disequilibrium pruning, see details below.
Default: \code{snp.ld = NULL}.}

\item{plot.number.snp.reads}{(optional, logical) Plot the distribution of SNP
per read.
Default: \code{plot.number.snp.reads = FALSE}.}

\item{maf.thresholds}{(string, double, optional) String with
local/populations and global/overall maf thresholds, respectively.
e.g. \code{maf.thresholds = c(0.05, 0.1)} for a local maf threshold
of 0.05 and a global threshold of 0.1. Available for VCF, PLINK and data frame
files.
Default: \code{maf.thresholds = NULL}.}

\item{maf.pop.num.threshold}{(integer, optional) When maf thresholds are used,
this argument is for the number of pop required to pass the maf thresholds
to keep the locus. Default: \code{maf.pop.num.threshold = 1}}

\item{maf.approach}{(character, optional).
\code{maf.approach = "haplotype"} : looks at the minimum MAF found on the
read/haplotype. Using this option will discard all the markers/snp on
that read based on the thresholds chosen. This method is only available
for VCF and haplotype files, or tidy data frame from those file types.
\code{maf.approach = "SNP"} : treats all the SNP on the same
haplotype/read as independent. Doesn't work with haplotype file,
but does work for all other file type.
Default is \code{maf.approach = "SNP"}.}

\item{maf.operator}{(character, optional) \code{maf.operator = "AND"} or
default \code{maf.operator = "OR"}.
When filtering over LOCUS or SNP, do you want the local \code{"AND"}
global MAF to pass the thresholds, or ... you want the local \code{"OR"}
global MAF to pass the thresholds, to keep the marker?}

\item{filename}{(optional) The filename prefix for the objet in the global environment
or the working directory. Default: \code{filename = NULL}. A default name will be used,
customized with the output file(s) selected.}

\item{imputation.method}{(character, optional)
Methods available for map-independent imputations of missing genotype:

(1) \code{imputation.method = "max"} Strawman imputation,
the most frequently observed genotypes (ties are broken at random).

(2) \code{imputation.method = "rf"} On-the-fly-imputations using
Random Forests algorithm.

(3) \code{imputation.method = "boost"} extreme gradient boosting trees.

(4) \code{imputation.method = "mca"} Multiple Correspondence Analysis.

\code{imputation.method = NULL} will return the original dataset, without
imputation.
Default: \code{imputation.method = "rf"}.}

\item{hierarchical.levels}{(character, optional) \code{c("global", "populations")}.
Should the imputations be computed globally or by populations. Note that
imputing genotype globally in conjunction with \code{imputation.method = "max"}
can create huge bias for example by
introducing foreign genotypes in some populations (see note for more info).
Default: \code{hierarchical.levels = "populations"}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during vcf import.
Default: \code{parallel::detectCores() - 1}.}
}
\value{
The function returns an object (list). The content of the object
can be listed with \code{names(object)} and use \code{$} to isolate specific
object (see examples). Some output format will write the output file in the
working directory. The tidy genomic data frame is generated automatically.
}
\description{
Import, filter and generate imputed dataset of DArT output file.
}
\examples{
\dontrun{
testing
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
