% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dart_metadata.R
\name{tidy_dart_metadata}
\alias{tidy_dart_metadata}
\title{Import and tidy \href{http://www.diversityarrays.com}{DArT} metadata.}
\usage{
tidy_dart_metadata(data, filename = NULL, verbose = FALSE,
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{data}{DArT output file. Note that most popular formats used by DArT are
recognised (1- and 2- row format, also called binary, and count data.).
If you encounter a problem, sent me your data so that I can update
the function. The function can import \code{.csv} or \code{.tsv} files.}

\item{filename}{(optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in
the working directory. The file extension appended to
the \code{filename} provided is \code{.rad}.
With default: \code{filename = NULL}, the tidy data frame is
in the global environment only (i.e. not written in the working directory...).}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel::detectCores() - 1}.}
}
\value{
A tidy dataframe with these columns:
\enumerate{
\item MARKERS: generated by radiator and correspond to CHROM + LOCUS + POS
separated by 2 underscores.
\item CHROM: the chromosome, for de novo: CHROM_1.
\item LOCUS: the locus.
\item POS: the SNP id on the LOCUS.
\item REF: the reference allele.
\item ALT: the alternate allele.
\item CALL_RATE: call rate output specific of DArT.
\item AVG_COUNT_REF: the coverage for the reference allele, output specific of DArT.
\item AVG_COUNT_SNP: the coverage for the alternate allele, output specific of DArT.
\item REP_AVG: the reproducibility average, output specific of DArT.
}
}
\description{
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
\code{\link{filter_dart}} and might be of interest for users.
The function generate a tidy dataset of
\href{http://www.diversityarrays.com}{DArT} markers and associated metadata.
Usefull to filter before importing the actual dataset.
}
\examples{
\dontrun{
clownfish.dart.tidy <- radiator::tidy_dart_metadata(
data = "clownfish.dart.tsv",
verbose = TRUE)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
