% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_genlight.R
\name{tidy_genlight}
\alias{tidy_genlight}
\title{Tidy a genlight object to a tidy dataframe}
\usage{
tidy_genlight(data)
}
\arguments{
\item{data}{A genlight object in the global environment.}
}
\description{
Tidy genlight object to a tidy dataframe.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator} 
and might be of interest for users.
}
\references{
Jombart T (2008) adegenet: a R package for the multivariate
analysis of genetic markers. Bioinformatics, 24, 1403-1405.

Jombart T, Ahmed I (2011) adegenet 1.3-1: 
new tools for the analysis of genome-wide SNP data. 
Bioinformatics, 27, 3070-3071.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
