% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_genomic_data.R
\name{tidy_genomic_data}
\alias{tidy_genomic_data}
\title{Transform common genomic dataset format in a tidy data frame}
\usage{
tidy_genomic_data(data, vcf.metadata = FALSE, whitelist.markers = NULL,
  monomorphic.out = TRUE, blacklist.genotype = NULL, snp.ld = NULL,
  common.markers = TRUE, maf.thresholds = NULL, maf.pop.num.threshold = 1,
  maf.approach = "SNP", maf.operator = "OR", max.marker = NULL,
  blacklist.id = NULL, pop.levels = NULL, pop.labels = NULL,
  strata = NULL, pop.select = NULL, filename = NULL,
  parallel.core = parallel::detectCores() - 1, verbose = TRUE)
}
\arguments{
\item{data}{12 options: VCF (SNPs or Haplotypes,
to make the vcf population ready, see details below),
plink, stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop, DArT,
and a data frame in long/tidy or wide format.
\emph{See details} of \code{\link[radiator]{tidy_genomic_data}}.}

\item{vcf.metadata}{(optional, logical or string) For VCF files only.
With \code{logical, TRUE/FALSE}, \code{vcf.metadata = FALSE}, only the genotype
information is kept (GT field). With \code{vcf.metadata = TRUE},
all the metadata contained in the \code{FORMAT} field will be kept in
the tidy data file. radiator is currently keeping/cleaning only these metadata:
\code{"DP", "AD", "GL", "PL", "GQ", "GOF", "NR", "NV"}.
If you need different one, submit a request.
With \code{string}, explicitely ask for a particular FORMAT field you want
to keep (except the GT field that is always imported).
e.g. \code{vcf.metadata = "PL"} or \code{vcf.metadata = c("DP", "GL")}.
Default: \code{vcf.metadata = FALSE}.}

\item{whitelist.markers}{(optional) A whitelist containing CHROM (character
or integer) and/or LOCUS (integer) and/or
POS (integer) columns header. To filter by chromosome and/or locus and/or by snp.
The whitelist is an object in your
global environment or a file in the working directory (e.g. "whitelist.txt").
de novo CHROM column with 'un' need to be changed to 1.
In the VCF, the column ID is the LOCUS identification (VCF generated from
stacks have the SNP position on the read embedded in the ID,
so the ID = no longer represent the LOCUS). Marker names are cleaned of
separators that interfere with some packages or codes:
\code{/}, \code{:}, \code{-} and \code{.} are changed to an underscore
\code{_}.
Default \code{whitelist.markers = NULL} for no whitelist of markers.}

\item{monomorphic.out}{(optional) Should the monomorphic
markers present in the dataset be filtered out ?
Default: \code{monomorphic.out = TRUE}.}

\item{blacklist.genotype}{(optional) Useful to erase genotype with below
average quality, e.g. genotype with more than 2 alleles in diploid likely
sequencing errors or genotypes with poor genotype likelihood or coverage.
The blacklist has a minimum of 2 column headers (markers and individuals).
Markers can be 1 column (CHROM or LOCUS or POS),
a combination of 2 (e.g. CHROM and POS or CHROM and LOCUS or LOCUS and POS) or
all 3 (CHROM, LOCUS, POS). The markers columns must be designated: CHROM (character
or integer) and/or LOCUS (integer) and/or POS (integer). The id column designated
INDIVIDUALS (character) columns header.
The blacklist is an object in your global environment or
a file in the working directory (e.g. "blacklist.genotype.txt").
For de novo VCF, CHROM column
with 'un' need to be changed to 1.
Marker names are cleaned of
separators that interfere with some packages or codes:
\code{/}, \code{:}, \code{-} and \code{.} are changed to an underscore
\code{_}.
Ids are also cleaned of separators that interfere with some packages or codes:
\code{_} and \code{:} are changed to a dash \code{-}.
Default: \code{blacklist.genotype = NULL} for no blacklist of
genotypes to erase.}

\item{snp.ld}{(optional) \strong{For data with locus and SNP info, like VCF and DArT file}.
SNP short distance linkage disequilibrium pruning. With anonymous markers from
RADseq/GBS de novo discovery, you can minimize linkage disequilibrium (LD) by
choosing among these 4 options:
\code{snp.ld = "random"} for a random selection of 1 SNP on the read,
\code{snp.ld = "first"} for the first one on the read...,
\code{snp.ld = "last"} for the last SNP on the read and
\code{snp.ld = "middle"} for locus with > 2 SNPs/read the option to select at random
one SNP between the first and the last SNP on the read. If the locus as <= 2
SNPs on the read, the first one is selected. Note that for that last option,
the numbers are reported. For long distance linkage
disequilibrium pruning, see details below.
Default: \code{snp.ld = NULL}.}

\item{common.markers}{(optional) Logical. Default: \code{common.markers = TRUE},
will only keep markers in common (genotyped) between all the populations.}

\item{maf.thresholds}{(string, double, optional) String with
local/populations and global/overall maf thresholds, respectively.
e.g. \code{maf.thresholds = c(0.05, 0.1)} for a local maf threshold
of 0.05 and a global threshold of 0.1. Available for VCF, PLINK and data frame
files.
Default: \code{maf.thresholds = NULL}.}

\item{maf.pop.num.threshold}{(integer, optional) When maf thresholds are used,
this argument is for the number of pop required to pass the maf thresholds
to keep the locus. Default: \code{maf.pop.num.threshold = 1}}

\item{maf.approach}{(character, optional).
\code{maf.approach = "haplotype"} : looks at the minimum MAF found on the
read/haplotype. Using this option will discard all the markers/snp on
that read based on the thresholds chosen. This method is only available
for VCF and haplotype files, or tidy data frame from those file types.
\code{maf.approach = "SNP"} : treats all the SNP on the same
haplotype/read as independent. Doesn't work with haplotype file,
but does work for all other file type.
Default is \code{maf.approach = "SNP"}.}

\item{maf.operator}{(character, optional) \code{maf.operator = "AND"} or
default \code{maf.operator = "OR"}.
When filtering over LOCUS or SNP, do you want the local \code{"AND"}
global MAF to pass the thresholds, or ... you want the local \code{"OR"}
global MAF to pass the thresholds, to keep the marker?}

\item{max.marker}{(integer, optional) For large PLINK files,
useful to subsample marker number. e.g. if the data set
contains 200 000 markers and \code{max.marker = 10000}, 10000 markers are
subsampled randomly from the 200000 markers. If you need specific markers,
use \code{whitelist.markers} argument.
Default: \code{max.marker = NULL}.}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is an object in your
global environment or a file in the working directory
(e.g. "blacklist.txt").
Default: \code{blacklist.id = NULL}.}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
White spaces in population names are replaced by underscore.
Default: \code{pop.levels = NULL}.}

\item{pop.labels}{(optional, string) Use this argument to rename/relabel
your pop or combine your pop. e.g. To combine \code{"QUE"} and \code{"ONT"}
into a new pop called \code{"NEW"}:
(1) First, define the levels for your pop with \code{pop.levels} argument:
\code{pop.levels = c("QUE", "ONT", "ALB")}.
(2) then, use \code{pop.labels} argument:
\code{pop.levels = c("NEW", "NEW", "ALB")}.
To rename \code{"QUE"} to \code{"TAS"}:
\code{pop.labels = c("TAS", "ONT", "ALB")}.
Default: \code{pop.labels = NULL}. If you find this too complicated,
there is also the \code{strata} argument that can do the same thing,
see below.
White spaces in population names are replaced by underscore.}

\item{strata}{(optional/required) Required for VCF and haplotypes files,
optional for the other formats supported.

The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}.
If a \code{strata} file is specified with all file formats that don't
require it, the strata argument will have precedence on the population
groupings used internally in those file formats.
The \code{STRATA} column can be any hierarchical grouping.
To create a strata file see \code{\link[radiator]{individuals2strata}}.
If you have already run
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data,
the strata file is similar to a stacks \emph{population map file},
make sure you
have the required column names (\code{INDIVIDUALS} and \code{STRATA}).
The strata column is cleaned of a white spaces that interfere with some
packages or codes: space is changed to an underscore \code{_}
Default: \code{strata = NULL}.}

\item{pop.select}{(string, optional) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops).
Default: \code{pop.select = NULL}}

\item{filename}{(optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in
the working directory. The file extension appended to
the \code{filename} provided is \code{.rad}.
With default: \code{filename = NULL}, the tidy data frame is
in the global environment only (i.e. not written in the working directory...).}

\item{parallel.core}{(optional) The number of core used for parallel
execution during vcf import.
Default: \code{parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\value{
The output in your global environment is a tidy data frame.
If \code{filename} is provided, the tidy data frame is also
written in the working directory with file extension \code{.rad}.
The file is written with the
\href{https://github.com/fstpackage/fst}{Lightning Fast Serialization of Data Frames for R} package.
To read the file back in R use \code{\link[fst]{read.fst}}.
}
\description{
Transform genomic data set produced by massive parallel
sequencing pipeline (e.g.GBS/RADseq,
SNP chip, DArT, etc) into a tidy format. The use of blacklist and whitelist along
several filtering options are available to prune the dataset.
Several arguments are available to make your data population-wise and easily
rename the pop id.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
}
\details{
\strong{Long distance SNP linkage disequilibrium pruning}
If you have markers position on a genome or a linkage map,
you can go further in removing linked markers by using
\href{https://github.com/zhengxwen/SNPRelate}{SNPRelate} or
\href{http://pngu.mgh.harvard.edu/~purcell/plink/}{PLINK}, \emph{linkage
disequilibrium based SNP pruning} option.

\strong{Input files:}
\enumerate{
\item VCF biallelic file (e.g. \code{data = "batch_1.vcf"})
To make the VCF population ready, you need the \code{strata} argument.

\item VCF haplotypic file (e.g. \code{data = "batch_1.haplotypes.vcf"})
To make the VCF population ready, you need the \code{strata} argument.

\item haplotype file created in STACKS (e.g. \code{data = "batch_1.haplotypes.tsv"}).
To make the haplotype file population ready, you need the \code{strata} argument.

\item Data frame
To discriminate the long from the wide format,
the function \pkg{radiator} \code{\link[radiator]{tidy_wide}} searches
for \code{MARKERS or LOCUS} in column names (TRUE = long format).
The data frame is tab delimitted.
\strong{Wide format:}
The wide format cannot store metadata info.
The wide format starts with these 2 id columns:
\code{INDIVIDUALS}, \code{POP_ID} (that refers to any grouping of individuals),
the remaining columns are the markers in separate columns storing genotypes.

\strong{Long/Tidy format:}
The long format is considered to be a tidy data frame and can store metadata info.
(e.g. from a VCF see \pkg{radiator} \code{\link[radiator]{tidy_genomic_data}}). A minimum of 4 columns
are required in the long format: \code{INDIVIDUALS}, \code{POP_ID},
\code{MARKERS or LOCUS} and \code{GENOTYPE or GT}. The rest are considered metata info.

\strong{2 genotypes formats are available:}
6 characters no separator: e.g. \code{001002 of 111333} (for heterozygote individual).
6 characters WITH separator: e.g. \code{001/002 of 111/333} (for heterozygote individual).
The separator can be any of these: \code{"/", ":", "_", "-", "."}.

\emph{How to get a tidy data frame ?}
\pkg{radiator} \code{\link[radiator]{tidy_genomic_data}} can transform 11 genomic
data formats in a tidy data frame.

\item PLINK file in
\code{tped/tfam} format (e.g. \code{data =  "data.assignment.tped"}).
The \code{tfam} file is found based on the name prefix of the supplied
\code{tped} file. The first 2 columns of the \code{tfam} file will be used
for the \code{strata} (population identification), unless the \code{strata}
argument is provided.
Columns 1, 3 and 4 of the \code{tped} are discarded. The remaining columns
correspond to the genotype in the format \code{01/04}
where \code{A = 01, C = 02, G = 03 and T = 04}. For \code{A/T} format, use
PLINK or bash to convert.
Use \href{http://vcftools.sourceforge.net/}{VCFTOOLS} with \code{--plink-tped}
to convert very large VCF file. For \code{.ped} file conversion to
\code{.tped} use \href{http://pngu.mgh.harvard.edu/~purcell/plink/}{PLINK}
with \code{--recode transpose}

\item \code{\link[adegenet]{genind}} object from \code{\link[adegenet]{adegenet}}.

\item \code{\link[adegenet]{genlight}} object from \code{\link[adegenet]{adegenet}}.

\item \code{\link[strataG]{gtypes}} object from \code{\link[strataG]{strataG}}.

\item \href{http://www.diversityarrays.com}{DArT} file.

\item genepop data file (e.g. \code{data = "kiwi_data.gen"}). Here, the function can only use
alleles encoded with 3 digits.
}


\strong{GATK VCF files:} Some VCF have an \code{ID} column filled with \code{.},
the LOCUS information is all contained along the linkage group in the
\code{CHROM} column. To make it work with
\href{https://github.com/thierrygosselin/radiator}{radiator},
the \code{ID} column is filled with the \code{POS} column info.

\strong{platypus VCF files:} Some VCF files don't have an ID filed with values,
here the same thing as GATK VCF files above is done.
}
\examples{
\dontrun{
tidy.vcf <- tidy_genomic_data(
data = "batch_1.vcf",
whitelist.markers = "whitelist.vcf.txt",
snp.ld = NULL,
common.markers = TRUE,
blacklist.id = "blacklist.id.treefrog.tsv",
strata = "strata.treefrog.tsv",
pop.levels = c("PAN", "COS")
)
}
}
\references{
Catchen JM, Amores A, Hohenlohe PA et al. (2011)
Stacks: Building and Genotyping Loci De Novo From Short-Read Sequences.
G3, 1, 171-182.

Catchen JM, Hohenlohe PA, Bassham S, Amores A, Cresko WA (2013)
Stacks: an analysis tool set for population genomics.
Molecular Ecology, 22, 3124-3140.

Danecek P, Auton A, Abecasis G et al. (2011)
The variant call format and VCFtools.
Bioinformatics, 27, 2156-2158.

Purcell S, Neale B, Todd-Brown K, Thomas L, Ferreira MAR,
Bender D, et al.
PLINK: a tool set for whole-genome association and population-based linkage
analyses.
American Journal of Human Genetics. 2007: 81: 559–575. doi:10.1086/519795

Jombart T, Devillard S, Balloux F.
Discriminant analysis of principal components: a new method for the analysis
of genetically structured populations.
BMC Genet. 2010:11: 94. doi:10.1186/1471-2156-11-94

Jombart T, Ahmed I. adegenet 1.3-1: new tools for the analysis
of genome-wide SNP data.
Bioinformatics. 2011:27: 3070–3071. doi:10.1093/bioinformatics/btr521

Raymond M. & Rousset F, (1995).
GENEPOP (version 1.2): population genetics software for exact tests
and ecumenicism.
J. Heredity, 86:248-249

Archer FI, Adams PE, Schneiders BB.
strataG: An r package for manipulating, summarizing and analysing population
genetic data.
Molecular Ecology Resources. 2016. doi:10.1111/1755-0998.12559
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
