% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_gtypes.R
\name{tidy_gtypes}
\alias{tidy_gtypes}
\title{Tidy a gtypes object to a tidy dataframe}
\usage{
tidy_gtypes(data)
}
\arguments{
\item{data}{A gtypes object (>= v.2.0.2) in the global environment.}
}
\description{
Transform a \code{\link[strataG]{gtypes}} object from
\code{\link[strataG]{strataG}} to a tidy dataframe.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\references{
Archer FI, Adams PE, Schneiders BB.
strataG: An r package for manipulating, summarizing and analysing population
genetic data.
Molecular Ecology Resources. 2017; 17: 5-11. doi:10.1111/1755-0998.12559
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
