% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_silico_dart.R
\name{tidy_silico_dart}
\alias{tidy_silico_dart}
\title{Tidy silico \href{http://www.diversityarrays.com}{DArT} output file.}
\usage{
tidy_silico_dart(data, strata = NULL, filename = NULL,
  verbose = TRUE)
}
\arguments{
\item{data}{Silico DArT output file.
The function can import \code{.csv} or \code{.tsv} files.}

\item{strata}{A tab delimited file or object with 3 columns.
Columns header is:
\code{TARGET_ID}, \code{INDIVIDUALS} and \code{STRATA}.
Note: the column \code{STRATA} refers to any grouping of individuals.
You need to make sure that
the column \code{TARGET_ID} match the id used by DArT.
The column \code{INDIVIDUALS} and \code{STRATA} will be kept in the tidy data.
Only individuals in the strata file are kept in the tidy, i.e. that the strata
is also used as a whitelist of individuals/strata.
Silico DArT data is currently used to detect sex markers, so the \code{STRATA}
column should be filed with sex information: \code{M} or \code{F}.}

\item{filename}{(optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in
the working directory. The file extension appended to
the \code{filename} provided is \code{.rad}.
With default: \code{filename = NULL}, the tidy data frame is
in the global environment only (i.e. not written in the working directory...).}

\item{verbose}{(optional, logical) When verbose = TRUE the function is a
little more chatty during execution.
Default: \code{verbose = FALSE}.}
}
\value{
A tidy dataframe with several columns depending on DArT file:
#TODO


Written in the working directory:
#TODO
}
\description{
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users. The function generate a tidy dataset using
\href{http://www.diversityarrays.com}{DArT} file.
}
\examples{
\dontrun{
clownfish.dart.tidy <- radiator::tidy_silico_dart(
data = "clownfish.dart.tsv",
strata = "clownfish.strata.tsv",
verbose = TRUE)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
