% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_vcf.R
\name{tidy_vcf}
\alias{tidy_vcf}
\title{Tidy a vcf file (bi/multi-allelic).}
\usage{
tidy_vcf(data, strata = NULL, parallel.core = parallel::detectCores() -
  1, verbose = FALSE, ...)
}
\arguments{
\item{data}{(character string) The VCF SNPs are biallelic or haplotypes.


To make the VCF population-ready, you have to use \code{strata} argument.
\strong{GATK VCF files:} Some VCF have an \code{ID} column filled with \code{.},
the LOCUS information is all contained along the linkage group in the
\code{CHROM} column. To make it work with
\href{https://github.com/thierrygosselin/radiator}{radiator},
the \code{ID} column is filled with the \code{POS} column info.
\strong{platypus VCF files:} Some VCF files don't have an ID filed with values,
here the same thing as GATK VCF files above is done.
\strong{stacks VCF files:} with \emph{de novo} approaches, the CHROM column is
filled with "1", the LOCUS column correspond to the CHROM section in stacks VCF and
the COL column is POS -1. With a reference genome, the ID column in stacks VCF is
separated into "LOCUS", "COL", "STRANDS".}

\item{strata}{(optional)
The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}.
If a \code{strata} file is specified with all file formats that don't
require it, the strata argument will have precedence on the population
groupings used internally in those file formats. For file formats without
population/strata groupings (e.g. vcf, haplotype files) if no strata file is
provided, 1 pop/strata grouping will automatically be created.
For vcf and haplotypes file, the strata can also be used as a whitelist of id.
Samples not in the strata file will be discarded from the data set.
The \code{STRATA} column can be any hierarchical grouping.
To create a strata file see \code{\link[radiator]{individuals2strata}}.
If you have already run
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data,
the strata file is similar to a stacks \emph{population map file},
make sure you
have the required column names (\code{INDIVIDUALS} and \code{STRATA}).
The strata column is cleaned of a white spaces that interfere with some
packages or codes: space is changed to an underscore \code{_}.
Default: \code{strata = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) To pass further argument for fine-tuning the tidying
(details below).}
}
\value{
The output in your global environment is a tidy data frame.
}
\description{
Used internally in
\href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
Highly recommended to use \code{\link[radiator]{tidy_genomic_data}} or
\code{\link[radiator]{genomic_converter}}. Those two functions allows
to manipulate and prune the dataset with blacklists and whitelists along
several other filtering options.
}
\details{
\strong{Advance mode, using \emph{dots-dots-dots ...}}
\enumerate{
\item \code{whitelist.markers: }documented in \code{\link[radiator]{tidy_genomic_data}}
\item \code{blacklist.id: }documented in \code{\link[radiator]{tidy_genomic_data}}
\item \code{pop.select: }documented in \code{\link[radiator]{tidy_genomic_data}}
\item \code{pop.levels: }documented in \code{\link[radiator]{tidy_genomic_data}}
\item \code{pop.labels: }documented in \code{\link[radiator]{tidy_genomic_data}}
\item \code{ref.calibration: } (optional, logical)
Default: \code{ref.calibration = FALSE}.
REF/ALT alleles are designated based on count.
The allele with the higher overall count is called REF.
When the count for REF/ALT alleles is equal, the REF allele is chosen randomly.
REF/ALT designation can change when individuals are blacklisted.
If individuals are blacklisted by
internal filters inside this function the argument is turned on automatically.
If you removed individuals in another software that doesn't check this, force
the calibration of REF/ALT alleles: \code{ref.calibration = TRUE}.
\item \code{vcf.stats: } (optional, logical) Generates individuals and markers important statistics
helpful for filtering. These are very fast to generate and because computational
cost is minimal, even for huge VCFs, the default is \code{vcf.stats = TRUE}.
\item \code{vcf.metadata: } (optional, logical or character string)
With \code{vcf.metadata = FALSE}, only the genotypes are kept (GT field).
With \code{vcf.metadata = TRUE},
all the metadata contained in the \code{FORMAT} field will be kept in
the tidy data file. radiator is currently keeping/cleaning only these metadata:
\code{"DP", "AD", "GL", "PL", "GQ", "HQ", "GOF", "NR", "NV"}.
e.g. you only wnat AD and PL, \code{vcf.metadata = c("AD", "PL")}.
If yours is not in the list, submit a request.
Default: \code{vcf.metadata = FALSE}.
}
}
\examples{
\dontrun{
# very basic with built-in defaults (not recommended):
 prep.data <- radiator::tidy_vcf(data = "populations.snps.vcf")

# Using more arguments and filters (recommended):
prep.data <- radiator::tidy_vcf(
data = "populations.snps.vcf",
strata = "strata_salamander.tsv",
vcf.stats = TRUE,
filter.individuals.missing = "outlier",
common.markers = TRUE,
keep.both.strands = FALSE,
filter.mac = 4,
filter.markers.missing = 50,
filter.snp.read.position = "outliers",
filter.short.ld = "maf",
filter.long.ld = NULL,
vcf.metadata = TRUE,
path.folder = "salamander/prep_data",
verbose = TRUE)

# To view the objects generated by the function:
names(prep.data)

# if you want to check the blacklisted markers:
bl <- prep.data$blacklist.markers
# or get the numbers associated with the filters applied:
 bl.numbers <- dplyr::count(prep.data$blacklist.markers, FILTER)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
