% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{update_bl_individuals}
\alias{update_bl_individuals}
\title{update_bl_individuals}
\usage{
update_bl_individuals(gds, generate = FALSE, update = NULL,
  extract = FALSE, bl.i.gds = NULL)
}
\description{
Generate and update blacklist of individuals on radiator GDS node
}
\keyword{internal}
