% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_parameters}
\alias{update_parameters}
\title{update_parameters}
\usage{
update_parameters(generate = FALSE, initiate = FALSE, update = TRUE,
  parameter.obj = NULL, data = NULL, filter.name = "",
  param.name = "", values = paste(NULL, NULL, sep = " / "),
  units = "individuals / strata / chrom / locus / markers",
  comments = "", path.folder = NULL, file.date = NULL,
  internal = FALSE, verbose = TRUE)
}
\description{
Generate or update a filters parameters file and object.
Used internally in radiator, not usefull outside the package.
}
\keyword{internal}
