% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{update_radiator_gds}
\alias{update_radiator_gds}
\title{update_radiator_gds}
\usage{
update_radiator_gds(
  gds,
  radiator.gds = TRUE,
  node.name,
  value,
  replace = TRUE,
  sync = FALSE,
  summary = FALSE,
  verbose = FALSE
)
}
\description{
Update radiator gds node
}
\keyword{internal}
