% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitelist_blacklist.R
\name{whitelist_loci_snp}
\alias{whitelist_loci_snp}
\title{Whitelist loci and snp}
\usage{
whitelist_loci_snp(data, filename, col.header)
}
\arguments{
\item{data}{A tidy vcf or sumstats prep file (using ".tsv") or object in
your Environment.}

\item{filename}{The name of the file written in the directory.}

\item{col.header}{TRUE and the loci will have a column header 'LOCUS'.}
}
\description{
This function creates a whitelist of loci and snp,
useful in the populations module of STACKS.
}
\seealso{
\link{tidy_genomic_data} and  \link{summary_stats_vcf_tidy}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
