% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_betadiv.R
\name{write_betadiv}
\alias{write_betadiv}
\title{Write a betadiv file from a tidy data frame}
\usage{
write_betadiv(data)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}
}
\value{
A betadiv object is returned.
}
\description{
Write a betadiv file from a tidy data frame.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
}
\references{
Lamy T, Legendre P, Chancerelle Y, Siu G, Claudet J (2015)
Understanding the Spatio-Temporal Response of Coral Reef Fish Communities to
Natural Disturbances: Insights from Beta-Diversity Decomposition.
PLoS ONE, 10, e0138696.
}
\seealso{
\code{beta.div} is available on Pierre Legendre web site \url{http://adn.biol.umontreal.ca/~numericalecology/Rcode/}
}
\author{
Laura Benestan \email{laura.benestan@icloud.com} and
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
