% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dadi.R
\name{write_dadi}
\alias{write_dadi}
\title{Write a \code{dadi} SNP input file from a tidy data frame.}
\usage{
write_dadi(
  data,
  fasta.ingroup = NULL,
  fasta.outgroup = NULL,
  sumstats.ingroup = NULL,
  sumstats.outgroup = NULL,
  dadi.input.filename = NULL
)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{fasta.ingroup}{(optional) The fasta file, sequences for the ingroup.
Leave empty if no outgroup.
Default: \code{fasta.ingroup = NULL}.}

\item{fasta.outgroup}{(optional) The fasta file, sequences for the outgroup.
Default: \code{fasta.outgroup = NULL}.}

\item{sumstats.ingroup}{(optional) The sumstats output file from STACKS
when running STACKS for the ingroup fasta file.This file is
required to use with an outgroup. Leave empty if no outgroup.
Default: \code{sumstats.ingroup = NULL}.}

\item{sumstats.outgroup}{(optional) The sumstats output file from STACKS
when running STACKS for the outgroup fasta file. This file is
required to use an outgroup.
Default: \code{sumstats.outgroup = NULL}.}

\item{dadi.input.filename}{(optional) Name of the \code{dadi} SNP input file
written to the working directory. e.g. \code{dadi.file.tsv}.
Default use date and time to make the file. If used, the file extension
need to finish with \code{.tsv or .txt}.
Default: \code{dadi.input.filename = NULL}.}
}
\value{
The function returns tibble and the dadi input file in the working
directory.
}
\description{
This function will generate a \code{dadi} SNP input file using a
radiator tidy tibble.

Note that missing data can potentially bias demographic inference,
consequently you might want to check the impact of missing data by running
two datset in \code{dadi}: one with the missing data or several with varying thresholds
of missingness, and one with imputed genotypes
(check my other package \href{https://thierrygosselin.github.io/grur/}{grur} for this).
}
\examples{
\dontrun{
# without outgroup:
dadi.data <- radiator::write_dadi(data = "my_tidy_dataset.rad")

# with outgroup and fasta generated by stacks:
dadi.data <- radiator::write_dadi(
   data = "my_tidy_dataset.rad",
   fasta.ingroup = "batch_1.ingroup.fa",
   fasta.outgroup = "batch_1.outgroup.fa",
   sumstats.ingroup = "batch_1.sumstats.ingroup.tsv",
   sumstats.outgroup = "batch_1.sumstats.outgroup.tsv"
)
}
}
\references{
Gutenkunst RN, Hernandez RD, Williamson SH, Bustamante CD (2009)
Inferring the Joint Demographic History of Multiple Populations
from Multidimensional SNP Frequency Data (G McVean, Ed,).
PLoS genetics, 5, e1000695.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
