% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_faststructure.R
\name{write_faststructure}
\alias{write_faststructure}
\title{Write a faststructure file from a tidy data frame}
\usage{
write_faststructure(data, plink.bed = FALSE, pop.levels = NULL,
  filename = NULL, ...)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{plink.bed}{To export in plink BED format. This will write 3 files:
\code{.bed}, \code{.bim}, \code{.fam}.
Default: \code{plink.bed = FALSE}.
Currently under development.}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
White spaces in population names are replaced by underscore.
Default: \code{pop.levels = NULL}.}

\item{filename}{(optional) The file name prefix for the faststructure file
written to the working directory. With default: \code{filename = NULL},
the date and time is appended to \code{radiator_faststructure_}.}

\item{...}{other parameters passed to the function.}
}
\value{
A faststructure file is saved to the working directory.
}
\description{
Write a faststructure file from a tidy data frame.
For biallelic dataset only.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
}
\references{
Raj A, Stephens M, Pritchard JK (2014)
fastSTRUCTURE: Variational Inference of Population Structure in Large SNP
Datasets. Genetics, 197, 573-589.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
