% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_genepopedit.R
\name{write_genepopedit}
\alias{write_genepopedit}
\title{Write a genepopedit flatten object}
\usage{
write_genepopedit(data)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}
}
\value{
A genepopedit object in the global environment.
}
\description{
Write a genepopedit object from a tidy data frame or GDS file/object.

Why not use \code{genepopedit::genepop_flatten}?
\itemize{
\item genepopedit requires a
\href{https://github.com/rystanley/genepopedit#genepopedit}{specific type of genepop},
so if you don't want to manipulate your genepop file, radiator is an alternative.
\item radiator follows guidelines highlighted here:
\href{http://genepop.curtin.edu.au/help_input.html}{genepop format}, but the
3 functions in radiator that reads genepop files:
\enumerate{
\item \href{https://thierrygosselin.github.io/radiator/reference/genomic_converter.html}{genomic_converter}
\item \href{https://thierrygosselin.github.io/radiator/reference/tidy_genomic_data.html}{tidy_genomic_data}
\item the underlying module: \href{https://thierrygosselin.github.io/radiator/reference/tidy_genepop.html}{tidy_genepop}
}
imports a \emph{larger variety of genepop alternatives}, similarly to
\href{https://github.com/thibautjombart/adegenet}{adegenet} \code{read.genepop}
 function, only faster.
}
}
\references{
Stanley RRE, Jeffery NW, Wringe BF, DiBacco C, Bradbury IR (2017)
genepopedit: a simple and flexible tool for manipulating multilocus molecular
data in R. Molecular Ecology Resources, 17, 12-18.
}
\seealso{
\href{https://github.com/rystanley/genepopedit}{genepopedit}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
