% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gtypes.R
\name{write_gtypes}
\alias{write_gtypes}
\title{Write a \code{\link[strataG]{gtypes}} object from a tidy data frame}
\usage{
write_gtypes(data)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}
}
\value{
An object of the class \code{\link[strataG]{gtypes}} is returned.
}
\description{
Write a \code{\link[strataG]{gtypes}} object from a tidy data frame.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\references{
Archer FI, Adams PE, Schneiders BB.
strataG: An r package for manipulating, summarizing and analysing population
genetic data.
Molecular Ecology Resources. 2017; 17: 5-11. doi:10.1111/1755-0998.12559
}
\seealso{
\code{strataG.devel} is available on github \url{https://github.com/EricArcher/}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
