% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hapmap.R
\name{write_hapmap}
\alias{write_hapmap}
\title{Write a HapMap file from a tidy data frame}
\usage{
write_hapmap(data, filename = NULL)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.

The data requires nucleotide \code{A, C, G, T} information for the genotypes.

\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{filename}{(optional) The file name prefix for the hapmap file
written to the working directory. With default: \code{filename = NULL},
the date and time is appended to \code{radiator_hapmap}.}
}
\description{
Write a \href{https://www.ncbi.nlm.nih.gov/variation/news/NCBI_retiring_HapMap/}{HapMap file} from a tidy data frame.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
