% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_maverick.R
\name{write_maverick}
\alias{write_maverick}
\title{Write a maverick file from a tidy data frame}
\usage{
write_maverick(data, filename = NULL, ...)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{filename}{(optional) The file name prefix for the maverick file
written to the working directory. With default: \code{filename = NULL},
the date and time is appended to \code{radiator_maverick_}.}

\item{...}{other parameters passed to the function.}
}
\value{
A maverick file is saved to the working directory (with \code{.txt}).
}
\description{
Write a maverick file from a tidy data frame
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\references{
Verity R, Nichols RA (2016) Estimating the Number of
Subpopulations (K) in Structured Populations.
Genetics, 203, genetics.115.180992-1839.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
