% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pcadapt.R
\name{write_pcadapt}
\alias{write_pcadapt}
\title{Write a \href{https://github.com/bcm-uga/pcadapt}{pcadapt}
file from a tidy data frame}
\usage{
write_pcadapt(data, pop.select = NULL, snp.ld = NULL,
  maf.thresholds = NULL, maf.pop.num.threshold = 1, maf.approach = "SNP",
  maf.operator = "OR", filename = NULL,
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{pop.select}{(string, optional) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops).
Default: \code{pop.select = NULL}}

\item{snp.ld}{(optional) \strong{For VCF file only}.
SNP short distance linkage disequilibrium pruning. With anonymous markers from
RADseq/GBS de novo discovery, you can minimize linkage disequilibrium (LD) by
choosing among these 3 options: \code{"random"} selection, \code{"first"} or
\code{"last"} SNP on the same short read/haplotype. For long distance linkage
disequilibrium pruning, see details below.
Default: \code{snp.ld = NULL}.}

\item{maf.thresholds}{(string, double, optional) String with
local/populations and global/overall maf thresholds, respectively.
e.g. \code{maf.thresholds = c(0.05, 0.1)} for a local maf threshold
of 0.05 and a global threshold of 0.1. Available for VCF, PLINK and data frame
files.
Default: \code{maf.thresholds = NULL}.}

\item{maf.pop.num.threshold}{(integer, optional) When maf thresholds are used,
this argument is for the number of pop required to pass the maf thresholds
to keep the locus. Default: \code{maf.pop.num.threshold = 1}}

\item{maf.approach}{(character, optional).
\code{maf.approach = "haplotype"} : looks at the minimum MAF found on the
read/haplotype. Using this option will discard all the markers/snp on
that read based on the thresholds chosen. This method is only available
for VCF and haplotype files, or tidy data frame from those file types.
\code{maf.approach = "SNP"} : treats all the SNP on the same
haplotype/read as independent. Doesn't work with haplotype file,
but does work for all other file type.
Default is \code{maf.approach = "SNP"}.}

\item{maf.operator}{(character, optional) \code{maf.operator = "AND"} or
default \code{maf.operator = "OR"}.
When filtering over LOCUS or SNP, do you want the local \code{"AND"}
global MAF to pass the thresholds, or ... you want the local \code{"OR"}
global MAF to pass the thresholds, to keep the marker?}

\item{filename}{(optional) The file name prefix for the pcadapt file
written to the working directory. With default: \code{filename = NULL},
the date and time is appended to \code{radiator_pcadapt_}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during vcf import.
Default: \code{parallel::detectCores() - 1}.}
}
\value{
A pcadapt file is written in the working directory a genotype matrix
object is also generated in the global environment.
}
\description{
Write a
\href{https://github.com/bcm-uga/pcadapt}{pcadapt}
file from a tidy data frame. The data is biallelic.
Used internally in
\href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.

Also integrated in this function:
\enumerate{
\item \emph{pop.select: } to choose population to include in the pcadapt file.
\item \emph{maf: } evaluate the impact of Minor Allele Frequency on
genome scans with radiator maf module related arguments.
\item \emph{snp.ld: } SNPs on the same locus should not be considered
independant (short linkage disequilibrium) and the snp.ld argument
integrated here give users the ability to adress this quickly before
running a genome scan.
\item by defaults only markers found in common between populations are used
\item by defaults monomorphic markers are automatically removed before
generating the pcadapt file.
}
}
\details{
\emph{Integrated filters:} Only markers found in common between
populations are used and monomorphic markers are automatically removed
before generating pcadapt file.
}
\references{
Luu, K., Bazin, E., & Blum, M. G. (2017).
pcadapt: an R package to perform genome scans for selection based on principal component analysis.
Molecular Ecology Resources, 17(1), 67-77.

Duforet-Frebourg, N., Luu, K., Laval, G., Bazin, E., & Blum, M. G. (2015).
Detecting genomic signatures of natural selection with principal component analysis: application to the 1000 Genomes data.
Molecular biology and evolution, msv334.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
