% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_related.R
\name{write_related}
\alias{write_related}
\title{Write a related file from a tidy data frame}
\usage{
write_related(data, filename = NULL,
  parallel.core = parallel::detectCores() - 1, ...)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{filename}{(optional) The file name prefix for the related file
written to the working directory. With default: \code{filename = NULL},
the date and time is appended to \code{radiator_related_}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel::detectCores() - 1}.}

\item{...}{other parameters passed to the function.}
}
\value{
A related file is saved to the working directory.
}
\description{
Write a related file from a tidy data frame.
This output file format enables to run the data in the
\href{https://github.com/timothyfrasier/related}{related} R package
(Pew et al. 2015), which is essantially the R version of
\href{https://www.zsl.org/science/software/coancestry}{COANCESTRY} fortran
program developed by Jinliang Wang.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\references{
Pew J, Muir PH, Wang J, Frasier TR (2015)
related: an R package for analysing pairwise relatedness from codominant
molecular markers.
Molecular Ecology Resources, 15, 557-561.

Wang, J. 2011.
COANCESTRY: A program for simulating, estimating and analysing relatedness
and inbreeding coefficients.
Molecular Ecology Resources 11(1): 141-145.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
