% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rubias.R
\name{write_rubias}
\alias{write_rubias}
\title{Write a rubias object}
\usage{
write_rubias(
  data,
  strata = NULL,
  filename = NULL,
  parallel.core = parallel::detectCores() - 1
)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{strata}{(optional, tibble file or object) This tibble of individual's
metadata must contain
four columns: \code{SAMPLE_TYPE, REPUNIT, COLLECTION, INDIVIDUALS}. Those columns
are described in \strong{rubias}.
Default:\code{strata = NULL}. With default, \code{SAMPLE_TYPE} is filled
with \code{reference}. \code{REPUNIT} and \code{COLLECTION} will be filled by
the \code{STRATA} or \code{POP_ID} column found in the data.}

\item{filename}{The prefix for the name of the file written to the working directory.
Default: \code{filename = NULL}.
With default, only the rubias object is generated. The filename will be appended
\code{_rubias.tsv}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
A rubias object in the global environment and a file is written in the
working directory if \code{filename} argument was used.
}
\description{
Write a rubias object from a tidy data frame or GDS file/object.
}
\references{
Moran BM, and Anderson, E. C. 2018 Bayesian inference from the
conditional genetic stock identification model.
Can. J. Fish. Aquat. Sci. 76: 551-560.

Anderson, Eric C., Robin S. Waples, and Steven T. Kalinowski. (2008)
An improved method for predicting the accuracy of genetic stock identification.
Canadian Journal of Fisheries and Aquatic Sciences 65, 7:1475-1486.

Anderson, E. C. (2010) Assessing the power of informative subsets of
loci for population assignment: standard methods are upwardly biased.
Molecular ecology resources 10, 4:701-710.
}
\seealso{
/href{https://github.com/eriqande/rubias}{rubias}: genetic stock
identification (GSI) in the tidyverse.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
