% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_seqarray.R
\name{write_seqarray}
\alias{write_seqarray}
\title{Write a SeqArray GDS file from a vcf file and generate a connection object.}
\usage{
write_seqarray(vcf, filename = NULL, vcf.stats = FALSE,
  parallel.core = parallel::detectCores() - 1, verbose = TRUE, ...)
}
\arguments{
\item{vcf}{The path to the vcf file.}

\item{filename}{(optional) The file name of the Genomic Data Structure (GDS) file.
radiator will append \code{.gds} to the filename.
If filename chosen is already present in the
working directory, the default \code{radiator_datetime.gds} is chosen.
Default: \code{filename = NULL}.}

\item{vcf.stats}{(logical) When \code{vcf.stats = TRUE}, individual's missing
genotype proportion, averaged heterozygosity, total coverage, mean genotype
coverage and marker's metadata along count for ref and alt alleles and mean
coverage is generated and written in the working directory.
Default: \code{vcf.stats = FALSE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) To pass further argument for fine-tuning the function.}
}
\description{
Write a SeqArray \href{http://zhengxwen.github.io/SeqArray/}{SeqArray}
file (Zheng et al. 2017) from a vcf file and generate a connection object.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
A vcf file of 35 GB with ~4 millions SNPs take about ~7 min with 8 CPU.
A vcf file of 21 GB with ~2 millions SNPs take about ~5 min with 7 CPU.

After the file is generated, it's a matter of sec to open a connection.
}
\references{
Zheng X, Gogarten S, Lawrence M, Stilp A, Conomos M, Weir BS,
Laurie C, Levine D (2017). SeqArray -- A storage-efficient high-performance
data format for WGS variant calls.
Bioinformatics.

Danecek P, Auton A, Abecasis G et al. (2011)
The variant call format and VCFtools.
Bioinformatics, 27, 2156-2158.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
