% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stockr.R
\name{write_stockr}
\alias{write_stockr}
\title{Write a stockR dataset from a tidy data frame or GDS file or object.}
\usage{
write_stockr(data, filename = NULL, verbose = TRUE)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{write_seqarray}} or
\code{\link{tidy_vcf}}.}

\item{filename}{(optional) The stockr object is written in the working
directory. The file is written with \code{radiator_stockr_DATE@TIME.RData} and
can be open with readRDS.
Default: \code{filename = NULL}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\value{
The object generated is a matrix with
dimension: MARKERS x INDIVIDUALS. The genotypes are coded like PLINK:
0, 1 or 2 alternate allele. 0: homozygote for the reference allele,
1: heterozygote, 2: homozygote for the alternate allele.
Missing genotypes have NA. The object also as 2 attributes.
\code{attributes(data)$grps} with \code{STRATA/POP_ID} of the individuals and
\code{attributes(data)$sample.grps} filled with \code{INDIVIDUALS}.
Both attributes can be used inside \emph{stockR}.
}
\description{
Write a stockR dataset (Fost et al. submitted).
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\references{
Foster et al. submitted
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
