% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_structure.R
\name{write_structure}
\alias{write_structure}
\title{Write a structure file from a tidy data frame}
\usage{
write_structure(data, pop.levels = NULL, markers.line = TRUE,
  filename = NULL, ...)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
White spaces in population names are replaced by underscore.
Default: \code{pop.levels = NULL}.}

\item{markers.line}{(optional, logical) In the structure
file, you can write the markers on a single line separated by
commas \code{markers.line = TRUE},
or have markers on a separate line, i.e. in one column, for the structure file
(not very useful with thousands of markers) and not printed at all for the
structure file.
Default: \code{markers.line = TRUE}.}

\item{filename}{(optional) The file name prefix for the structure file
written to the working directory. With default: \code{filename = NULL},
the date and time is appended to \code{radiator_structure_}.}

\item{...}{other parameters passed to the function.}
}
\value{
A structure file is saved to the working directory.
}
\description{
Write a structure file from a tidy data frame
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
}
\references{
Pritchard JK, Stephens M, Donnelly P. (2000)
Inference of population structure using multilocus genotype data.
Genetics. Genetics Society of America. 155: 945–959.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
