% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{write_vcf}
\alias{write_vcf}
\title{Write a vcf file from a tidy data frame}
\usage{
write_vcf(data, strata = FALSE, filename = NULL, source = NULL, empty = FALSE)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{strata}{(optional, logical) Should the strata information be
included in the FORMAT field (along the GT info for each samples ?). To make
the VCF population-ready use \code{strata = TRUE}. The strata information
must be included in the \code{STRATA} column of the tidy dataset.
Default: \code{strata = FALSE}. Experimental.}

\item{filename}{(optional) The file name prefix for the vcf file
written to the working directory. With default: \code{filename = NULL},
the date and time is appended to \code{radiator_vcf_file_}.}

\item{source}{source of vcf}

\item{empty}{generate an empty vcf.
Default: \code{empty = FALSE}.}
}
\description{
Write a vcf file (file format version 4.3, see details below)
from a tidy data frame.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
\strong{VCF file format version:}

If you need a different file format version than the current one, just change
the version inside the newly created VCF, that should do the trick.
\href{https://vcftools.github.io/specs.html}{For more
information on Variant Call Format specifications}.
}
\references{
Danecek P, Auton A, Abecasis G et al. (2011)
The variant call format and VCFtools.
Bioinformatics, 27, 2156-2158.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
