% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas_estimator.R
\name{betas_estimator}
\alias{betas_estimator}
\title{Estimate \eqn{\beta}s per population}
\usage{
betas_estimator(data, strata = NULL, filename = NULL, verbose = FALSE)
}
\arguments{
\item{data}{14 options for input (\strong{diploid data only}): VCFs (SNPs or Haplotypes,
to make the vcf population ready),
plink (tped, bed), stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop, DArT,
and a data frame in long/tidy or wide format. To verify that radiator detect
your file format use \code{\link{detect_genomic_format}} (see example below).
Documented in \strong{Input genomic datasets} of \code{\link{tidy_genomic_data}}.

\strong{DArT and VCF data}: \pkg{radiator} was not meant to generate alleles
and genotypes if you are using a VCF file with no genotype
(only genotype likelihood: GL or PL).
Neither is \pkg{radiator} able to magically generate a genind object
from a SilicoDArT dataset. Please look at the first few lines of your dataset
to understand it's limit before asking raditor to convert or filter your dataset.}

\item{strata}{(optional)
The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}. Documented in \code{\link{read_strata}}.
DArT data: a third column \code{TARGET_ID} is required.
Documented on \code{\link{read_dart}}. Also use the strata read function to
blacklist individuals.
Default: \code{strata = NULL}.}

\item{filename}{(optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in
the working directory. The file extension appended to
the \code{filename} provided is \code{.rad}.
With default: \code{filename = NULL}, the tidy data frame is
in the global environment only (i.e. not written in the working directory...).}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\value{
A list is created with 3 objects:
betaiovl: Average \eqn{\beta_i} over loci,
Hw: Within population gene diversities
Hb: Between populations gene diversities
}
\description{
Estimate \eqn{\beta}s per population.
}
\examples{
\dontrun{
# Using a  VCF file, the simplest for of the function:
fh <- ibdg_fh(
data = "batch_1.vcf",
strata = "strata.panda.tsv"
)
# To see what's inside the list
names(fh)
# To view the manhattan plot:
fh$fh.manhattan.plot
# To view the distribution of FH values:
fh$fh.distribution.plot
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
