% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_seqarray.R
\name{boxplot_stats}
\alias{boxplot_stats}
\title{boxplot_stats}
\usage{
boxplot_stats(data, title, x.axis.title = NULL, y.axis.title,
  facet.columns = FALSE, facet.rows = FALSE, bp.filename = NULL)
}
\description{
Generate a boxplot
}
\keyword{internal}
