% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{check_header_source_vcf}
\alias{check_header_source_vcf}
\title{Check the vcf header and detect vcf source}
\usage{
check_header_source_vcf(vcf)
}
\description{
Check the vcf header and detect vcf source
}
\keyword{internal}
