% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{dart2gds}
\alias{dart2gds}
\title{dart2gds}
\usage{
dart2gds(data, strata = NULL, markers.meta, filename, dart.format,
  gt.vcf = NULL, gt.vcf.nuc = NULL, gt = NULL,
  parallel.core = parallel::detectCores() - 1, verbose = TRUE)
}
\description{
Transform dart to GDS format
}
\keyword{internal}
