% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_biallelic_markers.R
\name{detect_biallelic_markers}
\alias{detect_biallelic_markers}
\title{Detect biallelic data}
\usage{
detect_biallelic_markers(data, verbose = FALSE,
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty
during execution.
Default: \code{verbose = FALSE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
A logical character string (TRUE/FALSE). That answer the question if
the data set is biallelic or not.
}
\description{
Detect if markers in tidy dataset are biallelic.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
