% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_genomic_format.R
\name{detect_genomic_format}
\alias{detect_genomic_format}
\title{Used internally in radiator to detect the file format}
\usage{
detect_genomic_format(data)
}
\arguments{
\item{data}{A genomic data set in the global environment}
}
\value{
One of these file format:
\itemize{
\item tbl_df: for a data frame
\item genind: for a genind object
\item genlight: for a genlight object
\item gtypes: for a gtypes object
\item vcf.file: for a vcf file
\item plink.file: for a plink file
\item genepop.file: for a genepop file
\item haplo.file: for a stacks haplotypes file
\item fstat.file: for a fstat file
\item dart: for a DArT file
\item fst.file: for a file ending with .rad
}
}
\description{
Detect file format of genomic data set.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
