% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_het_outliers.R
\name{detect_het_outliers}
\alias{detect_het_outliers}
\title{Detect heterozygotes outliers and estimate miscall rate}
\usage{
detect_het_outliers(
  data,
  nreps = 2000,
  burn.in = NULL,
  strata = NULL,
  filename = NULL,
  parallel.core = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{data}{14 options for input: VCFs (SNPs or Haplotypes,
to make the vcf population ready),
plink (tped, bed), stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop, DArT,
and a data frame in long/tidy or wide format. To verify that radiator detect
your file format use \code{\link{detect_genomic_format}} (see example below).
Documented in \strong{Input genomic datasets} of \code{\link{tidy_genomic_data}}.

\strong{DArT and VCF data}: \pkg{radiator} was not meant to generate alleles
and genotypes if you are using a VCF file with no genotype
(only genotype likelihood: GL or PL).
Neither is \pkg{radiator} able to magically generate a genind object
from a SilicoDArT dataset. Please look at the first few lines of your dataset
to understand it's limit before asking raditor to convert or filter your dataset.}

\item{nreps}{(integer, optional) The number of MCMC sweeps to do.
Default: \code{nreps = 2000}.}

\item{burn.in}{(integer, optional) The number of MCMC burn-in reps.
With default, during execution, you will be asked to enter the nuber of burn-in.
For this, a plot showing the heterozygote miscall rate for all
the MCMC sweeps will be printed. This plot will help pinpoint the
number of burn-in. The remaining MCMC sweeps will be used
to average the heterozygote miscall rate.
e.g. of common value \code{burn.in = 500}.
With default: \code{burn.in = NULL}.}

\item{strata}{(path or object) The strata file or object.
Additional documentation is available in \code{\link{read_strata}}.
Use that function to whitelist/blacklist populations/individuals.
Option to set \code{pop.levels/pop.labels} is also available.}

\item{filename}{(optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in
the working directory. The file extension appended to
the \code{filename} provided is \code{.rad}.
With default: \code{filename = NULL}, the tidy data frame is
in the global environment only (i.e. not written in the working directory...).}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{...}{(optional) To pass further arguments for fine-tuning the function.}
}
\value{
A folder generated automatically with date and time,
the file \code{het.summary.tsv} contains the summary statistics. The file
\code{markers.genotypes.boundaries.pdf} is the plot with boundaries.
The overall genotyping and heterozygotes miscall rate is writen in the file
\code{overall_error_rate.tsv}.
The function also returns a list inside the global environment with
8 objects:

\enumerate{
\item input the input data, cleaned if filters were used during import.
\item outlier.summary a list with a tibble and plot of genotypes frequencies
and boundaries (also written in the folder).
\item summary.alt.allele a tibble summarizing the number of markers with:
\itemize{
\item no homozygote for the alternate allele (NO_HOM_ALT)
\item no heterozygote genotype (NO_HET)
\item one homozygote for the alternate allele(ONE_HOM_ALT)
\item one heterozygote genotype (ONE_HET)
\item one homozygote for the alternate allele only (ONE_HOM_ALT_ONLY)
\item one heterozygote genotype only (ONE_HET_ONLY)
\item one homozygote for the alternate allele and one heterozygote genotype only (ONE_HOM_ALT_ONE_HET_ONLY)
}
\item m.nreps A tibble with the heterozygote miscall rate for each MCMC replicate
\item overall.genotyping.error.rate The overall genotyping error rate
\item overall.m The overall heterozygote miscall rate
\item simmed_genos The simulated genotypes
}

The statistics are summarized per population and overall,
the grouping is found in the last column called \code{POP_ID}.
}
\description{
Explore departure from H-W equilibrium in bi-allelic RADseq data.
Highlight excess of homozygotes present in numeros RADseq studies.
The function estimate the genotyping error rate and heterozygote miscall rate.
The model focus on heterozygotes being
incorrectly called as homozygotes. See details below for more info.
}
\details{
\strong{Before using the function:}
\enumerate{
\item Don't use raw RADseq data, this function will work best with filtered data
\item Remove duplicate \code{\link[radiator]{detect_duplicate_genomes}}.
\item Remove mixed samples \code{\link[radiator]{detect_mixed_genomes}}.
\item Look at other filters in radiator package...
}

\strong{During import:}

By default the function will keep only polymorphic markers and markers common
between all populations. If you supply a tidy data frame or a \code{.rad} file,
the function skip all the filters, pop selection, etc. It will however scan and
remove monomorphic markers automatically.

\strong{Keep track of the data:}

Use the argument filename to write the imported (and maybe further filtered)
tidy genomic data set inside the folder. The filename will be automatically
appended \code{.rad} to it. This file can be used again directly inside this
function and other radiator functions. See \code{\link[radiator]{read_rad}}.
}
\examples{
\dontrun{
het.prob <- radiator::detect_het_outliers(
data = "tuna.vcf", strata = "tuna.strata.tsv", nreps = 2000)
}
}
\author{
Eric Anderson \email{eric.anderson@noaa.gov} and
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
