% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_ref_genome.R
\name{detect_ref_genome}
\alias{detect_ref_genome}
\title{Detect reference genome}
\usage{
detect_ref_genome(chromosome = NULL, data = NULL, verbose = TRUE)
}
\arguments{
\item{chromosome}{(string) String with chromosome ID, unique or not.
Default: \code{chromosome = NULL}.}

\item{data}{(optional, path or object) radiator GDS object or file. When this
argument is used, the function look inside the GDS radiator node for:
\itemize{
\item \code{reference.genome} node and \code{TRUE/FALSE} value
\item \code{markers.meta} node and detect the information using the \code{CHROM}
column
\item \code{chromosome} node directly.
}
Default: \code{data = NULL}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\value{
TRUE/FALSE if dataset is assembled with a reference genome or not.
}
\description{
Detect if the dataset was de novo or reference genome assembled
}
\examples{
\dontrun{
ref.genome <- radiator::detect_ref_genome(chromosome = tidy.data$CHROM)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
