% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_maf.R
\name{diagnostic_maf}
\alias{diagnostic_maf}
\title{MAF diagnostic}
\usage{
diagnostic_maf(data, group.rank, filename = NULL)
}
\arguments{
\item{data}{A file in the working directory or object in the global environment
in wide or long (tidy) formats. To import, the function uses
\href{https://github.com/thierrygosselin/radiator}{radiator}
\code{\link[radiator]{tidy_wide}}.
\emph{See details of this function for more info}.}

\item{group.rank}{(Number) The number of group to class the MAF.}

\item{filename}{(optional) Name of the file written to the working directory.}
}
\description{
Minor Allele Frequency diagnostic, help choose a filter threshold.
}
\details{
Highly recommended to look at the distribution of MAF
\link{plot_density_distribution_maf}
}
\examples{
\dontrun{
problem <- radiator::diagnostic_maf(
data = tidy.salmon.data, group.rank = 10)
}
}
\seealso{
\link{filter_maf}
}
