% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{extract_coverage}
\alias{extract_coverage}
\title{extract_coverage}
\usage{
extract_coverage(gds, markers = TRUE, ind = TRUE, update.gds = FALSE,
  parallel.core = parallel::detectCores() - 2, verbose = FALSE)
}
\arguments{
\item{gds}{The gds object.}

\item{markers}{(optional, logical) Default:\code{markers = TRUE}.}

\item{ind}{(optional, logical) Default:\code{ind = TRUE}.}

\item{update.gds}{(optional, logical) Default:\code{update.gds = FALSE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\description{
Extract coverage information from a GDS file
}
