% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{extract_coverage}
\alias{extract_coverage}
\title{extract_coverage}
\usage{
extract_coverage(gds, markers = TRUE, ind = TRUE, depth = NULL)
}
\description{
Extract coverage information from a GDS file
}
\keyword{internal}
