% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{extract_dart_target_id}
\alias{extract_dart_target_id}
\title{Extract \href{http://www.diversityarrays.com}{DArT} target id}
\usage{
extract_dart_target_id(data, write = TRUE)
}
\arguments{
\item{data}{One of the DArT output files. 6 formats used by DArT are recognized
by radiator.
recognised:
\enumerate{
\item \code{1row}: Genotypes are in 1 row and coded (0, 1, 2, -).
\code{0 for 2 reference alleles REF/REF}, \code{1 for 2 alternate alleles ALT/ALT},
\code{2 for heterozygote REF/ALT}, \code{- for missing}.
\item \code{2rows}: No genotypes. It's absence/presence, 0/1, of the REF and ALT alleles.
Sometimes called binary format.
\item \code{counts}: No genotypes, It's counts/read depth for the REF and ALT alleles.
Sometimes just called count data.
\item \code{silico.dart}: SilicoDArT data. No genotypes, no REF or ALT alleles.
It's a file coded as absence/presence, 0/1, for the presence of sequence in
the clone id.
\item \code{silico.dart.counts}: SilicoDArT data. No genotypes, no REF or ALT alleles.
It's a file coded as absence/presence, with counts for the presence of sequence in
the clone id.
\item \code{dart.vcf}: For DArT VCFs, please use \code{\link{read_vcf}}.
}

Depending on the number of markers, these format will be recoded similarly to
VCF files (dosage of alternate allele, see details).

The function can import \code{.csv} or \code{.tsv} files.


If you encounter a problem, sent me your data so that I can update
the function.}

\item{write}{With default \code{write = TRUE}, the dart target id column is
written in a file in the working directory.}
}
\value{
A tidy dataframe with a \code{TARGET_ID} column. For cleaning, the \code{TARGET_ID}
column is treated like the column \code{INDIVIDUALS}. Spaces and \code{,}
are removed, \code{_} and \code{:} are changed to a dash \code{-} and
UPPER case is used.
\href{https://thierrygosselin.github.io/radiator/reference/clean_ind_names.html}{see cleaning doc for logic behind this}.
}
\description{
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users. The function allows to extract DArT
target id from a DArT file. To help prepare the appropriate STRATA file.
}
\examples{
\dontrun{
# Built a strata file:
strata <- radiator::extract_dart_target_id("mt.dart.file.csv") \%>\%
    dplyr::mutate(
        INDIVIDUALS = "new id you want to give",
        STRATA = "fill this"
    ) \%>\%
    readr::write_tsv(x = ., path = "my.new.dart.strata.tsv")
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com} and Peter Grewe \email{peter.grewe@csiro.au}
}
