% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{extract_individuals_metadata}
\alias{extract_individuals_metadata}
\title{extract_individuals_metadata}
\usage{
extract_individuals_metadata(gds, ind.field.select = NULL,
  radiator.node = TRUE, whitelist = FALSE, blacklist = FALSE,
  verbose = FALSE)
}
\arguments{
\item{gds}{The gds object.}

\item{ind.field.select}{(optional, character) Default:\code{ind.field.select = NULL}.}

\item{radiator.node}{(optional, logical) Default:\code{radiator.node = TRUE}.}

\item{whitelist}{(optional, logical) Default:\code{whitelist = FALSE}.}

\item{blacklist}{(optional, logical) Default:\code{blacklist = FALSE}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\description{
Import gds or radiator individuals node
}
