% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{extract_markers_metadata}
\alias{extract_markers_metadata}
\title{extract_markers_metadata}
\usage{
extract_markers_metadata(gds, markers.meta.select = NULL,
  radiator.node = TRUE, verbose = FALSE)
}
\description{
Import gds or radiator markers meta node
}
\keyword{internal}
