% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_common_markers.R
\name{filter_common_markers}
\alias{filter_common_markers}
\title{Filter common markers between strata}
\usage{
filter_common_markers(
  data,
  filter.common.markers = TRUE,
  fig = FALSE,
  parallel.core = parallel::detectCores() - 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{filter.common.markers}{(optional, logical)
Default: \code{filter.common.markers = TRUE}.}

\item{fig}{(optional, logical) \code{fig = TRUE} will produce an
\href{https://github.com/hms-dbmi/UpSetR}{UpSet fig} to visualize the number
of markers between populations. The package is required for this to work...
Default: \code{fig = FALSE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty
during execution.
Default: \code{verbose = FALSE}.}

\item{...}{(optional) To pass further arguments for fine-tuning the function
and legacy arguments.}
}
\value{
A list with the filtered input, whitelist and blacklist of markers..
}
\description{
The function will filter the markers by keeping only those
in common between all strata
(population or any groupings defined in \code{STRATA/POP_ID} column).

\strong{Filter targets}: SNPs

\strong{Statistics}: strata genotyping rate per SNPs


Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users who wants to keep only markers in common.
}
\examples{
\dontrun{
require(SeqArray) # when using gds
common <- radiator::filter_common_markers(data = "my.radiator.gds.rad", verbose = TRUE)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
