% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_radiator_functions.R
\name{filter_dart}
\alias{filter_dart}
\title{filter DArT file}
\usage{
filter_dart(...)
}
\arguments{
\item{...}{Read life cycle section below}
}
\description{
Use \code{\link{filter_rad}}.
}
\section{Life cycle}{


As of radiator v.0.0.22, \code{filter_dart} is now replaced by
\code{\link{filter_rad}} in an effort to have more meaningful functions names.
This is THE function to rule them all.
}

