% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dart.R
\name{filter_dart}
\alias{filter_dart}
\title{Swiss Army knife tool to prepare \href{http://www.diversityarrays.com}{DArT}
output file for population genetics analysis.}
\usage{
filter_dart(interactive.filter = TRUE, data, strata, output = NULL,
  pop.levels = NULL, blacklist.id = NULL, pop.select = NULL,
  monomorphic.out = TRUE, common.markers = TRUE,
  filter.reproducibility = NULL, filter.coverage = NULL,
  filter.call.rate = NULL, filter.ind.missing.geno = NULL,
  number.snp.reads = NULL, maf.thresholds = NULL,
  maf.pop.num.threshold = 1, maf.approach = "SNP", maf.operator = "OR",
  mixed.genomes.analysis = TRUE, ind.heterozygosity.threshold = NULL,
  duplicate.genomes.analysis = c(TRUE, FALSE), snp.ld = NULL,
  imputation.method = NULL, hierarchical.levels = "populations",
  num.tree = 50, filename = NULL, verbose = TRUE,
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. The user is asked to see figures of distribution before
making decisions for filtering.
Default: \code{interactive.filter = TRUE}.}

\item{data}{DArT output file with genotype in 1 or 2 row \emph{binary} format
used in CSIRO genomic projects.}

\item{strata}{A tab delimited file with columns header:
\code{INDIVIDUALS} and \code{STRATA}.
Note: the column \code{STRATA} refers to any grouping of individuals. If a third column
named \code{NEW_ID} is used, this column will be used to replace the
\code{INDIVIDUALS} column in the main data file.}

\item{output}{19 genomic data formats can be exported: tidy (by default),
genind, genlight, vcf (for file format version, see details below), plink, genepop,
structure, arlequin, hierfstat, gtypes (strataG), bayescan, betadiv, pcadapt,
hzar, snprelate (for SNPRelate's GDS).
Use a character string,
e.g. \code{output = c("genind", "genepop", "structure")}, to have preferred
output formats generated. The tidy format is generated automatically.
Default: \code{output = NULL}.}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
White spaces in population names are replaced by underscore.
Default: \code{pop.levels = NULL}.}

\item{blacklist.id}{(optional) A blacklist with individual ID and
a column header 'INDIVIDUALS'. The blacklist is an object in your
global environment or a file in the working directory
(e.g. "blacklist.txt").
Default: \code{blacklist.id = NULL}.}

\item{pop.select}{(string, optional) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops).
Default: \code{pop.select = NULL}}

\item{monomorphic.out}{(optional) Should the monomorphic
markers present in the dataset be filtered out ?
Default: \code{monomorphic.out = TRUE}.}

\item{common.markers}{(optional) Logical. Default: \code{common.markers = TRUE},
will only keep markers in common (genotyped) between all the populations.}

\item{filter.reproducibility}{(optional, numerical) Filter the \code{RepAvg}
column in the data set. Default: \code{filter.reproducibility = NULL}.
e.g to keep markers with reproducibility >= 99%,
use: \code{filter.reproducibility = 0.99}.}

\item{filter.coverage}{(optional, string, numerical) Filter the lower and
upper bound of locus/read coverage. The locus/read coverage combines the markers
average count for REF and ALT allele (respectively the \code{AvgCountRef} and
\code{AvgCountSnp} info).
Default: \code{filter.coverage = NULL}.
e.g to keep markers with coverage inbetween 7 and 200,
use : \code{filter.coverage = c(7, 200)}.}

\item{filter.call.rate}{(optional, numerical) Filter the \code{CallRate}
column in the data set. Default: \code{filter.call.rate = NULL}. e.g to keep
markers genotyped in more than 95% of the individuals use :
\code{filter.call.rate = 0.95}}

\item{filter.ind.missing.geno}{(optional, string) Similar to call rate, but
more adapted to the data. 3 values are required in the string, corresponding
to the \code{\link[radiator]{filter_individual}} module of radiator.
First value is the approach to count genotyped individuals, \code{"overall"}
or by \code{"pop"}. Second value is the percent threshold for the marker, with
\code{70}, 70 percent of genotyped individuals are required to keep the marker.
The last threshold is the number of problematic population that are allowed to skip
the threshold. In doubt, use the interactive mode that take step by step these
arguments. e.g to keep individuals genotyped at >= 70 percent for the markers,
without considering the population info and allowing 1 population to be problematic for the
threshold, use: \code{c("overall", 70, 1)}.
Default: \code{filter.ind.missing.geno = NULL}.}

\item{number.snp.reads}{(optional, integer) This filter removes outlier markers
with too many SNP number per locus/read.
Having a higher than "normal" SNP number is usually the results of
assembly artifacts or bad assembly parameters.
This filter is population-agnostic. This is best decide after viewing the figures,
with the interactive mode.
If the argument is set to \code{number.snp.reads = 2},
locus with 3 and more SNPs will be blacklisted.
Default: \code{number.snp.reads = NULL}.}

\item{maf.thresholds}{(string, double, optional) String with
local/populations and global/overall maf thresholds, respectively.
e.g. \code{maf.thresholds = c(0.05, 0.1)} for a local maf threshold
of 0.05 and a global threshold of 0.1. Available for VCF, PLINK and data frame
files.
Default: \code{maf.thresholds = NULL}.}

\item{maf.pop.num.threshold}{(integer, optional) When maf thresholds are used,
this argument is for the number of pop required to pass the maf thresholds
to keep the locus. Default: \code{maf.pop.num.threshold = 1}}

\item{maf.approach}{(character, optional).
\code{maf.approach = "haplotype"} : looks at the minimum MAF found on the
read/haplotype. Using this option will discard all the markers/snp on
that read based on the thresholds chosen. This method is only available
for VCF and haplotype files, or tidy data frame from those file types.
\code{maf.approach = "SNP"} : treats all the SNP on the same
haplotype/read as independent. Doesn't work with haplotype file,
but does work for all other file type.
Default is \code{maf.approach = "SNP"}.}

\item{maf.operator}{(character, optional) \code{maf.operator = "AND"} or
default \code{maf.operator = "OR"}.
When filtering over LOCUS or SNP, do you want the local \code{"AND"}
global MAF to pass the thresholds, or ... you want the local \code{"OR"}
global MAF to pass the thresholds, to keep the marker?}

\item{mixed.genomes.analysis}{(optional, logical) Highlight outliers individual's
observed heterozygosity for a quick
diagnostic of mixed samples or poor polymorphism discovery due to DNA quality,
sequencing effort, etc.
See this function for more info: \code{\link[radiator]{detect_mixed_genomes}}.
Default: \code{detect_mixed_genomes = TRUE}.}

\item{ind.heterozygosity.threshold}{(string, double, optional)
Blacklist individuals based on observed heterozygosity (averaged across markers).


The string contains 2 thresholds values (min and max).
The values are proportions (0 to 1), where 0 turns off the min threshold and
1 turns off the max threshold.
Individuals with mean observed heterozygosity higher (>) or lower (<)
than the thresholds will be blacklisted.

Default: \code{ind.heterozygosity.threshold = NULL} will turn off completely
the filter and the function will only output the plots and table of heterozygosity.}

\item{duplicate.genomes.analysis}{(optional, string) Detect duplicate individuals.
The function can compute two methods (distance or genome pairwise similarity)
to highligh potential duplicate individuals.
See this function for more info: \code{\link[radiator]{detect_duplicate_genomes}}.
The string required to run the analysis as 2 values:
\enumerate{
\item TRUE/FALSE to run the analysis;
\item Computes pairwise genome similarity (TRUE/FALSE),
with FALSE just the distance measure is used.
The pairwise genome similarity is longer to run, but is better because it
integrates markers in common/missing data.
Using \code{interactive.filter = TRUE}, can overide this value,
you can opt in for the pairwise genome similarity after viewing the figures
used with distance measure... handy!
}
Default: \code{detect_duplicate_genomes = c(TRUE, FALSE)}.}

\item{snp.ld}{(optional) \strong{For data with locus and SNP info, like VCF and DArT file}.
SNP short distance linkage disequilibrium pruning. With anonymous markers from
RADseq/GBS de novo discovery, you can minimize linkage disequilibrium (LD) by
choosing among these 3 options: \code{"random"} selection, \code{"first"} or
\code{"last"} SNP on the same short read/haplotype. For long distance linkage
disequilibrium pruning, see details below.
Default: \code{snp.ld = NULL}.}

\item{imputation.method}{(character, optional)
Methods available for map-independent imputations of missing genotype:

(1) \code{imputation.method = "max"} Strawman imputation,
the most frequently observed genotypes (ties are broken at random).

(2) \code{imputation.method = "rf"} On-the-fly-imputations using
Random Forests algorithm.

(3) \code{imputation.method = "boost"} extreme gradient boosting trees.

(4) \code{imputation.method = "mca"} Multiple Correspondence Analysis.

\code{imputation.method = NULL} the function will stop.
Default: \code{imputation.method = NULL}.}

\item{hierarchical.levels}{(character, optional) \code{c("global", "populations")}.
Should the imputations be computed globally or by populations. Note that
imputing genotype globally in conjunction with \code{imputation.method = "max"}
can create huge bias for example by
introducing foreign genotypes in some populations (see note for more info).
Default: \code{hierarchical.levels = "populations"}.}

\item{num.tree}{(integer, optional) The number of trees to grow
when \code{imputation.method = "rf"} or \code{imputation.method = "rf_pred"}.
Default: \code{num.tree = 50}.}

\item{filename}{(optional) The filename prefix for the objet in the global environment
or the working directory. Default: \code{filename = NULL}. A default name will be used,
customized with the output file(s) selected.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during vcf import.
Default: \code{parallel::detectCores() - 1}.}
}
\value{
The function returns an object (list). The content of the object
can be listed with \code{names(object)} and use \code{$} to isolate specific
object (see examples). Some output format will write the output file in the
working directory. The tidy genomic data frame is generated automatically.
}
\description{
Import, filter and generate imputed dataset of DArT output file.
}
\examples{
\dontrun{
testing
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com} and Peter Grewe \email{peter.grewe@csiro.au}
}
