% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genotyping.R
\name{filter_genotyping}
\alias{filter_genotyping}
\title{Filter markers based on genotyping/missing rate}
\usage{
filter_genotyping(
  data,
  interactive.filter = TRUE,
  filter.genotyping = NULL,
  filename = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. With default: \code{interactive.filter == TRUE}, figures and
tables are shown before making decisions for filtering.}

\item{filter.genotyping}{(optional, string) 2 options:
\itemize{
\item character string \code{filter.genotyping = "outliers"} will use as
thresholds the higher outlier values in the box plot.
\item double \code{filter.genotyping = 0.2}. Will allow up to 0.2 missing genotypes.
}
Default: \code{filter.genotyping = NULL}.}

\item{filename}{(optional, character)
Default: \code{filename = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function. Also used for legacy arguments (see details or
special section)}
}
\value{
With \code{interactive.filter = FALSE}, a list in the global environment,
with 7 objects:
\enumerate{
\item $tidy.filtered.mac
\item $whitelist.markers
\item $blacklist.markers
\item $mac.data
\item $filters.parameters
}

With \code{interactive.filter = TRUE}, a list with 4 additionnal objects are generated.
\enumerate{
\item $distribution.mac.global
\item $distribution.mac.local
\item $mac.global.summary
\item $mac.helper.table
}
}
\description{
This function is designed to remove/blacklist markers
based on genotyping/call rate.

\strong{Filter targets}: SNPs

\strong{Statistics}: mean genotyping/call rate (missingness information).
}
\section{Advance mode}{


\emph{dots-dots-dots ...} allows to pass several arguments for fine-tuning the function:
\enumerate{
\item \code{filter.common.markers} (optional, logical).
Default: \code{filter.common.markers = FALSE},
Documented in \code{\link{filter_common_markers}}.
\item \code{filter.monomorphic} (logical, optional) Should the monomorphic
markers present in the dataset be filtered out ?
Default: \code{filter.monomorphic = TRUE}.
Documented in \code{\link{filter_monomorphic}}.
\item \code{path.folder}: to write ouput in a specific path
(used internally in radiator).
Default: \code{path.folder = getwd()}.
If the supplied directory doesn't exist, it's created.
}
}

\section{Interactive version}{


To help choose a threshold use the interactive version.

2 steps in the interactive version:

Step 1. Visualization and helper table.

Step 2. Filtering markers based on mean genotyping/missing rate
}

\examples{
\dontrun{
# The minumum
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
