% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_maf.R
\name{filter_maf}
\alias{filter_maf}
\title{MAF/MAC filter}
\usage{
filter_maf(
  interactive.filter = TRUE,
  data,
  strata = NULL,
  maf.thresholds = NULL,
  filename = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. With default: \code{interactive.filter == TRUE}, figures and
tables are shown before making decisions for filtering.}

\item{data}{13 options for input: VCFs (SNPs or Haplotypes,
to make the vcf population ready),
plink, stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop, DArT,
and a data frame in long/tidy or wide format. To verify that radiator detect
your file format use \code{\link{detect_genomic_format}} (see example below).
Documented in \strong{Input genomic datasets} of \code{\link{tidy_genomic_data}}.

\strong{DArT and VCF data}: \pkg{radiator} was not meant to generate alleles
and genotypes if you are using a VCF file with no genotype
(only genotype likelihood: GL or PL).
Neither is \pkg{radiator} able to magically generate a genind object
from a SilicoDArT dataset. Please look at the first few lines of your dataset
to understand it's limit before asking raditor to convert or filter your dataset.}

\item{strata}{(optional)
The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}. Documented in \code{\link{read_strata}}.
DArT data: a third column \code{TARGET_ID} is required.
Documented on \code{\link{read_dart}}. Also use the strata read function to
blacklist individuals.
Default: \code{strata = NULL}.}

\item{maf.thresholds}{(optional, string) 5 values required in the string.

Example using maf.thresholds with SNP and frequency approach:
\code{maf.thresholds = c("SNP", 0.001, "OR", 0.001, 1)}.

Example using maf.thresholds with locus and count approach:
\code{maf.thresholds = c("locus", 3, "OR", 6, 1)}.


\enumerate{
\item \code{maf approach} (character: "SNP"/"locus"):
MAF filtering is conducted by SNPs or locus.
\code{"SNP"}: will consider all the SNPs on the same locus/read as independent
and will be filtered independently of their locus id.
\code{"locus"}: looks at the minimum MAF found on the
read/locus. Using this option will discard all the markers/snp on
that read based on the thresholds chosen. For the locus approach to work, your dataset
requires SNP and Locus info (e.g. from a VCF file).

\item \code{local} threshold (double or integer): For a frequency threshold use
a double (e.g. 0.05). For a count threshold, use an integer
(e.g. 3 for the number of alternate allele required in a population). This
threshold is applied by population.
Not sure about the threshold to use, choose the interactive mode argument.

\item \code{operator} (character: "OR" / "AND"):
To consider both the local \code{"AND"} the global thresholds, use \code{"AND"}.
To consider the local \code{"OR"} the global thresholds, use \code{"OR"}.

\item \code{global} threshold (double or integer): For a frequency threshold
use a double (e.g. 0.02). For a count threshold, use an integer
(e.g. 6 for the number of alternate allele required). This threshold is
applied at the dataset level (no population).
Not sure about the threshold to use, choose the interactive mode argument.

\item \code{maf pop num threshold} (integer)
The number of pop required to pass the thresholds
to keep the marker. Usually, I always use \code{1}, for 1 pop is required to
pass thresholds and keep the marker.
}}

\item{filename}{(optional, character) Write to folder the MAF filtered
tidy dataset. The name will be appended \code{.rad}. With default, the filtered
data is only in the global environment.
Default: \code{filename = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) To pass further arguments for fine-tuning the function
and legacy arguments (see Life cycle section below).}
}
\value{
With \code{interactive.filter = FALSE}, a list in the global environment,
with 7 objects:
\enumerate{
\item $tidy.filtered.maf
\item $whitelist.markers
\item $blacklist.markers
\item $maf.data
\item $filters.parameters
}

With \code{interactive.filter = TRUE}, a list with 4 additionnal objects are generated.
\enumerate{
\item $distribution.maf.global
\item $distribution.maf.local
\item $maf.global.summary
\item $maf.helper.table
}

\strong{maf.helper.table:}

First and second variables (in columns) represents POP_ID and sample size (n).

The last 2 rows are the local MAF (suggested based on the lowest pop value)
and the TOTAL/GLOBAL observations.

Columns starting with ALT are the variable corresponding
to the number of alternative (ALT) allele (ranging from 1 to 20).
The observations in the ALT allele variable columns are the local (for the pop)
and global (last row) MAF of your dataset.
e.g. ALT_3 can potentially represent 3 heterozygote individuals with
the ALT allele or 1 homozygote individuals for the ALT allele and
1 heterozygote individual. And so on...
}
\description{
Minor Allele Filter.
Remove markers based on Minor/Alternate Allele Frequency (MAF) or Count (MAC).

Most arguments are inherited from
\code{\link[radiator]{tidy_genomic_data}}.
}
\section{Life cycle}{


The \strong{maf.thresholds} arguments used in the past in several of
radiator functions
is still available inside \code{filter_maf} arguments.
But it is in evaluation mode and the function will probaly be deprecated in
future releases. The arguments are complicated to use and
most RADseq datesets encountered so far are better filtered
using \code{\link{filter_mac}}.
}

\section{MAC or MAF ?}{

Using count or frequency to remove a SNPs ? The preferred choice in radiator
as changed from frequency to count, because we think the filtering should not
alter the spectrum and this is only achieved if the same criteria is applied
for each SNP.


Even small differences in missing data between RADseq markers
generates differences in MAF frequency thresholds applied.

Example with a datset consisting of N = 36 individuals and 3 SNPs
with varying level of missing genotypes:

\itemize{
\item \code{SNP number : number samples genotypes : REF/ALT counts}
\item SNP1 : 36 : 69/3
\item SNP2 : 30 : 65/3
\item SNP3 : 24 : 45/3
}

Each SNPs have the same alternate allele count, corresponding to
2 individuals with the polymorphism: 1 homozygote + 1 heterozygote.
Applying a MAF threshold of
0.05 would mean that SNP3 would be blacklisted
(\code{24 * 2 * 0.05 = 2.4 alt alleles required to pass}).


\strong{Using count instead of frequency allows each RADseq markers,
with varying missing data, to be treated equally.}
}

\section{Interactive version}{


To help choose a threshold for the local and global MAF
use the interactive version.

2 steps in the interactive version:

Step 1. Global and Local MAF visualization and helper table.

Step 2. Filtering markers based on the different MAF arguments
}

\examples{
\dontrun{
# The minumum
turtle.maf <- radiator::filter_maf(
        data = "turtle.vcf",
        strata = "turtle.strata.tsv")
# This will use the default: interactive version,
# a list is created and to view the filtered tidy data:
tidy.data <- turtle.maf$tidy.filtered.maf

# The remaining argument are used in tidy_genomic_data during import and allow
# the user to apply some filtering or selection before doing the MAF filtering.

# If I want to filter Alternate alleles based on count.
turtle.maf <- radiator::filter_maf(
        data = "turtle.vcf",
        strata = "turtle.strata.tsv",
        maf.thresholds = c("SNP", 3, "OR", "5", 1),
        filename = "turtle.maf")

# This will remove monomorphic markers (by default filter.monomorphic = TRUE)
# it will use the SNP approache (SNPs are considered independent and
# filtered independent)
# I keep markers if: they have a local count of 3 alternate allele OR global
# count of 5 alternate allele.
# I keep the marker if at leat 1 pop pass that.
# finally the filtered data will be written in the directory under the name:
# turtle.maf.rad
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
