% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_monomorphic.R
\name{filter_monomorphic}
\alias{filter_monomorphic}
\title{Filter monomorphic markers}
\usage{
filter_monomorphic(data, verbose = FALSE)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.
The function can also use a Genomic Data Structure (GDS) object generated from
\pkg{radiator} \code{\link{write_seqarray}}, \code{\link{tidy_vcf}},
\code{\link{tidy_genomic_data}}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty
during execution.
Default: \code{verbose = FALSE}.}
}
\value{
A list with the filtered input file, whitelist and blacklist of markers removed.
}
\description{
Filter monomorphic markers. This filter will remove from the dataset
markers with just one genotype coding
(e.g. all homozygotes for REF or ALT, but also all heterozygotes).


Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\seealso{
\code{\link{write_seqarray}}, \code{\link{tidy_vcf}},
\code{\link{tidy_genomic_data}}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
