% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_monomorphic.R
\name{filter_monomorphic}
\alias{filter_monomorphic}
\title{Filter monomorphic markers}
\usage{
filter_monomorphic(
  data,
  filter.monomorphic = TRUE,
  parallel.core = parallel::detectCores() - 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{filter.monomorphic}{(optional, logical)
Default: \code{filter.monomorphic = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function. Also used for legacy arguments (see details or
special section)}
}
\value{
A list with the filtered input, whitelist and blacklist of markers..
}
\description{
Filter monomorphic markers.
This filter will remove from the dataset
markers with just \emph{one genotype phenotype}:
\itemize{
\item genotypes are ALL homozygotes REF/REF (pp)
\item genotypes are ALL heterozygotes REF/ALT, ALT/REF (pq or qp)
\item genotypes are ALL homozygotes ALT/ALT (qq)
}

\strong{Filter targets}: SNPs

\strong{Statistics}: the number of genotype phenotypes

Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users who wants to keep only polymorphic markers in
their dataset.
}
\examples{
\dontrun{
require(SeqVarTools) # when using gds
mono <- radiator::filter_monomorphic(data = "my.radiator.gds.rad", verbose = TRUE)
}
}
\seealso{
\code{\link{filter_rad}},
\code{\link{tidy_genomic_data}}, \code{\link{read_vcf}},
\code{\link{tidy_vcf}}.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
