% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rad.R
\name{filter_rad}
\alias{filter_rad}
\title{ONE FUNCTION TO RULE THEM ALL}
\usage{
filter_rad(data, strata = NULL, interactive.filter = TRUE,
  output = NULL, filename = NULL, verbose = TRUE,
  parallel.core = parallel::detectCores() - 1, ...)
}
\arguments{
\item{data}{12 options for input: VCFs (SNPs or Haplotypes,
to make the vcf population ready),
plink, stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop, DArT,
and a data frame in long/tidy or wide format. To verify that radiator detect
your file format use \code{\link{detect_genomic_format}} (see example below).
Documented in \strong{Input genomic datasets} of \code{\link{tidy_genomic_data}}.}

\item{strata}{(optional)
The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}. Documented in \code{\link{read_strata}}.
DArT data: a third column \code{TARGET_ID} is required.
Documented on \code{\link{read_dart}}. Also use the strata read function to
blacklist individuals.
Default: \code{strata = NULL}.}

\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. Figures of distribution are shown before asking for filtering
thresholds.
Default: \code{interactive.filter = TRUE}.}

\item{output}{26 genomic data formats can be exported: tidy (by default),
genepop, genind, genlight, vcf (for file format version, see details below),
plink, structure, faststructure, arlequin, hierfstat, gtypes (strataG), bayescan, betadiv,
pcadapt, hzar, fineradstructure, related, seqarray, snprelate and maverick.
Use a character string,
e.g. \code{output = c("genind", "genepop", "structure")}, to have preferred
output formats generated. With default, only the tidy format is generated.
Default: \code{output = NULL}.}

\item{filename}{(optional) The filename prefix for the objet in the global environment
or the working directory. Default: \code{filename = NULL}. A default name will be used,
customized with the output file(s) selected.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{...}{(optional) To pass further argument for fine-tuning the function.}
}
\value{
The function returns an object (list). The content of the object
can be listed with \code{names(object)} and use \code{$} to isolate specific
object (see examples). Some output format will write the output file in the
working directory. The tidy genomic data frame is generated automatically.
}
\description{
Designed for RADseq data, it's radiator integrated pipeline that links
several \code{filter_} functions of \pkg{radiator}.
Rapidly get an idea of what you can and cannot do with your dataset.
Novices, start with this one!
}
\section{Filtering steps and folders generated}{

\enumerate{
\item \strong{radiator}:
\itemize{
\item the GDS file ending with \code{.gds.rad}. To open in R use \code{\link{read_rad}}.
\item \code{filters_parameters.tsv}: file containing all the parameters and
values of the filtering process.
\item \code{radiator_filter_rad_args.tsv}: the function calls and values, for
reproducibility.
\item \code{radiator_tidy_dart_metadata.rad}: for DArT data, this is the markers
metadata file. To open in R use \code{\link{read_rad}}.
\item \code{random.seed}: for reproducibility.
}
\item \strong{filter_dart_reproducibility}: filter reproducibility of markers
(DArT data only). Described in \code{\link[radiator]{filter_dart_reproducibility}}.
\itemize{
\item \code{blacklist.dart.reproducibility.tsv}: blacklisted markers.
\item \code{whitelist.dart.reproducibility.tsv}: whitelisted markers.
\item \code{dart_reproducibility_boxplot.pdf}: the boxplot of reproducibility.
\item \code{dart_reproducibility_stats.tsv}: reproducibility associated summary statistics.
\item \code{dart.reproducibility.helper.plot.pdf}: the helper plot showing
the impact of thresholds
on the number of markers blacklisted and whitelisted.
\item \code{dart.reproducibility.helper.table.tsv}: a tibble with
the impact of thresholds on the the number
of markers blacklisted and whitelisted.
\item \code{radiator_filter_dart_reproducibility_args.tsv}: the function calls and values.
}

\item \strong{filter_monomorphic}: removes monomorphic markers.
Described in \code{\link[radiator]{filter_monomorphic}}.
\itemize{
\item \code{blacklist.monomorphic.markers.tsv}: blacklisted markers.
\item \code{whitelist.polymorphic.markers}: whitelisted markers.
\item \code{radiator_filter_monomorphic_args}: the function calls and values.
}
\item \strong{filter_common_markers}: keep only markers in common between strata.
Described in \code{\link[radiator]{filter_common_markers}}.
\itemize{
\item \code{common.markers.upsetrplot.pdf}: the UpSetR plot highlight the number
of markers in common between strata.
\item \code{blacklist.common.markers.tsv}: blacklisted markers.
\item \code{whitelist.common.markers.tsv}: whitelisted markers.
\item \code{radiator_filter_common_markers_args.tsv}: the function calls and values.
}

\item \strong{filter_individuals}: blacklist individuals based on missingness
and/or heterozygosity and/or total coverage.
Described in \code{\link[radiator]{filter_individuals}}.
\itemize{
\item \code{individuals.qc.pdf}: several boxplots showing different individual
quality metrics.
\item \code{individuals.qc.stats.tsv}: tibble with individual's
proportion of missing genotypes, heterozygosity, total and mean coverage.
\item \code{individuals.qc.stats.summary.tsv}: individual's summary statistics.
\item \code{blacklist.individuals.missing.tsv}: blacklisted individuals based
on missingness (genotyping rate of individuals).
\item \code{radiator_filter_individuals_args.tsv}: the function calls and values.
}
The function will remove automatically monomorphic markers if individuals are removed.
\item \strong{filter_mac}: remove/blacklist markers based on Minor/Alternate Allele Count (MAC).
Described in \code{\link[radiator]{filter_mac}}.
\itemize{
\item \code{distribution.mac.global.pdf}: distribution of overall MAC.
\item \code{mac.boxplot.pdf}: boxplot of the MAC.
\item \code{mac.global.tsv}: a tibble with the global MAC and MAF.
\item \code{mac.markers.plot.pdf}: the helper plot showing
the impact of thresholds
on the number of markers blacklisted and whitelisted.
\item \code{mac.helper.table.tsv}: a tibble with
the impact of thresholds on the the number
of markers blacklisted and whitelisted.
\item \code{mac.summary.stats.tsv}: MAC summary statistics.
\item \code{blacklist.markers.mac.tsv}: blacklisted markers.
\item \code{whitelist.markers.mac.tsv}: whitelisted markers.
\item \code{radiator_filter_mac_args.tsv}: the function calls and values.
}
\item \strong{filter_coverage}: remove/blacklist markers based on mean coverage information.
Described in \code{\link[radiator]{filter_coverage}}.
\itemize{
\item \code{markers_metadata.tsv}: all the markers coverage metadata available.
\item \code{markers_metadata_stats.tsv}: summary statistics of coverage information.
\item \code{markers_qc.pdf}: several coverage boxplots (total coverage, mean coverage, etc.)
\item \code{coverage.low.helper.plot.pdf}: the helper plot showing
the impact of thresholds
on the number of markers blacklisted and whitelisted.
\item \code{coverage.low.helper.table.tsv}: a tibble with
the impact of thresholds on the the number
of markers blacklisted and whitelisted.
\item \code{coverage.high.helper.plot.pdf}: the helper plot showing
the impact of thresholds
on the number of markers blacklisted and whitelisted.
\item \code{coverage.high.helper.table.tsv}: a tibble with
the impact of thresholds on the the number
of markers blacklisted and whitelisted.
\item \code{blacklist.markers.coverage.tsv}: blacklisted markers.
\item \code{whitelist.markers.coverage.tsv}: whitelisted markers.
\item \code{radiator_filter_coverage_args.tsv}: the function calls and values.
}
\item \strong{filter_genotyping}: remove/blacklist markers based on genotyping/call rate.
Described in \code{\link[radiator]{filter_genotyping}}.
\itemize{
\item \code{markers_qc.pdf}: the missing genotypes boxplot.
\item \code{markers_metadata.tsv}: the missing proportion per markers along other stats.
\item \code{markers_metadata_stats.tsv}: the summary statistics of markes genotyping rate..
\item \code{markers.genotyping.helper.plot.pdf}: the helper plot showing
the impact of thresholds
on the number of markers blacklisted and whitelisted.
\item \code{genotyping.helper.table.tsv}: a tibble with
the impact of thresholds on the the number
of markers blacklisted and whitelisted (overall).
\item \code{markers.pop.missing.helper.table.tsv}: a tibble with
the impact of thresholds on the the number
of markers blacklisted and whitelisted (strata).
\item \code{blacklist.markers.genotyping.tsv}: blacklisted markers.
\item \code{whitelist.markers.genotyping.tsv}: whitelisted markers.
\item \code{radiator_filter_genotyping_args.tsv}: the function calls and values.
}
\item \strong{filter_snp_position_read}: removes markers/SNPs based on their
position on the read
Described in \code{\link[radiator]{filter_snp_position_read}}.
\itemize{
\item \code{snp.position.read.boxplot.pdf}: boxplot of SNP position on the read
\item \code{snp.position.read.helper.table.tsv}: a tibble with
the impact of thresholds on the the number
of markers blacklisted and whitelisted.
\item \code{snp.position.read.distribution.pdf}: distribution of the SNP position
on the read.
\item \code{blacklits.markers.snp.position.read.tsv}: blacklisted markers.
\item \code{whitelist.markers.snp.position.read.tsv}: whitelisted markers.
\item \code{radiator_filter_snp_position_read_args.tsv}: the function calls and values.
}
\item \strong{filter_snp_number}: removes outlier markers with too many SNP
number per locus/read.
Described in \code{\link[radiator]{filter_snp_number}}.
\itemize{
\item \code{markers_metadata.tsv}: metadata associated with the number of SNP/locus
\item \code{snp_per_locus.pdf}: boxplot of number of SNP/locus.
\item \code{snp_per_locus_distribution.pdf}: distribution on the number od
\item \code{snp.per.locus.helper.plot.pdf}: the helper plot showing
the impact of thresholds
on the number of markers blacklisted and whitelisted.
\item \code{blacklist.snp.per.locus.tsv}: blacklisted markers.
\item \code{whitelist.snp.per.locus.tsv}: whitelisted markers.
\item \code{radiator_filter_snp_number_args.tsv}: the function calls and values.
}
\item \strong{filter_ld}: SNP short and long distance linkage disequilibrium
pruning.
Described in \code{\link[radiator]{filter_ld}}.
\itemize{
\item \code{short.ld.locus.stats.tsv}: number of locus with more than 1, 2, 3 SNPs.
\item \code{blacklist.short.ld.tsv}: blacklisted markers.
\item \code{whitelist.short.ld.tsv}: whitelisted markers.
\item \code{blacklist.long.ld.tsv}: blacklisted markers.
\item \code{whitelist.long.ld.tsv}: whitelisted markers.
\item \code{radiator_filter_ld_args.tsv}: the function calls and values.
}
\item \strong{detect_mixed_genomes}: highlight outliers individual's.
observed heterozygosity.
Described in \code{\link[radiator]{detect_mixed_genomes}}.
\itemize{
\item \code{individuals.qc.stats.tsv}: individual's heterozygosity and
missingness proportion.
\item \code{individuals.qc.stats.summary.tsv}: heterozygosity summary statistics.
\item \code{heterozygosity.statistics.tsv}: heterozygosity summary statistics
per strata and overall.
\item \code{individual.heterozygosity.manhattan.plot.pdf}: manhattan plot
highlighting potential patterns of heterozygosity and missingness.
\item \code{individual.heterozygosity.boxplot.pdf}: boxplot
highlighting potential patterns of heterozygosity and missingness.
\item \code{blacklist.ind.het.tsv}: blacklisted individuals.
\item \code{radiator_detect_mixed_genomes_args.tsv}: the function calls and values.
}
The function will remove automatically monomorphic markers if individuals are removed.
\item \strong{detect_duplicate_genomes}: highligh potential duplicate individuals.
Described in \code{\link[radiator]{detect_duplicate_genomes}}.
\itemize{
\item \code{genotyped.statistics.tsv}: genotyping statistics used in the analysis.
\item \code{individuals.pairwise.dist.tsv}: pairwise distances measures.
\item \code{individuals.pairwise.distance.stats.tsv}: summary statistics of
pairwise distances measures.
\item \code{individuals.qc.stats.tsv}: individuals summary produced by default.
\item \code{individuals.qc.stats.summary.tsv}: individuals summary produced by default.
\item \code{manhattan.plot.distance.pdf}: manhattan plot highlighting duplicates.
\item \code{violin.plot.distance.pdf}: boxplot highlighting duplicates.
\item \code{blacklist.id.similar.tsv}: blacklisted individuals.
\item \code{radiator_detect_duplicate_genomes_args.tsv}: the function calls and values.
}
\item \strong{filter_hwe}: remove/blacklist markers based on Hardy-Weinberg Equilibrium
Described in \code{\link[radiator]{filter_hwe}}.
\itemize{
\item \code{genotypes.summary.tsv}: summary of genotypes groups.
\item \code{hwd.helper.table.tsv}: number of markers blacklisted based on thresholds.
\item \code{hw.pop.sum.tsv}: summary of hwe/hwd.
\item \code{hwd.plot.blacklist.markers.pdf}: helper plot.
\item \code{hwe.manhattan.plot.pdf}: overview of markers in HWD.
\item \code{hwe.ternary.plots.missing.data.pdf}: overview of markers in HWD.
\item \code{blacklisted and whitelisted markers based on thresholds}.
\item \code{radiator_filter_hwe_args.tsv}: the function calls and values.
}

\item \strong{filtered}:
\itemize{
\item \code{.rad} file is the tidy data set. To open in R use \code{\link{read_rad}}.
\item \code{strata.filtered.tsv} the filtered strata file.
\item \code{blacklist.id.tsv}: blacklisted individuals (total).
\item \code{blacklist.markers.tsv}: blacklisted markers.
\item \code{whitelist.markers.tsv}: whitelisted markers.
\item \code{markers_qc.pdf}: markers qc after filters.
\item \code{individuals.qc.pdf}: individuals qc after filters.
\item \code{individuals.qc.stats.tsv}: individuals qc stats.
\item \code{individuals.qc.stats.summary.tsv}: individuals qc stats summary.
\item \code{markers_metadata.tsv}: markers final metadata.
\item \code{markers_metadata_stats.tsv}: markers metadata stats.
\item \code{radiator_genind_.RData} and \code{radiator_stockr.RData}, the
genind and stockr files generated by the output argument.
}
}
}

\section{Advance mode}{


Ideally, use the function in interactive mode and forget about this section.
For advance user who want to run in batch mode with pre-defined arguments.
Below are the arguments available. Read the function documentation associated
with the arguments. Most of them reside in separate module that can be explored
separately.
\emph{dots-dots-dots ...} allows to pass several arguments for fine-tuning the function:
\enumerate{
\item \strong{filter.reproducibility}: detailed in \code{\link[radiator]{filter_dart_reproducibility}}.
\item \strong{filter.monomorphic}: detailed in \code{\link[radiator]{filter_monomorphic}}.
\item \strong{filter.common.markers}: detailed in \code{\link[radiator]{filter_common_markers}}.
\item \code{whitelist.markers}: detailed in \code{\link[radiator]{filter_whitelist}}.
\item \strong{filter.individuals.missing}: detailed in \code{\link[radiator]{filter_individuals}}.
\item \strong{filter.individuals.heterozygosity}: detailed in \code{\link[radiator]{filter_individuals}}.
\item \strong{filter.individuals.coverage.total}: detailed in \code{\link[radiator]{filter_individuals}}.
\item \strong{filter.mac}: detailed in \code{\link[radiator]{filter_mac}}.
\item \strong{filter.coverage}: detailed in \code{\link[radiator]{filter_coverage}}.
\item \strong{filter.genotyping}: described in \code{\link[radiator]{filter_genotyping}}.
\item \strong{filter.snp.position.read}: described in \code{\link[radiator]{filter_snp_position_read}}.
\item \strong{filter.snp.number}: described in \code{\link[radiator]{filter_snp_number}}.
\item \strong{filter.short.ld}: described in \code{\link[radiator]{filter_ld}}.
\item \strong{filter.long.ld}: described in \code{\link[radiator]{filter_ld}}.
\item \strong{long.ld.missing}: described in \code{\link[radiator]{filter_ld}}.
\item \strong{ld.method}: described in \code{\link[radiator]{filter_ld}}.
\item \strong{detect.mixed.genomes}: described in \code{\link[radiator]{detect_mixed_genomes}}.
\item \strong{detect.duplicate.genomes}: described in \code{\link[radiator]{detect_duplicate_genomes}}.
\item \strong{dup.threshold}: described in \code{\link[radiator]{detect_duplicate_genomes}}.
\item \strong{filter.hwe}: described in \code{\link[radiator]{filter_hwe}}.
\item \strong{hw.pop.threshold}: described in \code{\link[radiator]{filter_hwe}}.
\item \strong{midp.threshold}: described in \code{\link[radiator]{filter_hwe}}.
\item \strong{filter.hwe}: described in \code{\link[radiator]{filter_hwe}}.
\item \strong{filter.hwe}: described in \code{\link[radiator]{filter_hwe}}.
\item \strong{ind.heterozygosity.threshold}: described in \code{\link[radiator]{detect_mixed_genomes}}.
\item \strong{ind.heterozygosity.threshold}: described in \code{\link[radiator]{detect_mixed_genomes}}.
\item \strong{ind.heterozygosity.threshold}: described in \code{\link[radiator]{detect_mixed_genomes}}.
}
}

\examples{
\dontrun{
# Some of the packages used in this function are not installed by default.
# Installed the required packages:
radiator_pkg_install()

# Very simple:
shark <- radiator::filter_rad(
    data = "data.shark.vcf",
    strata = "strata.shark.tsv")


# With filename and output
shark <- radiator::filter_rad(
    data = "data.shark.vcf",
    strata = "strata.shark.tsv",
    output = "genind",
    filename = "shark")
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com} and
Peter Grewe \email{peter.grewe@csiro.au}
}
