% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_snp_number.R
\name{filter_snp_number}
\alias{filter_snp_number}
\title{Filter SNP number per locus/read}
\usage{
filter_snp_number(interactive.filter = TRUE, data, strata = NULL,
  filter.snp.number = NULL, filename = NULL,
  parallel.core = parallel::detectCores() - 1, verbose = TRUE, ...)
}
\arguments{
\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. Figures of distribution are shown before asking for filtering
thresholds.
Default: \code{interactive.filter = TRUE}.}

\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{strata}{(path or object) The strata file or object.
Additional documentation is available in \code{\link{read_strata}}.
Use that function to whitelist/blacklist populations/individuals.
Option to set \code{pop.levels/pop.labels} is also available.}

\item{filter.snp.number}{(integer) This is best decided after viewing the figures.
If the argument is set to 2, locus with 3 and more SNPs will be blacklisted.
Default: \code{filter.snp.number = NULL}.}

\item{filename}{(optional) Name of the filtered tidy data frame file
written to the working directory (ending with \code{.tsv})
Default: \code{filename = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function. Also used for legacy arguments (see details or
special section)}
}
\value{
A list in the global environment with 6 objects:
\enumerate{
\item $snp.number.markers
\item $number.snp.reads.plot
\item $whitelist.markers
\item $tidy.filtered.snp.number
\item $blacklist.markers
\item $filters.parameters
}

The object can be isolated in separate object outside the list by
following the example below.
}
\description{
This filter removes outlier markers
with too many SNP number per locus/read.
The data requires snp and locus information (e.g. from a VCF file).
Having a higher than "normal" SNP number is usually the results of
assembly artifacts or bad assembly parameters.
This filter is population-agnostic, but still requires a strata
file if a vcf file is used as input.

\strong{Filter targets}: Markers

\strong{Statistics}: The number of SNPs per locus.
}
\details{
\strong{Interactive version}

There are 2 steps in the interactive version to visualize and filter
the data based on the number of SNP on the read/locus:

Step 1. SNP number per read/locus visualization

Step 2. Choose the filtering thresholds
}
\examples{
\dontrun{
turtle.outlier.snp.number <- radiator::filter_snp_number(
data = "turtle.vcf",
strata = "turtle.strata.tsv",
max.snp.number = 4,
filename = "tidy.data.turtle.tsv"
)

tidy.data <- turtle.outlier.snp.number$tidy.filtered.snp.number

#Inside the same list, to isolate the markers blacklisted:
blacklist <- turtle.outlier.snp.number$blacklist.markers

}
}
