% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{gds2tidy}
\alias{gds2tidy}
\title{gds2tidy}
\usage{
gds2tidy(
  gds,
  markers.meta = NULL,
  markers.meta.select = NULL,
  wide = FALSE,
  individuals = NULL,
  pop.id = TRUE,
  calibrate.alleles = TRUE,
  strip.rad = FALSE,
  parallel.core = parallel::detectCores() - 1,
  close.gds = FALSE,
  ...
)
}
\description{
GDS to tidy...
}
\keyword{internal}
