% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator.R
\name{generate_folder}
\alias{generate_folder}
\title{generate_folder}
\usage{
generate_folder(f, rad.folder = NULL, internal = FALSE,
  file.date = NULL, prefix_int = TRUE, verbose = FALSE)
}
\arguments{
\item{f}{Folder name}

\item{rad.folder}{Name of the rad folder}

\item{internal}{(optional, logical) Is the function internal or not}

\item{file.date}{The file date included}

\item{prefix_int}{Use an integer prefix padded left with 0.
Default: \code{prefix_int = TRUE}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\description{
Generate a folder based on ...
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
