% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator.R
\name{generate_folder}
\alias{generate_folder}
\title{generate_folder}
\usage{
generate_folder(
  rad.folder = NULL,
  path.folder = NULL,
  internal = FALSE,
  append.date = TRUE,
  file.date = NULL,
  prefix.int = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{rad.folder}{Name of the rad folder}

\item{path.folder}{path of the folder}

\item{internal}{(optional, logical) Is the function internal or not}

\item{append.date}{Include the date and time with the folder.
Default: \code{append.date = TRUE}.}

\item{file.date}{The file date included in as argument/value or with
default \code{file.date = NULL}, generated by the fucntion.}

\item{prefix.int}{Use an integer prefix padded left with 0.
Default: \code{prefix.int = TRUE}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\description{
Generate a folder based on ...
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
