% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duplicate_genomes.R
\name{genome_similarity}
\alias{genome_similarity}
\title{genome_similarity}
\usage{
genome_similarity(list.pair, input.prep = NULL, ...)
}
\description{
for the parallel part
}
\keyword{internal}
