% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_radiator_functions.R
\name{keep_common_markers}
\alias{keep_common_markers}
\title{Keep markers in common between all populations}
\usage{
keep_common_markers(data, plot = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{plot}{(optional, logical) \code{plot = TRUE} will produce an
\href{https://github.com/hms-dbmi/UpSetR}{UpSet plot} to visualize the number
of markers between populations. The package is required for this to work...
Default: \code{plot = FALSE}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty
during execution.
Default: \code{verbose = FALSE}.}
}
\value{
A list with the filtered input data set,
the whitelist of markers in common and the blacklist of markers discarded.
}
\description{
Keep markers in common between all populations.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\section{Life cycle}{


As of radiator v.0.0.22, keep.common.markers function is deprecated and was replaced by
\code{\link{filter_common_markers}} in an effort to have more meaningful functions names.
}

\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
