% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ma.R
\name{minor_allele_stats}
\alias{minor_allele_stats}
\title{minor_allele_stats}
\usage{
minor_allele_stats(
  data,
  calibrate.alleles = c("count", "depth", "ancestral"),
  blacklist.markers = NULL,
  markers.meta = NULL,
  path.folder = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE
)
}
\description{
Generate Minor Allele necessary statistics
}
\keyword{internal}
