% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_vcf.R
\name{parse_gds_metadata}
\alias{parse_gds_metadata}
\title{parse_gds_metadata}
\usage{
parse_gds_metadata(x, data = NULL, verbose = TRUE,
  parallel.core = parallel::detectCores() - 1)
}
\description{
function to parse the format field and tidy the results of VCF
}
\keyword{internal}
