% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_vcf.R
\name{parse_genomic}
\alias{parse_genomic}
\title{parse_genomic}
\usage{
parse_genomic(x, data = NULL, mask = FALSE, gather.data = FALSE,
  return.alleles = FALSE, strata = NULL, verbose = TRUE)
}
\description{
function to parse the format field and tidy the results of VCF
}
\keyword{internal}
