% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{pi}
\alias{pi}
\title{Nucleotide diversity}
\usage{
pi(
  data,
  read.length = NULL,
  parallel.core = parallel::detectCores() - 1,
  path.folder = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{read.length}{(integer, optional) The length in nucleotide of your reads.
By default it is estimated from the data using the column \code{COL}.
Default: \code{read.length = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{path.folder}{(path, optional) By default will print results in the working directory.
Default: \code{path.folder = NULL}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\value{
The function returns a list with the function call and:
\enumerate{
\item $pi.individuals: the pi estimated for each individual
\item $pi.populations: the pi statistics estimated per populations and overall.
\item $boxplot.pi: showing the boxplot of Pi for each populations and overall.
}
use $ to access each #' objects in the list.
}
\description{
Calculates the nucleotide diversity (Nei & Li, 1979).

To get an estimate with the consensus reads, use the
function \emph{summary_haplotypes} found in the package
[stackr](https://github.com/thierrygosselin/stackr). The estimate in
\emph{summary_haplotypes} integrates the consensus markers found in
[STACKS](http://catchenlab.life.illinois.edu/stacks/)
populations.haplotypes.tsv file.
Both radiator and stackr functions requires \code{stringdist} package.

The \code{read.length} argument below is used directly in the calculations.
To be correctly estimated, the reads obviously need to be of identical size...
}
\examples{
\dontrun{
require(stringdist)
# The simplest way to run the function:
pi.sum <- radiator::pi(data = "brook.charr.gds")
}
}
\references{
Nei M, Li WH (1979)
Mathematical model for studying genetic variation in terms of
restriction endonucleases.
Proceedings of the National Academy of Sciences of
the United States of America, 76, 5269–5273.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
