% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_boxplot_diversity}
\alias{plot_boxplot_diversity}
\title{Box plot of the diversity (Gene and Haplotypes)}
\usage{
plot_boxplot_diversity(data, aes.x.y, y.title)
}
\arguments{
\item{data}{The hapstats summary file or object.}

\item{aes.x.y}{The GGPLOT2 aesthetics,
e.g. aes.x.y = aes(x = factor(POP_ID), y = GENE_DIVERSITY, na.rm = T).}

\item{y.title}{Title of the y-axis.}
}
\description{
GGPLOT2 Box plot of the diversity (Gene and Haplotypes).
}
\keyword{internal}
