% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_histogram}
\alias{plot_histogram}
\alias{plot_density_distribution_coverage}
\title{Figure of distribution (histogram)}
\usage{
plot_histogram(data, variable, x.axis.title, y.axis.title)

plot_density_distribution_coverage(data, aes.colour, adjust.bin)
}
\arguments{
\item{data}{Coverage summary file.}

\item{aes.colour}{GGPLOT2 aesthetics colour,
e.g. aes(y = ..scaled.., color = COVERAGE_GROUP).}

\item{adjust.bin}{Adjust GGPLOT2 bin size (0 to 1).}
}
\description{
Create density distribution of coverage summary statistics.
Use the coverage summary file created with coverage_summary function.

Create density distribution of coverage summary statistics.
Use the coverage summary file created with coverage_summary function.
}
\keyword{internal}
