% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_common_markers.R
\name{plot_upset}
\alias{plot_upset}
\title{plot_upset}
\usage{
plot_upset(x, data.type, plot.filename = NULL,
  parallel.core = parallel::detectCores() - 2, verbose = FALSE)
}
\description{
Generate UpSetR plot
}
\keyword{internal}
