% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_alleles.R
\name{private_alleles}
\alias{private_alleles}
\title{Find private alleles}
\usage{
private_alleles(data, strata = NULL)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{strata}{(optional)
The strata file is a tab delimited file (in directory or global environment)
with 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}.
The \code{STRATA} column can be any hierarchical grouping.
The default uses the grouping in the dataset.
Default: \code{strata = NULL}.}
}
\value{
A list with an object highlighting private alleles by markers and strata and
a second object with private alleles summary by strata.
}
\description{
The function uses a tidy genomic data frame
}
\examples{
\dontrun{
corals.private.alleles.by.pop <- radiator::private_alleles(data = tidy, strata = strata.pop)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
