% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator.R
\name{radiator_folder}
\alias{radiator_folder}
\title{radiator_folder}
\usage{
radiator_folder(f, path.folder = NULL, prefix_int = TRUE)
}
\arguments{
\item{f}{Folder name}

\item{path.folder}{path of the folder}

\item{prefix_int}{Use an integer prefix padded left with 0.
Default: \code{prefix_int = TRUE}.}
}
\description{
Generate the rad folders
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
