% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_future.R
\name{radiator_future}
\alias{radiator_future}
\title{radiator parallel function}
\usage{
radiator_future(
  .x,
  .f,
  parallel.core = parallel::detectCores() - 1,
  split.with = NULL,
  flat.future = c("int", "chr", "dfr", "dfc", "walk", "drop"),
  ...
)
}
\description{
Updating radiator to use future
}
\keyword{internal}
