% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{radiator_future}
\alias{radiator_future}
\title{radiator parallel function}
\usage{
radiator_future(
  .x,
  .f,
  flat.future = c("int", "chr", "dfr", "dfc", "walk", "drop"),
  split.vec = FALSE,
  split.with = NULL,
  split.chunks = 4L,
  parallel.core = parallel::detectCores() - 1,
  ...
)
}
\description{
Updating radiator to use future
}
\keyword{internal}
