% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{radiator_gds}
\alias{radiator_gds}
\title{radiator gds constructor}
\usage{
radiator_gds(
  genotypes = NULL,
  biallelic = TRUE,
  data.source = NULL,
  geno.coding = c("alt.dos", "ref.dos"),
  strata = NULL,
  markers.meta = NULL,
  genotypes.meta = NULL,
  dp = NULL,
  ad = NULL,
  gl = NULL,
  filename = NULL,
  open = FALSE,
  verbose = TRUE
)
}
\description{
helper function to construct a radiator gds
}
\keyword{internal}
