% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_parallel.R
\name{mclapply_win}
\alias{mclapply_win}
\alias{.radiator_parallel_mc}
\alias{.radiator_parallel}
\title{hack to switch function for parallel computation based on OS}
\usage{
mclapply_win(X, FUN, ..., mc.preschedule = TRUE, mc.set.seed = TRUE,
  mc.silent = FALSE, mc.cores = NULL, mc.cleanup = TRUE,
  mc.allow.recursive = TRUE)

.radiator_parallel_mc(X, FUN, ..., mc.preschedule = TRUE,
  mc.set.seed = TRUE, mc.silent = FALSE, mc.cores = getOption("mc.cores",
  2L), mc.cleanup = TRUE, mc.allow.recursive = TRUE)

.radiator_parallel(X, FUN, ..., mc.style = "ETA", mc.substyle = NA,
  mc.cores = getOption("mc.cores", 2L),
  ignore.interactive = getOption("ignore.interactive", F))
}
\arguments{
\item{X}{see \pkg{parallel} \code{\link{mclapply}}}

\item{FUN}{see \pkg{parallel} \code{\link{mclapply}}}

\item{...}{see \pkg{parallel} \code{\link{mclapply}}}

\item{mc.preschedule}{see \pkg{parallel} \code{\link{mclapply}}}

\item{mc.set.seed}{see \pkg{parallel} \code{\link{mclapply}}}

\item{mc.silent}{see \pkg{parallel} \code{\link{mclapply}}}

\item{mc.cores}{see \pkg{parallel} \code{\link{mclapply}}}

\item{mc.cleanup}{see \pkg{parallel} \code{\link{mclapply}}}

\item{mc.allow.recursive}{see \pkg{parallel} \code{\link{mclapply}}}
}
\value{
For mclapply, a list of the same length as X and named by X.

For mclapply, a list of the same length as X and named by X.
}
\description{
\code{parallel::mclapply} doesn't work on Windows,
because forking is not supported.
This function defines a socket version of mclapply for windows computer
An implementation that switch automatically the parallel process when detecting
the os.
The code below was inspired from
\pkg{parallel} \code{\link{mclapply}},
\href{https://github.com/nathanvan}{Nathan VanHoudnos},
\href{https://github.com/kvnkuang/pbmcapply}{Kevin Kuang},
\href{https://github.com/psolymos/pbapply}{Peter Solymos} and
\href{https://github.com/EricArcher/}{Eric Archer}.

functions on Windows.

functions on Windows.
}
\keyword{internal}
