% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rad.R
\name{read_rad}
\alias{read_rad}
\title{Read radiator file ending \code{.gds, .rad, .gds.rad, .arrow.parquet}.}
\usage{
read_rad(
  data,
  columns = NULL,
  allow.dup = FALSE,
  check = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{data}{A file in the working directory ending with .arrow.parquet or .gds,
and produced by radiator, assigner or grur.}

\item{columns}{(optional) For arrow.parquet file. Column names to read. The default is to read all all columns.
Default: \code{columns = NULL}.}

\item{allow.dup}{(optional, logical) To allow the opening of a GDS file with
read-only mode when it has been opened in the same R session.
Default: \code{allow.dup = FALSE}.}

\item{check}{(optional, logical) Verify that GDS number of samples and markers
match.
Default: \code{check = TRUE}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty
during execution.
Default: \code{verbose = FALSE}.}
}
\value{
A radiator tidy data frame or
GDS object (with read/write permissions) in the global environment.
}
\description{
Fast read of \code{.gds, .rad, .gds.rad, .arrow.parquet} files.
The function uses \code{\link[arrow]{read_parquet}} or
CoreArray Genomic Data Structure (\href{https://github.com/zhengxwen/gdsfmt}{GDS})
file system.


Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
}
\details{
For GDS file system, \strong{read_rad} will open the GDS connection file
set the filters (variants and samples) based on the info found in the file.
}
\examples{
\dontrun{
require(SeqArray)
shark <- radiator::read_rad(data = "data.shark.gds")
turtle <- radiator::read_rad(data = "data.turtle.arrow.parquet")
}
}
\seealso{
\href{https://github.com/apache/arrow/}{arrow}
\href{https://github.com/zhengxwen/gdsfmt}{GDS}
\code{\link{read_rad}}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
