% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rad.R
\name{read_rad}
\alias{read_rad}
\title{Read tidy genomic data file ending .rad}
\usage{
read_rad(data, columns = NULL, from = 1, to = NULL,
  as.data.table = FALSE, old_format = FALSE)
}
\arguments{
\item{data}{A file in the working directory ending with .rad,
a tidy genomic data produced by radiator, assigner or grur.}

\item{columns}{Column names to read. The default is to read all all columns.}

\item{from}{Read data starting from this row number.}

\item{to}{Read data up until this row number. The default is to read to the last row of the stored dataset.}

\item{as.data.table}{If TRUE, the result will be returned as a \code{data.table} object. Any keys set on
dataset \code{x} before writing will be retained. This allows for storage of sorted datasets.}

\item{old_format}{use TRUE to read fst files generated with a fst package version lower than v0.8.0}
}
\value{
A tidy data frame in the global environment.
}
\description{
Fast read .rad file.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
The function uses \code{\link[fst]{read_fst}}.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
