% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{read_strata}
\alias{read_strata}
\title{read strata}
\usage{
read_strata(strata, pop.id = FALSE, pop.levels = NULL,
  pop.labels = NULL, pop.select = NULL, blacklist.id = NULL,
  keep.two = TRUE, path.folder = NULL, verbose = FALSE)
}
\arguments{
\item{strata}{(path or object) The strata file or object.
Additional documentation is available in \code{\link{read_strata}}.
Use that function to whitelist/blacklist populations/individuals.
Option to set \code{pop.levels/pop.labels} is also available.}

\item{pop.id}{(logical) When \code{pop.id = TRUE}, the strata returns
the stratification colname \code{POP_ID}.
Default: \code{pop.id = FALSE}, returns \code{STRATA}.}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
White spaces in population names are replaced by underscore.
Default: \code{pop.levels = NULL}.}

\item{pop.labels}{(optional, string) Use this argument to rename/relabel
your pop or combine your pop. e.g. To combine \code{"QUE"} and \code{"ONT"}
into a new pop called \code{"NEW"}:
(1) First, define the levels for your pop with \code{pop.levels} argument:
\code{pop.levels = c("QUE", "ONT", "ALB")}.
(2) then, use \code{pop.labels} argument:
\code{pop.labels = c("NEW", "NEW", "ALB")}.
To rename \code{"QUE"} to \code{"TAS"}:
\code{pop.labels = c("TAS", "ONT", "ALB")}.
Default: \code{pop.labels = NULL}. If you find this too complicated,
there is also the \code{strata} argument that can do the same thing,
see below.
White spaces in population names are replaced by underscore.}

\item{pop.select}{(optional, string) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops).
Default: \code{pop.select = NULL}}

\item{blacklist.id}{(optional, path or object) A blacklist file in the working directory
or object in the global environment. The data frame
as 1 column (named \code{INDIVIDUALS}) and is filled with the individual IDs
The ids are cleaned with \code{\link{clean_ind_names}} for separators,
only \code{-} are tolerated. Duplicates are removed automatically.
Default: \code{blacklist.id = NULL}.}

\item{keep.two}{(optional, logical) The output is limited to 2 columns:
\code{INDIVIDUALS, STRATA}.
Default: \code{keep.two = TRUE}.}

\item{path.folder}{(optional, path)
If \code{!is.null(blacklist.id) || !is.null(pop.select)}, the modified strata
is written by default in the working directory unless specified otherwise.
Default: \code{path.folder = getwd()}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\value{
\strong{A list} with several components:
\enumerate{
\item $strata
\item $pop.levels
\item $pop.labels
\item $pop.select
\item $blacklist.id
}
}
\description{
Read a strata object or file. The strata file contains thes
individual's metadata, the stratification: e.g. the population id and/or
the sampling sites (see details).
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
The strata file used in radiator is a tab delimited file with
a minimum of 2 columns headers (3 for DArT data users):
\code{INDIVIDUALS} and \code{STRATA}.
If a \code{strata} file is specified with all file formats that don't
require it, the strata argument will have precedence on the population
groupings used internally in those file formats. For file formats without
population/strata groupings (e.g. vcf, haplotype files) if no strata file is
provided, 1 pop/strata grouping will automatically be created.
For vcf and haplotypes file, the strata can also be used as a whitelist of id.
Samples not in the strata file will be discarded from the data set.
The \code{STRATA} column can be any hierarchical grouping.
To create a strata file see \code{\link[radiator]{individuals2strata}}.
If you have already run
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data,
the strata file is similar to a stacks \emph{population map file},
make sure you
have the required column names (\code{INDIVIDUALS} and \code{STRATA}).
The strata column is cleaned of a white spaces that interfere with some
packages or codes: space is changed to an underscore \code{_}.

For DArT data see \code{\link{read_dart}}

\href{https://www.dropbox.com/s/g0vsek0dmtpxntt/example.strata.tsv?dl=0}{example.strata.tsv}.

\href{https://www.dropbox.com/s/utq2h6o00v55kep/example.dart.strata.tsv?dl=0}{example.dart.strata.tsv}.
}
\section{VCF}{

VCF file users, not sure about the sample id inside your file ?
See the example in \code{\link{extract_individuals_vcf}}
}

\section{DArT}{

DArT file users, not sure about the sample id inside your file ?
See the example in \code{\link{extract_dart_target_id}}
}

\examples{
\dontrun{
strata.info <- radiator::read_strata(strata)

# the return object is a list with 5 objects:
names(strata.info)

# to get the strata
new.strata <- strata.info$strata

# if naything is changed from the original strata, a new strata file is
# generated automatically:

new.strata <- radiator::read_strata(
    strata = strata,
    blacklist.id = "blacklisted.ids.tsv"
    )

}
}
\seealso{
\code{\link{summary_strata}},
\code{\link{individuals2strata}}, \code{\link{change_pop_names}},
\code{\link{join_strata}}, \code{\link{generate_strata}}
}
