% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read VCF files for radiator and also write a SeqArray GDS file.}
\usage{
read_vcf(
  data,
  strata = NULL,
  filename = NULL,
  vcf.stats = TRUE,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{(path, character) The VCF markers are biallelic SNPs or haplotypes.
To make the VCF population-ready, you have to use \code{strata} argument.

\itemize{
\item \strong{GATK, platypus and ipyrad VCFs:}
Some VCFs have an \code{ID} column filled with \code{.},
the LOCUS information is all contained along the linkage group in the
\code{CHROM} column. Consequently, the short read locus information is unknown.
To make it work with
\href{https://github.com/thierrygosselin/radiator}{radiator},
the \code{ID} column is filled with the \code{POS} column info.
\item \strong{stacks VCFs:} with \emph{de novo} approaches, the CHROM column is
filled with "1", the LOCUS column correspond to the CHROM section in stacks VCF and
the COL column is POS -1. With a reference genome, the ID column in stacks VCF is
separated into "LOCUS", "COL", "STRANDS".
\item \strong{DArT VCFs:} CHROM with \code{.} are replaced by \code{denovo}.
\code{POS} with \code{NA} are replaced by {50}.
\code{COL}: the position of the SNP on the read is extracted from the
\code{LOCUS} as any other DArT data.
\code{LOCUS}: are the first group of digits the rest of the info is discarded
and \code{LOCUS} is then joined with \code{POS} by a \code{_}.
\code{POS} is replaced by \code{COL}, (col info is duplicated).
This is necessary to make the lines in the VCF unique.
}}

\item{strata}{(optional)
The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}. Documented in \code{\link{read_strata}}.
DArT data: a third column \code{TARGET_ID} is required.
Documented on \code{\link{read_dart}}. Also use the strata read function to
blacklist individuals.
Default: \code{strata = NULL}.}

\item{filename}{(optional) The file name of the Genomic Data Structure (GDS) file.
radiator will append \code{.gds.rad} to the filename.
If the filename chosen exists in the working directory,
the default \code{radiator_datetime.gds} is chosen.
Default: \code{filename = NULL}.}

\item{vcf.stats}{(optional, logical) Generates individuals and markers
statistics helpful for filtering.
These are very fast to generate and because computational
cost is minimal, even for huge VCFs, the default is \code{vcf.stats = TRUE}.
Individual's missing genotype proportion, averaged heterozygosity,
total coverage, mean genotype coverage and marker's metadata
along count for ref and alt alleles and mean
coverage is generated and written in the working directory.
Default: \code{vcf.stats = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) To pass further arguments for fine-tuning the function.}
}
\value{
The function returns a GDS object.
}
\description{
The function reads VCF files for radiator and
generate a connection SeqArray \href{https://github.com/zhengxwen/SeqArray}{SeqArray}
GDS object/file of class \code{SeqVarGDSClass} (Zheng et al. 2017)
The Genomic Data Structure (GDS) file format is detailed in
\href{https://github.com/zhengxwen/gdsfmt}{gdsfmt}.

The function as an advance mode that allows various filtering arguments to
be used. Handy to prune the dataset based on various QCs but also to
remove artifacts, duplicated information and thus lower the overall noise in
the VCF.

Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
A vcf file of 35 GB with ~4 millions SNPs take about ~7 min with 8 CPU.
A vcf file of 21 GB with ~2 millions SNPs take about ~5 min with 7 CPU.

After the file is generated, you can close your computer and
come back to it a month later and it's now a matter of sec to open a connection.


\strong{markers heterozygosity and inbreeding coefficient}:
The heterozygosity statistics (het obs and Fis) presented in this function
are calculated globally across strata, without the possibility to filter out
markers. We think this filtering for these statistics are best
addressed in \code{\link{filter_het}}, \code{\link{filter_fis}} and
\code{\link{filter_hwe}}, after outlier individuals and
markers (for other stats) are blacklisted. Why ? The reason is that an excess of
heterozygotes and negative Fis values are not a bad thing \emph{per se}.
Genomic regions under balancing selection may contain such markers and
statistics.
}
\section{VCF file format}{


\strong{PLINK:} radiator fills the \code{LOCUS} column of PLINK VCFs with
a unique integer based on the \code{CHROM} column
(\code{as.integer(factor(x = CHROM))}).
The \code{COL} column is filled with 1L for lack of bettern info on this.
Not what you need ? Open an issue on GitHub for a request.

\strong{ipyrad:} the pattern \code{locus_} in the \code{CHROM} column
is removed and used. The \code{COL} column is filled with the same value as
\code{POS}.

\strong{GATK:} Some VCF have an \code{ID} column filled with \code{.},
the LOCUS information is all contained along the linkage group in the
\code{CHROM} column. To make it work with
\href{https://github.com/thierrygosselin/radiator}{radiator},
the \code{ID} column is filled with the \code{POS} column info.

\strong{platypus:} Some VCF files don't have an ID filed with values,
here the same thing is done as GATK VCF files above.

\strong{freebayes:} Some VCF files don't have an ID filed with values,
here the same thing is done as GATK VCF files above.

\strong{stacks:} with \emph{de novo} approaches, the CHROM column is
filled with "1", the LOCUS column correspond to the CHROM section in stacks VCF and
the COL column is POS -1. With a reference genome, the ID column in stacks VCF is
separated into "LOCUS", "COL", "STRANDS".

\emph{stacks problem: } current version as some intrinsic problem with
missing allele depth info, during the tidying process a message will
highlight the number of genotypes impacted by the problem. When possible, the
problem is corrected by adding the read depth info into the allele depth field.
}

\section{Advance mode}{


\emph{dots-dots-dots ...} allows to pass several arguments for fine-tuning the function:
\enumerate{
\item \code{whitelist.markers}: detailed in \code{\link[radiator]{filter_whitelist}}.
\item \code{blacklist.id}: detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{pop.select}: detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{pop.levels}: detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{pop.labels}: detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{filter.strands}: (optional, character) Filter duplicate SNPs
found on different strands (+/-), options are:
\code{"keep.both", "best.stats", "blacklist"}. \code{"keep.both"}: does nothing
and duplicated markers are kept (not recommended, but here for testing purposes),
\code{"best.stats"}: will keep only one, based on the best statistics
(MAC and missingness). \code{"blacklist"}: discard all duplicated markers.
Default (\code{filter.strands = "blacklist"}).
\item \code{filter.common.markers}: (logical) Default: \code{filter.common.markers = TRUE}.
Documented in \code{\link{filter_common_markers}}.
\item \code{filter.mac}: (integer) Default: \code{filter.mac = NULL}.
To blacklist markers below a specific Minor Allele Count (calculated overall/global).
Documented in \code{\link{filter_mac}}.

\item \code{filter.coverage}: (optional, string)
Default: \code{filter.coverage = NULL}. To blacklist markers based on mean coverage.
Documented in \code{\link{filter_coverage}}.

\item \code{filter.genotyping}: (integer) To blacklist markers with too
many missing data. e.g. \code{filter.genotyping = 0.1}, will only keep
markers with missing rate <= to 10 percent.
Documented in \code{\link{filter_genotyping}}.
\item \code{filter.snp.position.read}: 3 options available, \code{"outliers", "q75", "iqr"}.
This argument will blacklist markers based on it's position on the read.
\code{filter.snp.position.read = "outliers"}, will remove markers based
on outlier statistics of the position on the reads. e.g. if majority of SNPs
are found between 10 and 90 pb, and very few above and below, those markers are
discarded. Use this function argument with dataset with problematic assembly,
or \emph{de novo} assembly with undocumented or poorly selected
mismatch threshold.
\item \code{filter.short.ld}: Described in \code{\link[radiator]{filter_ld}}
\item \code{filter.long.ld}: Described in  \code{\link[radiator]{filter_ld}}
\item \code{long.ld.missing}: Described in \code{\link[radiator]{filter_ld}}
\item \code{ld.method}: (optional, character) The values available are
\code{"composite"}, for LD composite measure, \code{"r"} for R coefficient
(by EM algorithm assuming HWE, it could be negative), \code{"r2"} for r^2,
\code{"dprime"} for D',
\code{"corr"} for correlation coefficient. The method corr and composite are
equivalent when SNPs are coded based on the presence of the alternate allele
(\code{0, 1, 2}).
Default: \code{ld.method = "r2"}.
\item \code{filter.individuals.missing}: (double) Use this argument to
blacklist individuals with too many missing data.
e.g. \code{filter.individuals.missing = 0.7}, will remove individuals with >
0.7 or 70% missing genotypes. This can help discover more polymorphic markers
with some dataset.
\item \code{markers.info}: (character) With default: \code{markers.info = NULL},
all the variable in the vcf INFO field are imported.
To import only DP (the SNP total read depth) and AF (the SNP allele frequency),
use \code{markers.info = c("DP", "AF")}.
Using, \code{markers.info = character(0)} will not import INFO variables.
\item \code{path.folder}: to write ouput in a specific path
(used internally in radiator). Default: \code{path.folder = getwd()}.
If the supplied directory doesn't exist, it's created.
\item \code{random.seed}: (integer, optional) For reproducibility, set an integer
that will be used inside codes that uses randomness. With default,
a random number is generated, printed and written in the directory.
Default: \code{random.seed = NULL}.
\item \code{subsample.markers.stats}: By default, when no filters are
requested and that the number of markers is > 200K,
0.2 of markers are randomly selected to generate the
statistics (individuals and markers). This is an all-around and
reliable number.
In doubt, overwrite this value by using 1 (all markers selected) and
expect a small computational cost.
}
}

\examples{
\dontrun{
#require(SeqVarTools)
# with built-in defaults:
 prep.data <- radiator::read_vcf(data = "populations.snps.vcf")

# Using more arguments and filters (recommended):
prep.data <- radiator::read_vcf(
    data = "populations.snps.vcf",
    strata = "strata_salamander.tsv",
    path.folder = "salamander",
    filter.individuals.missing = "outliers",
    filter.common.markers = TRUE,
    filter.strands = "blacklist",
    filter.mac = 4,
    filter.genotyping = 0.3,
    filter.snp.position.read = "outliers",
    filter.short.ld = "mac",
    filter.long.ld = NULL,
    verbose = TRUE
    )
}
}
\references{
Zheng X, Gogarten S, Lawrence M, Stilp A, Conomos M, Weir BS,
Laurie C, Levine D (2017). SeqArray -- A storage-efficient high-performance
data format for WGS variant calls.
Bioinformatics.

Danecek P, Auton A, Abecasis G et al. (2011)
The variant call format and VCFtools.
Bioinformatics, 27, 2156-2158.
}
\seealso{
\code{\link{filter_ld}}
\code{\link[radiator]{tidy_genomic_data}}
\code{\link[radiator]{tidy_vcf}}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
