% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read VCF files for radiator and also write a SeqArray GDS file.}
\usage{
read_vcf(data, strata = NULL, filename = NULL, vcf.stats = TRUE,
  parallel.core = parallel::detectCores() - 1, verbose = TRUE, ...)
}
\arguments{
\item{data}{(path, character) The VCF markers are biallelic SNPs or haplotypes.
To make the VCF population-ready, you have to use \code{strata} argument.

\itemize{
\item \strong{GATK, platypus and ipyrad VCFs:}
Some VCFs have an \code{ID} column filled with \code{.},
the LOCUS information is all contained along the linkage group in the
\code{CHROM} column. Consequently, the short read locus information is unknown.
To make it work with
\href{https://github.com/thierrygosselin/radiator}{radiator},
the \code{ID} column is filled with the \code{POS} column info.
\item \strong{stacks VCFs:} with \emph{de novo} approaches, the CHROM column is
filled with "1", the LOCUS column correspond to the CHROM section in stacks VCF and
the COL column is POS -1. With a reference genome, the ID column in stacks VCF is
separated into "LOCUS", "COL", "STRANDS".
}}

\item{strata}{(optional)
The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}. Documented in \code{\link{read_strata}}.
Default: \code{strata = NULL}.}

\item{filename}{(optional) The file name of the Genomic Data Structure (GDS) file.
radiator will append \code{.gds.rad} to the filename.
If the filename chosen exists in the working directory,
the default \code{radiator_datetime.gds} is chosen.
Default: \code{filename = NULL}.}

\item{vcf.stats}{(optional, logical) Generates individuals and markers
statistics helpful for filtering.
These are very fast to generate and because computational
cost is minimal, even for huge VCFs, the default is \code{vcf.stats = TRUE}.
Individual's missing genotype proportion, averaged heterozygosity,
total coverage, mean genotype coverage and marker's metadata
along count for ref and alt alleles and mean
coverage is generated and written in the working directory.
Default: \code{vcf.stats = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) To pass further arguments for fine-tuning the function.}
}
\value{
The function returns a list with:
\enumerate{
\item \code{vcf.connection}: the name of the GDS file.
To close the connection with the GDS file: use \code{SeqArray::seqClose(OBJECT_NAME$vcf.connection)}
\item \code{individuals}: a tibble with the names of samples in the original
VCF \code{INDIVIDUALS_VCF} and corrected for use in radiator \code{INDIVIDUALS}
\item \code{biallelic}: is the data biallelic or not.
\item \code{markers.meta}: a tibble with the markers metadata, including:
\code{VARIANT_ID, CHROM, LOCUS, POS, COL, MARKERS, REF, ALT} when available.
\item \code{n.ind}: the number of individuals
\item \code{n.markers}: the number of markers
\item \code{n.chromosome}: the numer of chromosome
\item \code{n.locus}: the number of locus
\item \code{filename}: the name of the file used.
}
}
\description{
The function reads VCF files for radiator and
generate a connection SeqArray \href{http://zhengxwen.github.io/SeqArray/}{SeqArray}
GDS object/file of class \code{SeqVarGDSClass} (Zheng et al. 2017)
The Genomic Data Structure (GDS) file format is detailed in
\href{http://zhengxwen.github.io/gdsfmt}{gdsfmt}.

The function as an advance mode that allows various filtering arguments to
be used. Handy to prune the dataset based on various QCs but also to
remove artifacts, duplicated information and thus lower the overall noise in
the VCF.

Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
A vcf file of 35 GB with ~4 millions SNPs take about ~7 min with 8 CPU.
A vcf file of 21 GB with ~2 millions SNPs take about ~5 min with 7 CPU.

After the file is generated, you can close your computer and
come back to it a month later and it's now a matter of sec to open a connection.


\strong{markers heterozygosity and inbreeding coefficient}:
The heterozygosity statistics (het obs and Fis) presented in this function
are calculated globally across strata, without the possibility to filter out
markers. We think this filtering for these statistics are best
addressed in \code{\link{filter_het}}, \code{\link{filter_fis}} and
\code{\link{filter_hwe}}, after outlier individuals and
markers (for other stats) are blacklisted. Why ? The reason is that an excess of
heterozygotes and negative Fis values are not a bad thing \emph{per se}.
Genomic regions under balancing selection may contain such markers and
statistics.

\strong{Advance mode, using \emph{dots-dots-dots ...}}
\enumerate{
\item \code{whitelist.markers}: detailed in \code{\link[radiator]{filter_whitelist}}.
\item \code{blacklist.id}: detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{pop.select}: detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{pop.levels}: detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{pop.labels}: detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{filter.strands}: (optional, character) Filter duplicate SNPs
found on different strands (+/-), options are:
\code{"keep.both", "best.stats", "blacklist"}. \code{"keep.both"}: does nothing
and duplicated markers are kept (not recommended, but here for testing purposes),
\code{"best.stats"}: will keep only one, based on the best statistics
(MAC and missingness). \code{"blacklist"}: discard all duplicated markers.
Default (\code{filter.strands = "blacklist"}).
\item \code{filter.common.markers}: (logical) Default: \code{filter.common.markers = TRUE}.
Documented in \code{\link{filter_common_markers}}.
\item \code{filter.mac}: (integer) Default: code{filter.mac = NULL}.
To blacklist markers below a specific Minor Allele Count (calculated overall/global).
Documented in \code{\link{filter_mac}}.

\item \code{filter.coverage}: (optional, string)
Default: code{filter.coverage = NULL}. To blacklist markers based on mean coverage.
Documented in \code{\link{filter_coverage}}.

\item \code{filter.genotyping}: (integer) To blacklist markers with too
many missing data. e.g. \code{filter.genotyping = 10}, will only keep
markers with missing rate <= to 10 percent.
\item \code{filter.snp.position.read}: 3 options available, \code{"outliers", "q75", "iqr"}.
This argument will blacklist markers based on it's position on the read.
\code{filter.snp.position.read = "outliers"}, will remove markers based
on outlier statistics of the position on the reads. e.g. if majority of SNPs
are found between 10 and 90 pb, and very few above and below, those markers are
discarded. Use this function argument with dataset with problematic assembly,
or \emph{de novo} assembly with undocumented or poorly selected
mismatch threshold.
\item \code{filter.short.ld}: Described in \code{\link[radiator]{filter_ld}}
\item \code{filter.long.ld}: Described in  \code{\link[radiator]{filter_ld}}
\item \code{long.ld.missing}: Described in \code{\link[radiator]{filter_ld}}
\item \code{ld.method}: (optional, character) The values available are
\code{"composite"}, for LD composite measure, \code{"r"} for R coefficient
(by EM algorithm assuming HWE, it could be negative), \code{"r2"} for r^2,
\code{"dprime"} for D',
\code{"corr"} for correlation coefficient. The method corr and composite are
equivalent when SNPs are coded based on the presence of the alternate allele
(\code{0, 1, 2}).
Default: \code{ld.method = "r2"}.
\item \code{filter.individuals.missing}: (double) Use this argument to
blacklist individuals with too many missing data.
e.g. \code{filter.individuals.missing = 0.7}, will remove individuals with >
0.7 or 70% missing genotypes. This can help discover more polymorphic markers
with some dataset.
\item \code{markers.info}: (character) With default: \code{markers.info = NULL},
all the variable in the vcf INFO field are imported.
To import only DP (the SNP total read depth) and AF (the SNP allele frequency),
use \code{markers.info = c("DP", "AF")}.
Using, \code{markers.info = character(0)} will not import INFO variables.
\item \code{path.folder}: to write ouput in a specific path
(used internally in radiator). Default: \code{path.folder = getwd()}.
If the supplied directory doesn't exist, it's created.
\item \code{random.seed}: (integer, optional) For reproducibility, set an integer
that will be used inside codes that uses randomness. With default,
a random number is generated, printed and written in the directory.
Default: \code{random.seed = NULL}.
\item \code{subsample.markers.stats}: By default, when no filters are
requested and that the number of markers is > 200K,
0.2 of markers are randomly selected to generate the
statistics (individuals and markers). This is an all-around and
reliable number.
In doubt, overwrite this value by using 1 (all markers selected) and
expect a small computational cost.
}
}
\examples{
\dontrun{
#require(SeqVarTools)
# with built-in defaults:
 prep.data <- radiator::read_vcf(data = "populations.snps.vcf")

# Using more arguments and filters (recommended):
prep.data <- radiator::read_vcf(
    data = "populations.snps.vcf",
    strata = "strata_salamander.tsv",
    filter.individuals.missing = "outliers",
    filter.common.markers = TRUE,
    filter.strands = "blacklist",
    filter.mac = 4,
    filter.genotyping = 0.3,
    filter.snp.position.read = "outliers",
    filter.short.ld = "mac",
    filter.long.ld = NULL,
    path.folder = "salamander",
    verbose = TRUE)
}
}
\references{
Zheng X, Gogarten S, Lawrence M, Stilp A, Conomos M, Weir BS,
Laurie C, Levine D (2017). SeqArray -- A storage-efficient high-performance
data format for WGS variant calls.
Bioinformatics.

Danecek P, Auton A, Abecasis G et al. (2011)
The variant call format and VCFtools.
Bioinformatics, 27, 2156-2158.
}
\seealso{
\code{\link{filter_ld}} and \code{\link[radiator]{tidy_genomic_data}}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
