% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duplicate_genomes.R, R/snp_ld.R
\name{remove_duplicates}
\alias{remove_duplicates}
\alias{ld_pruning}
\title{Read tidy genomic data file ending .rad}
\usage{
remove_duplicates(data = "individuals.pairwise.dist.tsv",
  stats = "genotyped.statistics.tsv", dup.threshold = 0.25,
  diff.pop.remove = TRUE)

ld_pruning(ld.tibble = NULL, stats = NULL, ld.threshold = 0.8)
}
\arguments{
\item{data}{(path) The individual's pairwise data.
Default: \code{data = "individuals.pairwise.dist.tsv"}.}

\item{stats}{(path) The genotype statistics
Default: \code{stats = "genotyped.statistics.tsv"}.}

\item{dup.threshold}{(double) The threshold to filter out duplicates
Default: \code{dup.threshold = 0.25}.}

\item{diff.pop.remove}{Remove all individuals in pairs from different pop.
Both samples are potentially problems. With defautl, the function will not keep
one sample in the duplicate pair.
Default: \code{diff.pop.remove = TRUE}.}

\item{ld.tibble}{(path) The markers LD pairwise data.
Default: \code{ld.tibble = NULL}.}

\item{ld.threshold}{(double) The threshold to prune SNPs in LD.
Default: \code{ld.threshold = 0.8}.}

\item{stats}{(path) The markers missingness info statistics.
Default: \code{stats = NULL}.}
}
\value{
A list with blacklisted duplicates. Write the blacklist in the working
directory.

A list with blacklisted SNPs.  Write the blacklist in the working
directory.
}
\description{
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
To remove duplicate individuals based on threshold established from the
visualization figures.

Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
Prune dataset based on LD. Use missingness to keep 1 SNP.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}

Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
