% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{separate_gt}
\alias{separate_gt}
\title{separate_gt}
\usage{
separate_gt(x, sep = "/", gt = "GT_VCF_NUC", gather = TRUE,
  exclude = c("LOCUS", "INDIVIDUALS", "POP_ID"), cpu.rounds = 10,
  parallel.core = parallel::detectCores() - 1)
}
\description{
Separate genotype field
}
\keyword{internal}
