% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_ld.R
\name{snp_ld}
\alias{snp_ld}
\title{GBS/RADseq short and long distance linkage disequilibrium pruning}
\usage{
snp_ld(data, snp.ld = "maf", maf.data = NULL, ld.threshold = NULL,
  parallel.core = parallel::detectCores() - 1, filename = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{data}{A tidy data set with ID (LOCUS) and POS (SNP) information.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.
Usually the LOCUS and POS info is taken from a VCF file.

The function also accept \code{SeqVarGDSClass} object generated
with the SeqArray package
or \pkg{radiator} \code{\link{write_seqarray}}.

\strong{Biallelic genotypes required...}}

\item{snp.ld}{(character) 5 options:
\enumerate{
\item \code{snp.ld = "random"} for a random selection of 1 SNP on the read,
\item \code{snp.ld = "first"} for the first one on the read...,
\item \code{snp.ld = "last"} for the last SNP on the read and
\item \code{snp.ld = "middle"} for locus with > 2 SNPs/read the option to select at random
one SNP between the first and the last SNP on the read. If the locus as <= 2
SNPs on the read, the first one is selected. Note that for that last option,
the numbers are reported.
\item \code{snp.ld = "maf"} will select the SNP on the locus with the maximum global
Minor Allele Frequency (MAF).
}
Default: \code{snp.ld = "maf"}.}

\item{maf.data}{(path) When \code{snp.ld = "maf"} is selected,
to speed up calculations, you can provide
the maf information generated by the function \code{\link{filter_maf}} or
let this function calculates it from the data (default).
Designated alternate allele embedded in the data are not taken for granted
and this information is re-computed from the data to get actual
minor allele frequency. In this context, to speed up calculations,
the MAF is calculated only for locus with multiple SNPs .
Default: \code{maf.data = NULL}.}

\item{ld.threshold}{(optional, double) The threshold to prune SNP based on
Long Distance Linkage Disequilibrium.
Default: \code{ld.threshold = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel::detectCores() - 1}.}

\item{filename}{(optional, character) File name prefix for file written in
the working directory.
Default: \code{filename = NULL}.}

\item{verbose}{(optional, logical) The function will write more details.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function (see details).}
}
\value{
A list in the global environment, with these objects:
\enumerate{
\item $ld.summary: tibble with LD statistics used for the boxplot
\item $ld.boxplot: box plot of LD values
\item $whitelist.snp.ld: whitelist of markers kept after filtering for LD.
The argument \code{ld.threshold} must be used to generate the whitelist.
\item $blacklist.snp.ld: blacklist of markers prunned during the filtering
for LD.
The argument \code{ld.threshold} must be used to generate the blacklist.
\item $data: The filtered tidy dataset.
\item $gds: the path to the GDS file.
}
}
\description{
SNP short and long distance linkage disequilibrium pruning.

What sets appart radiator LD pruning is the RADseq data tailored arguments.
First, minimize short linkage disequilibrium (LD) by
choosing between 5 values for \code{snp.ld} argument (see below).
Second, you can optionally reduce long distance LD by adjusting the argument
\code{ld.threshold}. Values between 0.7 and 0.9 are good starting point.
You can also choose to refilter the data using the outlier statistics
generated by the function.
To prune markers in long distance LD, instead of choosing randomly one SNP,
radiator can incorporate missing data statistics
to determine the best SNP to keep (see details).

Long distance LD pruning is usually advised to avoid capturing the variance LD
in PCA analysis.


This function is used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
The function requires \href{https://github.com/zhengxwen/SNPRelate}{SNPRelate}
(see example below on how to install).


\strong{Advance mode, using \emph{dots-dots-dots}}
\itemize{
\item \code{long.ld.missing} (logical) With \code{long.ld.missing = TRUE}.
The function first generates long distance LD values between markers along
the same chromosome or scaffold with \emph{SNPRelate::snpgdsLDMat}.
Based on the LD threshold (\code{ld.threshold}) SNPs in LD will be pruned
based on missingness.
e.g. if 4 SNPs are in LD, the 1 SNP selected in
the end is base on genotyping rate/missingness. If this statistic is equal
between the SNPs in LD, 1 SNP is chosen randomly.

Using missigness add extra computational time. To speed the analysis when
missingness between markers is not an issue, use \code{long.ld.missing = FALSE}.
The function will use \emph{SNPRelate::snpgdsLDpruning}
to prune the dataset. SNPs in LD are selected randomly.
Default: \code{long.ld.missing = FALSE}.
\item \code{keep.gds} (logical) Default: \code{keep.gds = FALSE}, when the input data is a
tidy data frame of genotypes, the GDS file
generated for the long distance LD is removed after completion. With GDS input,
the GDS file is always updated and kept in the working directory.
\item \code{ld.figures}: (logical) Generate long distance LD statistics and
figures.
Default: \code{ld.figures = TRUE}
\item \code{path.folder}: to write ouput in a specific path
(used internally in radiator). Default: \code{path.folder = getwd()}.
If the supplied directory doesn't exist, it's created.
}
}
\examples{
\dontrun{
#require(SNPRelate)
#To install SNPRelate:
#source("https://bioconductor.org/biocLite.R")
#biocLite("SNPRelate")
#library(radiator)
data <- radiator::tidy_vcf(
data = "my.vcf", strata = "my.strata.tsv",
vcf.metadata = FALSE, vcf.stats = TRUE, verbose = TRUE)

# short distance LD, no long distance LD:
check.short.ld <- radiator::snp_ld(
data = data$tidy.data,
snp.ld = "maf")

# short distance LD and long distance LD:
pruned.ld <- radiator::snp_ld(
data = data$tidy.data,
snp.ld = "maf", ld.threshold = 0.8)

# short distance LD and long distance LD, incorporating missing data:
pruned.ld <- radiator::snp_ld(
data = data, # a GDS object generated by radiator
snp.ld = "maf", ld.threshold = 0.8, long.ld.missing = TRUE)
}
}
\references{
Zheng X, Levine D, Shen J, Gogarten SM, Laurie C, Weir BS.
(2012) A high-performance computing toolset for relatedness and principal component
analysis of SNP data. Bioinformatics. 28: 3326-3328.
doi:10.1093/bioinformatics/bts606
}
\seealso{
\href{https://github.com/zhengxwen/SNPRelate}{SNPRelate}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
