% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_ld.R
\name{snp_ld}
\alias{snp_ld}
\title{GBS/RADseq short distance linkage disequilibrium pruning}
\usage{
snp_ld(data, snp.ld = "maf", maf.data = NULL,
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{data}{A tidy data set with ID (LOCUS) and POS (SNP) information.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.
Usually the LOCUS and POS info is taken from a VCF file.}

\item{snp.ld}{(character) 5 options:
\enumerate{
\item \code{snp.ld = "random"} for a random selection of 1 SNP on the read,
\item \code{snp.ld = "first"} for the first one on the read...,
\item \code{snp.ld = "last"} for the last SNP on the read and
\item \code{snp.ld = "middle"} for locus with > 2 SNPs/read the option to select at random
one SNP between the first and the last SNP on the read. If the locus as <= 2
SNPs on the read, the first one is selected. Note that for that last option,
the numbers are reported.
\item \code{snp.ld = "maf"} will select the SNP on the locus with the maximum global
Minor Allele Frequency (MAF).
}
Default: \code{snp.ld = "maf"}.}

\item{maf.data}{(path) When \code{snp.ld = "maf"} is selected,
to speed up calculations, you can provide
the maf information generated by the function \code{\link{filter_maf}} or
let this function calculates it from the data (default).
Designated alternate allele embedded in the data are not taken for granted
and this information is re-computed from the data to get actual
minor allele frequency. MAF in this context is calculated only for locus with
multiple SNPs.
Default: \code{maf.data = NULL}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during vcf import.
Default: \code{parallel::detectCores() - 1}.}
}
\description{
SNP short distance linkage disequilibrium pruning.
With anonymous markers from
RADseq/GBS de novo discovery, you can minimize linkage disequilibrium (LD) by
choosing among these 5 options (see argument below).

For long distance linkage disequilibrium pruning, see details below.
This function is used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
