% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_markers.R
\name{summarize_sex}
\alias{summarize_sex}
\title{summarize_sex}
\usage{
summarize_sex(data, silicodata, data.source, coverage.thresholds = NULL,
  tau = 0.3)
}
\description{
Function to summarize the presence-absence, heterozygosity and read depth per sex.
}
\keyword{internal}
