% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiator_tidy.R
\name{tidy2wide}
\alias{tidy2wide}
\title{tidy2wide}
\usage{
tidy2wide(x = NULL, gds = NULL, individuals = NULL, markers = NULL,
  tidy = TRUE, wide = TRUE, wide.markers = TRUE)
}
\description{
tidy2wide
}
\keyword{internal}
