% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{tidy_dart}
\alias{tidy_dart}
\title{Tidy \href{http://www.diversityarrays.com}{DArT} output file.}
\usage{
tidy_dart(data, strata, filename = NULL, verbose = FALSE,
  parallel.core = parallel::detectCores() - 1, ...)
}
\arguments{
\item{data}{DArT output file. Note that most popular formats used by DArT are
recognised (1- and 2- row format, also called binary, and count data.).
If you encounter a problem, sent me your data so that I can update
the function. The function can import \code{.csv} or \code{.tsv} files.}

\item{strata}{A tab delimited file or object with 3 columns.
Columns header is:
\code{TARGET_ID}, \code{INDIVIDUALS} and \code{STRATA}.
Note: the column \code{STRATA} refers to any grouping of individuals.
You need to make sure that
the column \code{TARGET_ID} match the id used by DArT.
The column \code{INDIVIDUALS} and \code{STRATA} will be kept in the tidy data.
Only individuals in the strata file are kept in the tidy, i.e. that the strata
is also used as a whitelist of individuals/strata.}

\item{filename}{(optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in
the working directory. The file extension appended to
the \code{filename} provided is \code{.rad}.
With default: \code{filename = NULL}, the tidy data frame is
in the global environment only (i.e. not written in the working directory...).}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{...}{(optional) To pass further argument for fine-tuning the function.}
}
\value{
A radiator GDS file and tidy dataframe with several columns depending on DArT file:
\enumerate{
\item MARKERS: generated by radiator and correspond to CHROM + LOCUS + POS separated by 2 underscores.
\item CHROM: the chromosome, for de novo: CHROM_1.
\item LOCUS: the locus.
\item POS: the SNP id on the LOCUS.
\item REF: the reference allele.
\item ALT: the alternate allele.
\item INDIVIDUALS: the sample name.
\item POP_ID: populations id of the sample.
\item GT: the genotype in 6 digit format à la genepop.
\item GT_VCF: the genotype in VCF format.
\item GT_VCF_NUC: the genotype in VCF format, but keeping the nucleotide information.
\item GT_BIN: the genotype in binary format similar to PLINK. The number correspond to the number of alternate allele in the genotype.
\item CALL_RATE: call rate output specific of DArT.
\item AVG_COUNT_REF: the coverage for the reference allele, output specific of DArT.
\item AVG_COUNT_SNP: the coverage for the alternate allele, output specific of DArT.
\item REP_AVG: the reproducibility average, output specific of DArT.
\item READ_DEPTH: the number of reads used for the genotype (count data).
\item ALLELE_REF_DEPTH: the number of reads of the reference allele (count data).
\item ALLELE_ALT_DEPTH: the number of reads of the alternate allele (count data).
}


Written in the working directory:
\itemize{
\item The radiator GDS file
\item The DArT metadata information
\item The tidy DArT data
\item The strata file associated with this tidy dataset
\item The allele dictionary is a tibble with columns:
\code{MARKERS, CHROM, LOCUS, POS, REF, ALT}.
}
\strong{Advance mode, using \emph{dots-dots-dots ...}}
\enumerate{
\item \code{whitelist.markers}: detailed in \code{\link[radiator]{filter_whitelist}}.
Defautl: \code{whitelist.markers = NULL}.
\item \code{missing.memory} (option, path)
This argument allows to erase genotypes that have bad statistics.
It's the path to a file \code{.rad} file that contains 3 columns:
\code{MARKERS, INDIVIDUALS, ERASE}. The file is produced by several radiator
functions. For DArT data, \code{\link[radiator]{filter_rad}} generate the file.
Defautl: \code{missing.memory = NULL}.
\item \code{dart.sequence}: (optional, logical) To keep the sequence of markers in
the GDS. Theres no computational cost to keep it in the GDS...
Default: \code{dart.sequence = TRUE}.
\item \code{path.folder}: (optional, path) To write output in a specific folder.
Default: \code{path.folder = NULL}. With defautl, the working directory is used.
}
}
\description{
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users. The function generate a tidy dataset using
\href{http://www.diversityarrays.com}{DArT} file.
}
\examples{
\dontrun{
clownfish.dart.tidy <- radiator::tidy_dart(
data = "clownfish.dart.tsv",
strata = "clownfish.strata.tsv",
verbose = TRUE)

# To get a strip, bare minimal version of the DArT tidy data:
clownfish.dart.tidy <- radiator::tidy_dart(
data = "clownfish.dart.tsv",
strata = "clownfish.strata.tsv",
tidy = "strip")
# Only MARKERS, INDIVIDUALS, POP_ID, GT_BIN is generated in the output
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
