% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dart.R
\name{tidy_dart}
\alias{tidy_dart}
\title{Tidy \href{http://www.diversityarrays.com}{DArT} output file.}
\usage{
tidy_dart(data, strata, whitelist.markers = NULL, filename = NULL,
  verbose = FALSE, parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{data}{DArT output file. Note that most popular formats used by DArT are
recognised (1- and 2- row format, also called binary, and count data.).
If you encounter a problem, sent me your data so that I can update
the function. The function can import \code{.csv} or \code{.tsv} files.}

\item{strata}{A tab delimited file with 3 columns.
Columns header is:
\code{TARGET_ID}, \code{INDIVIDUALS} and \code{STRATA}.
Note: the column \code{STRATA} refers to any grouping of individuals.
You need to make sure that
the column \code{TARGET_ID} match the id used by DArT.
The column \code{INDIVIDUALS} and \code{STRATA} will be kept in the tidy data.
Only individuals in the strata file are kept in the tidy, i.e. that the strata
is also used as a whitelist of individuals/strata.}

\item{whitelist.markers}{(optional) A whitelist containing CHROM (character
or integer) and/or LOCUS (integer) and/or
POS (integer) columns header. To filter by chromosome and/or locus and/or by snp.
The whitelist is an object in your
global environment or a file in the working directory (e.g. "whitelist.txt").
Note that \emph{de novo} CHROM column with 'un' need to be changed to 1.
In the VCF, the column ID is the LOCUS identification (VCF generated from
stacks have the SNP position on the read embedded in the ID,
so the ID = no longer represent the LOCUS). Marker names are cleaned of
separators that interfere with some packages or codes:
\code{/}, \code{:}, \code{-} and \code{.} are changed to an underscore
\code{_}.
Default \code{whitelist.markers = NULL} for no whitelist of markers.}

\item{filename}{(optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in
the working directory. The file extension appended to
the \code{filename} provided is \code{.rad}.
With default: \code{filename = NULL}, the tidy data frame is
in the global environment only (i.e. not written in the working directory...).}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during vcf import.
Default: \code{parallel::detectCores() - 1}.}
}
\value{
A tidy dataframe with these columns:
\enumerate{
\item MARKERS: generated by radiator and correspond to CHROM + LOCUS + POS separated by 2 underscores.
\item CHROM: the chromosome, for de novo: CHROM_1.
\item LOCUS: the locus.
\item POS: the SNP id on the LOCUS.
\item REF: the reference allele.
\item ALT: the alternate allele.
\item INDIVIDUALS: the sample name.
\item POP_ID: populations id of the sample.
\item GT: the genotype in 6 digit format à la genepop.
\item GT_VCF: the genotype in VCF format.
\item GT_VCF_NUC: the genotype in VCF format, but keeping the nucleotide information.
\item GT_BIN: the genotype in binary format similar to PLINK. The number correspond to the number of alternate allele in the genotype.
\item CALL_RATE: call rate output specific of DArT.
\item AVG_COUNT_REF: the coverage for the reference allele, output specific of DArT.
\item AVG_COUNT_SNP: the coverage for the alternate allele, output specific of DArT.
\item REP_AVG: the reproducibility average, output specific of DArT.
}
}
\description{
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users. The function generate a tidy dataset using
\href{http://www.diversityarrays.com}{DArT} file.
}
\examples{
\dontrun{
clownfish.dart.tidy <- radiator::tidy_dart(
data = "clownfish.dart.tsv",
strata = "clownfish.strata.tsv",
verbose = TRUE)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com} and Peter Grewe \email{peter.grewe@csiro.au}
}
