% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_genind.R
\name{tidy_genind}
\alias{tidy_genind}
\title{Tidy a genind object to a tidy dataframe}
\usage{
tidy_genind(data)
}
\arguments{
\item{data}{A genind object in the global environment.}
}
\description{
Tidy genind object to a tidy dataframe.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\references{
Jombart T (2008) adegenet: a R package for the multivariate
analysis of genetic markers. Bioinformatics, 24, 1403-1405.

Jombart T, Ahmed I (2011) adegenet 1.3-1:
new tools for the analysis of genome-wide SNP data.
Bioinformatics, 27, 3070-3071.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
