% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genind.R
\name{tidy_genind}
\alias{tidy_genind}
\title{Tidy a genind object to a tidy dataframe}
\usage{
tidy_genind(
  data,
  keep.allele.names = FALSE,
  tidy = TRUE,
  gds = TRUE,
  write = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{(path or object) A genind object in the global environment or
path to a genind file that will be open with \code{readRDS}.}

\item{keep.allele.names}{Allows to keep allele names for the tidy dataset.
Requires the alleles to be numeric. To have this argument in
\pkg{radiator} \code{\link{tidy_genomic_data}} or
\pkg{radiator} \code{\link{genomic_converter}}
use it at the end. \code{...} in those function looks for it.
Default: \code{keep.allele.names = FALSE}.}

\item{tidy}{(logical) Generate a tidy dataset.
Default: \code{tidy = TRUE}.}

\item{gds}{(optional, logical) To write a radiator gds object.
Currently, for biallelic datasets only.
Default: \code{gds = TRUE}.}

\item{write}{(optional, logical) To write in the working directory the tidy
data. The file is written with \code{radiator_genind_DATE@TIME.rad}.
Default: \code{write = FALSE}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\description{
Tidy genind object from
\href{https://github.com/thibautjombart/adegenet}{adegenet} to a tidy dataframe.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\references{
Jombart T (2008) adegenet: a R package for the multivariate
analysis of genetic markers. Bioinformatics, 24, 1403-1405.

Jombart T, Ahmed I (2011) adegenet 1.3-1:
new tools for the analysis of genome-wide SNP data.
Bioinformatics, 27, 3070-3071.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
