% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genlight.R
\name{tidy_genlight}
\alias{tidy_genlight}
\title{Tidy a genlight object to a tidy dataframe and/or GDS object/file}
\usage{
tidy_genlight(
  data,
  tidy = TRUE,
  gds = TRUE,
  write = FALSE,
  verbose = FALSE,
  parallel.core = parallel::detectCores() - 1
)
}
\arguments{
\item{data}{(path or object) A genlight object in the global environment or
path to a genlight file that will be open with \code{readRDS}.}

\item{tidy}{(logical) Generate a tidy dataset.
Default: \code{tidy = TRUE}.}

\item{gds}{(optional, logical) To write a radiator gds object.
Default: \code{gds = TRUE}.}

\item{write}{(optional, logical) To write in the working directory the tidy
data. The file is written with \code{radiator_genlight_DATE@TIME.rad}.
Default: \code{write = FALSE}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\description{
Tidy genlight object to a tidy dataframe and/or GDS object/file.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
A string of the same dimension is generated when genlight:
\enumerate{
\item \code{is.null(genlight@pop)}: pop will be integrated
in the tidy dataset.
\item \code{is.null(data@chromosome)}: CHROM1 will be integrated
in the tidy dataset.
\item \code{is.null(data@loc.names)}: LOCUS1 to dim(genlight)[2]
will be integrated in the tidy dataset.
\item \code{is.null(data@position)}: an integer string of
length = dim(genlight)[2] will be integrated in the tidy dataset.
}

\strong{Note: that if all CHROM, LOCUS and POS is missing the function will be terminated}
}
\references{
Jombart T (2008) adegenet: a R package for the multivariate
analysis of genetic markers. Bioinformatics, 24, 1403-1405.

Jombart T, Ahmed I (2011) adegenet 1.3-1:
new tools for the analysis of genome-wide SNP data.
Bioinformatics, 27, 3070-3071.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
