% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{tidy_plink}
\alias{tidy_plink}
\title{Tidy plink \code{.tped} and {.tfam} files in a tidy data frame}
\usage{
tidy_plink(data, strata = NULL, verbose = FALSE, ...)
}
\arguments{
\item{data}{The plink file ending with \code{.tped}, the corresponding
\code{.tfam} must be in the same directory.}

\item{strata}{(path or object) The strata file or object.
Additional documentation is available in \code{\link{read_strata}}.
Use that function to whitelist/blacklist populations/individuals.
Option to set \code{pop.levels/pop.labels} is also available.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function. Also used for legacy arguments (see details or
special section)}
}
\description{
Transform genomic data set produced by massive parallel
sequencing pipeline (e.g.GBS/RADseq,
SNP chip, DArT, etc) into a tidy format.
The use of blacklist and whitelist along
several filtering options are available to prune the dataset.
Several arguments are available to make your data population-wise and easily
rename the pop id.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
}
\references{
PLINK: a tool set for whole-genome association and population-based linkage
analyses.
American Journal of Human Genetics. 2007: 81: 559–575. doi:10.1086/519795
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
