% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bayescan.R
\name{write_bayescan}
\alias{write_bayescan}
\title{Write a \href{http://cmpg.unibe.ch/software/BayeScan/}{BayeScan}
file from a tidy data frame}
\usage{
write_bayescan(data, pop.select = NULL, filename = NULL,
  parallel.core = parallel::detectCores() - 1, ...)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{pop.select}{(optional, string) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops).
Default: \code{pop.select = NULL}}

\item{filename}{(optional) The file name prefix for the bayescan file
written to the working directory. With default: \code{filename = NULL},
the date and time is appended to \code{radiator_bayescan_}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{...}{(optional) To pass further arguments for fine-tuning the function.}
}
\value{
A bayescan file is written in the working directory.
}
\description{
Write a \href{http://cmpg.unibe.ch/software/BayeScan/}{BayeScan}
file from a tidy data frame. The data is bi- or multi-allelic.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
\emph{Integrated filters:}
\enumerate{
\item by defaults only markers found in common between populations are used
\item by defaults monomorphic markers are automatically removed before
generating the dataset.
}
}
\references{
Foll, M and OE Gaggiotti (2008) A genome scan method to identify
selected loci appropriate
for both dominant and codominant markers: A Bayesian perspective.
Genetics 180: 977-993

Foll M, Fischer MC, Heckel G and L Excoffier (2010)
Estimating population structure from
AFLP amplification intensity. Molecular Ecology 19: 4638-4647

Fischer MC, Foll M, Excoffier L and G Heckel (2011) Enhanced AFLP
genome scans detect
local adaptation in high-altitude populations of a small rodent (Microtus arvalis).
Molecular Ecology 20: 1450-1462
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
