% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_colony.R
\name{write_colony}
\alias{write_colony}
\title{Write a \code{COLONY} input file}
\usage{
write_colony(
  data,
  strata = NULL,
  sample.markers = NULL,
  pop.select = NULL,
  allele.freq = NULL,
  inbreeding = 0,
  mating.sys.males = 0,
  mating.sys.females = 0,
  clone = 0,
  run.length = 2,
  analysis = 1,
  allelic.dropout = 0,
  error.rate = 0.02,
  print.all.colony.opt = FALSE,
  random.seed = NULL,
  verbose = FALSE,
  parallel.core = parallel::detectCores() - 1,
  filename = NULL,
  ...
)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{strata}{(path or object) The strata file or object.
Additional documentation is available in \code{\link{read_strata}}.
Use that function to whitelist/blacklist populations/individuals.
Option to set \code{pop.levels/pop.labels} is also available.}

\item{sample.markers}{(number) \code{COLONY} can take a long time to run,
use a random subsample of your markers to speed test \code{COLONY}
e.g. \code{sample.markers = 500} to use only 500 randomly chosen markers.
Default: \code{sample.markers = NULL}, will use all markers.}

\item{pop.select}{(optional, string) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops).
Default: \code{pop.select = NULL}}

\item{allele.freq}{(optional, string) Allele frequency can be computed from
a select group.
e.g. \code{allele.freq = "QUE"} or \code{allele.freq = c("QUE", "ONT")}.
Using \code{allele.freq = "overall"} will use all the samples to compute the
allele frequency.
Default: \code{allele.freq = NULL}, will not compute allele frequency.}

\item{inbreeding}{(boolean) 0/1 no inbreeding/inbreeding.
Default: \code{inbreeding = 0}}

\item{mating.sys.males}{(boolean) Mating system in males.
0/1 polygyny/monogyny.
Default: \code{mating.sys.males = 0}.}

\item{mating.sys.females}{(boolean) Mating system in females.
0/1 polygyny/monogyny.
Default: \code{mating.sys.females = 0}.}

\item{clone}{(boolean) Should clones and duplicated individuals be inferred.
0/1, yes/no. Default: \code{clone = 0}.}

\item{run.length}{(integer) Length of run. 1 (short), 2 (medium), 3 (long),
4 (very long). Start with short or medium run and consider longer run if your
estimates probability are not stable or really good.
Default: \code{run.length = 2}.}

\item{analysis}{(integer) Analysis method.
0 (Pairwise-Likelihood Score), 1 (Full Likelihood),
2 (combined Pairwise-Likelihood Score and Full Likelihood).
Default: \code{analysis = 1}.}

\item{allelic.dropout}{Locus allelic dropout rate.
Default : \code{allelic.dropout = 0}.}

\item{error.rate}{Locus error rate.
Default:\code{error.rate = 0.02}.}

\item{print.all.colony.opt}{(logical) Should all \code{COLONY} options be printed in the file.

\strong{This require manual curation, for the file to work directly with \code{COLONY}}.
Default = \code{print.all.colony.opt = FALSE}.}

\item{random.seed}{(integer, optional) For reproducibility, set an integer
that will be used inside the function that requires randomness. With default,
a random number is generated and printed in the appropriate output.
Default: \code{random.seed = NULL}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{filename}{Name of the acronym for filenaming in the working directory.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function. Also used for legacy arguments (see details or
special section)}
}
\value{
A \code{COLONY} file in your working directory (2 if you selected imputations arguments...)
}
\description{
Write a \code{COLONY} input file.
}
\details{
\strong{It is highly recommended to read (twice!) the user guide distributed with
\code{COLONY} to find out the details for input and output of the software.}

Not all options are provided here.

But to ease the process, all the required options to properly run \code{COLONY}
will be printed in the file written in your working directory.
Change the values accordingly and wisely.
}
\examples{
\dontrun{
# Simplest way to run the function with a tidy dataset:
colony.file <- radiator::write_colony(data = "turtle.data.rad")
}
}
\references{
Jones OR, Wang J (2010) COLONY: a program for parentage and
sibship inference from multilocus genotype data.
Molecular Ecology Resources, 10, 551–555.

Wang J (2012) Computationally Efficient Sibship and
Parentage Assignment from Multilocus Marker Data. Genetics, 191, 183–194.
}
\seealso{
\code{COLONY} is available on Jinliang Wang web site
\url{https://www.zsl.org/science/software/colony}


\href{https://thierrygosselin.github.io/radiator/articles/rad_genomics_computer_setup.html#colony}{colony installation instructions}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
