% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_faststructure.R
\name{write_faststructure}
\alias{write_faststructure}
\title{Write a faststructure file from a tidy data frame}
\usage{
write_faststructure(
  data,
  plink.bed = FALSE,
  pop.levels = NULL,
  filename = NULL,
  ...
)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}

\item{plink.bed}{To export in plink BED format. This will write 3 files:
\code{.bed}, \code{.bim}, \code{.fam}. For this option to work, the argument
data must be a GDS file or object.
Default: \code{plink.bed = FALSE}.}

\item{pop.levels}{(optional, string) This refers to the levels in a factor. In this
case, the id of the pop.
Use this argument to have the pop ordered your way instead of the default
alphabetical or numerical order. e.g. \code{pop.levels = c("QUE", "ONT", "ALB")}
instead of the default \code{pop.levels = c("ALB", "ONT", "QUE")}.
White spaces in population names are replaced by underscore.
Default: \code{pop.levels = NULL}.}

\item{filename}{(optional) The file name prefix for the faststructure file
written to the working directory. With default: \code{filename = NULL},
the date and time is appended to \code{radiator_faststructure_}.}

\item{...}{other parameters passed to the function.}
}
\value{
A faststructure file is saved to the working directory.
}
\description{
Write a faststructure file from a tidy data frame.
For biallelic dataset only.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
}
\references{
Raj A, Stephens M, Pritchard JK (2014)
fastSTRUCTURE: Variational Inference of Population Structure in Large SNP
Datasets. Genetics, 197, 573-589.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
