% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{write_gds}
\alias{write_gds}
\title{Write a GDS object from a tidy data frame}
\usage{
write_gds(
  data,
  data.source = NULL,
  filename = NULL,
  open = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{data.source}{(optional, character) The name of the software that
generated the data. e.g. \code{data.source = "Stacks v.2.4"}.
Default: \code{data.source = NULL}.}

\item{filename}{(optional) The file name of the Genomic Data Structure (GDS) file.
radiator will append \code{.gds.rad} to the filename.
If filename chosen is already present in the
working directory, the default \code{radiator_datetime.gds.rad} is chosen.
Default: \code{filename = NULL}.}

\item{open}{(optional, logical) Open or not the radiator connection.
Default: \code{open = TRUE}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}
}
\value{
An object in the global environment of class \code{SeqVarGDSClass} and
a file in the working directory.
}
\description{
Write a Genomic Data Structure (GDS) file format
\href{https://github.com/zhengxwen/gdsfmt}{gdsfmt}
and object of class \code{SeqVarGDSClass} from a tidy data frame.

Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\examples{
\dontrun{
data.gds <- radiator::write_gds(data = "shark.rad")
}
}
\references{
Zheng X, Levine D, Shen J, Gogarten SM, Laurie C, Weir BS.
A high-performance computing toolset for relatedness and principal component
analysis of SNP data. Bioinformatics. 2012;28: 3326-3328.
doi:10.1093/bioinformatics/bts606

Zheng X, Gogarten S, Lawrence M, Stilp A, Conomos M, Weir BS,
Laurie C, Levine D (2017). SeqArray -- A storage-efficient high-performance
data format for WGS variant calls.
Bioinformatics.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
