% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_genepopedit.R
\name{write_genepopedit}
\alias{write_genepopedit}
\title{Write a genepopedit flatten object}
\usage{
write_genepopedit(data)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}
}
\value{
A genepopedit object in the global environment.
}
\description{
Write a genepopedit object from a tidy data frame or GDS file/object.
}
\references{
Stanley RRE, Jeffery NW, Wringe BF, DiBacco C, Bradbury IR (2017)
genepopedit: a simple and flexible tool for manipulating multilocus molecular
data in R. Molecular Ecology Resources, 17, 12-18.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
