% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtypes.R
\name{write_gtypes}
\alias{write_gtypes}
\title{Write a \href{https://github.com/EricArcher/}{strataG} object from
radiator GDS and tidy data file or object.}
\usage{
write_gtypes(data)
}
\arguments{
\item{data}{(4 options) A file or object generated by radiator:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Look into \code{\link{tidy_genomic_data}},
\code{\link{read_vcf}} or
\code{\link{tidy_vcf}}.}
}
\value{
An object of the class \href{https://github.com/EricArcher/}{strataG} is returned.
}
\description{
Write a\href{https://github.com/EricArcher/}{strataG} object from a tidy data frame.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\examples{
\dontrun{
require(strataG)
# with radiator GDS
turtle <- radiator::write_gtypes(data = "my.radiator.gds.rad")

# with tidy data
turtle <- radiator::write_gtypes(data = "my.radiator.rad")
}
}
\references{
Archer FI, Adams PE, Schneiders BB.
strataG: An r package for manipulating, summarizing and analysing population
genetic data.
Molecular Ecology Resources. 2017; 17: 5-11. doi:10.1111/1755-0998.12559
}
\seealso{
\href{https://github.com/EricArcher/}{strataG}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
