% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pcadapt.R
\name{write_pcadapt}
\alias{write_pcadapt}
\title{Write a \href{https://github.com/bcm-uga/pcadapt}{pcadapt}
file from a tidy data frame}
\usage{
write_pcadapt(
  data,
  pop.select = NULL,
  filename = NULL,
  parallel.core = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}

\item{pop.select}{(optional, string) Selected list of populations for
the analysis. e.g. \code{pop.select = c("QUE", "ONT")} to select \code{QUE}
and \code{ONT} population samples (out of 20 pops). If \code{pop.labels}
argument was used to rename the strata column, use the new names with
\code{pop.select}.
Default: \code{pop.select = NULL}}

\item{filename}{(optional) The file name prefix for the pcadapt file
written to the working directory. With default: \code{filename = NULL},
the date and time is appended to \code{radiator_pcadapt_}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{...}{(optional) To pass further arguments for fine-tuning the function.}
}
\value{
A pcadapt file is written in the working directory a genotype matrix
object is also generated in the global environment.
}
\description{
Write a
\href{https://github.com/bcm-uga/pcadapt}{pcadapt}
file from a tidy data frame. The data is biallelic.
Used internally in
\href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
\emph{Integrated filters:}
\enumerate{
\item by defaults only markers found in common between populations are used
(See advance section).
\item by defaults monomorphic markers are automatically removed before
generating the pcadapt file.
}
}
\section{Advance mode}{


\emph{dots-dots-dots ...} allows to pass several arguments for fine-tuning the function:
\enumerate{
\item \strong{Filtering for linkage disequilibrium}: 3 arguments
\code{filter.long.ld, long.ld.missing, ld.method}
described in \code{\link{filter_ld}} are available.
Reducing linkage before running genome scan is essential. At least start by
removing SNPs on the same RADseq locus (short linkage disequilibrium).

\item \strong{Filtering markers with low Minor Allele Count, Frequency or Depth}
Use the 2 arguments provided by the function \code{\link{filter_ma}} (read doc):
\code{filter.ma} and \code{ma.stats} to evaluate the impact of MAC/MAF/MAD on genome scans.
 is called.

\item Turning off the filter that keeps markers in common between strata:
This is not recommended, but users who wants to explore the impact of such filtering
and know the biais it can potentially generate can use the argument
\code{filter.common.markers}.
The function \code{\link{filter_common_markers}} is called.
Default: \code{filter.common.markers = NULL}
}
}

\references{
Luu, K., Bazin, E., & Blum, M. G. (2017).
pcadapt: an R package to perform genome scans for selection based on principal component analysis.
Molecular Ecology Resources, 17(1), 67-77.

Duforet-Frebourg, N., Luu, K., Laval, G., Bazin, E., & Blum, M. G. (2015).
Detecting genomic signatures of natural selection with principal component analysis: application to the 1000 Genomes data.
Molecular biology and evolution, msv334.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
