% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rad.R
\name{write_rad}
\alias{write_rad}
\title{Write tidy genomic data file or close GDS file}
\usage{
write_rad(
  data,
  filename,
  internal = FALSE,
  write.message = "standard",
  verbose = FALSE
)
}
\arguments{
\item{data}{An object in the global environment: tidy genomic dataset
or GDS connection file}

\item{filename}{(optional) Name of the file. If default, \code{radiator_date_time.arrow.parquet} is used.
Default: \code{filename = NULL}.}

\item{internal}{(optional, logical) This is used inside radiator internal code and it stops
from writing the file.
Default: \code{internal = FALSE}.}

\item{write.message}{(optional, character) Print a message in the console
after writing file.
With \code{write.message = NULL}, nothing is printed in the console.
Default: \code{write.message = "standard"}. This will print
\code{message("File written: ", basename(filename))}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty
during execution.
Default: \code{verbose = FALSE}.}
}
\value{
A file written in the working directory or nothing if it's a GDS connection file.
}
\description{
For tidy genomic datasets, the function provides a
fast way to write a \code{.arrow.parquet} file from Apache.
This new file ending replaces \code{.rad} file format that was essentially
the \code{.fst}, provided by
the package \href{https://github.com/fstpackage/fst}{fst}.
See explanation in section below.


When the object is a CoreArray Genomic Data Structure
(\href{https://github.com/zhengxwen/gdsfmt}{GDS}) file system, the function
\strong{close the connection with the GDS file}. Before doing so it sets the
filters (variants and samples) based on the info found in the file.

Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
}
\section{Why is .rad is no longer supported}{


Originally, the ending \code{.fst} from
package \href{https://github.com/fstpackage/fst}{fst} was replaced
by \code{.rad} to remove the confusion with population genetics statistic fst ...
The decision to stop using the package \code{fst} was taken because:
\itemize{
\item The package was \strong{always difficult} to install when you
wanted all cores to function.
\item Successful recipe to install with one OS was rarely working
after changing R version or OS version = painful.
\item Asking users to play with \code{.R/Makevars} was always time
consuming to troubleshoot after for me.
\item \href{https://arrow.apache.org/docs/r/}{arrow} is easy to install,
files are smaller in size and read/write faster!
}
}

\examples{
\dontrun{
require(SeqArray)
radiator::write_rad(data = tidy.data, filename = "data.shark.arrow.parquet")
radiator::write_rad(data = gds.object)
}
}
\seealso{
\href{https://arrow.apache.org}{appache arrow}

\href{https://github.com/fstpackage/fst}{fst}

\href{https://github.com/zhengxwen/gdsfmt}{GDS}

\code{\link{read_rad}}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
