% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rad.R
\name{write_rad}
\alias{write_rad}
\title{Write tidy genomic data file ending .rad}
\usage{
write_rad(data, path)
}
\arguments{
\item{data}{A file in the working directory ending with .rad,
a tidy genomic data produced by radiator, assigner or grur.}

\item{path}{Path to write the data on disk}
}
\value{
A tidy data frame in the global environment.
}
\description{
Fast write .rad file.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
The function uses \code{\link[fst]{write.fst}} with a compression level of 85,
that work well with RADseq dataset.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
