% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rad.R
\name{write_radiator_tsv}
\alias{write_radiator_tsv}
\title{Write radiator tsv}
\usage{
write_radiator_tsv(
  data,
  path.folder = NULL,
  filename,
  append = FALSE,
  col.names = TRUE,
  date = TRUE,
  internal = FALSE,
  write.message = "standard",
  verbose = FALSE
)
}
\arguments{
\item{data}{An object in the global environment. Ideally a tibble...}

\item{path.folder}{(optional) Path to write the file.
With default, the working directory is used.
Default: \code{path.folder = NULL}.}

\item{filename}{(required, character string) Name of the file.}

\item{append}{If FALSE, will overwrite existing file.
If TRUE, will append to existing file.
In both cases, if the file does not exist a new file is created.
Default: \code{append = FALSE}.}

\item{col.names}{If FALSE, column names will not be included at the top of the file.
If TRUE, column names will be included.
If not specified, col_names will take the opposite value given to append.
Default: \code{col.names = TRUE}.}

\item{date}{(logical) To append time to the filename.
Default: \code{date = TRUE}.}

\item{internal}{(optional, logical) This is used inside radiator internal code and it stops
from writting the file.
Default: \code{internal = FALSE}.}

\item{write.message}{(optional, character) Print a message in the console
after writing file.
With \code{write.message = NULL}, nothing is printed in the console.
Default: \code{write.message = "standard"}. This will print
\code{message("File written: ", basename(filename))}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty
during execution.
Default: \code{verbose = FALSE}.}
}
\value{
A file written in the working directory.
}
\description{
Write tsv file. Using the package \code{readr}. Takes cares of
the file path and proper filenaming. The complete filename is built with the
\code{path.folder} + \code{filename} + \code{datetime} + \code{.tsv}.

Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and \href{https://github.com/thierrygosselin/assigner}{assigner}
and might be of interest for users.
}
\examples{
\dontrun{
radiator::write_radiator_tsv(data = tidy.data, filename = "data.shark")
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}

\code{\link{read_rad}}
}
