% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_seqarray.R
\name{write_seqarray}
\alias{write_seqarray}
\title{Write a SeqArray GDS file from a vcf file and generate a connection object.}
\usage{
write_seqarray(vcf, strata = NULL, filename = NULL,
  vcf.stats = FALSE, parallel.core = parallel::detectCores() - 1,
  verbose = TRUE, ...)
}
\arguments{
\item{vcf}{The path to the vcf file.}

\item{strata}{(optional)
The strata file is a tab delimited file with a minimum of 2 columns headers:
\code{INDIVIDUALS} and \code{STRATA}.
If a \code{strata} file is specified with all file formats that don't
require it, the strata argument will have precedence on the population
groupings used internally in those file formats. For file formats without
population/strata groupings (e.g. vcf, haplotype files) if no strata file is
provided, 1 pop/strata grouping will automatically be created.
For vcf and haplotypes file, the strata can also be used as a whitelist of id.
Samples not in the strata file will be discarded from the data set.
The \code{STRATA} column can be any hierarchical grouping.
To create a strata file see \code{\link[radiator]{individuals2strata}}.
If you have already run
\href{http://catchenlab.life.illinois.edu/stacks/}{stacks} on your data,
the strata file is similar to a stacks \emph{population map file},
make sure you
have the required column names (\code{INDIVIDUALS} and \code{STRATA}).
The strata column is cleaned of a white spaces that interfere with some
packages or codes: space is changed to an underscore \code{_}.
Default: \code{strata = NULL}.}

\item{filename}{(optional) The file name of the Genomic Data Structure (GDS) file.
radiator will append \code{.gds} to the filename.
If filename chosen is already present in the
working directory, the default \code{radiator_datetime.gds} is chosen.
Default: \code{filename = NULL}.}

\item{vcf.stats}{(logical) When \code{vcf.stats = TRUE}, individual's missing
genotype proportion, averaged heterozygosity, total coverage, mean genotype
coverage and marker's metadata along count for ref and alt alleles and mean
coverage is generated and written in the working directory.
Default: \code{vcf.stats = FALSE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during import.
Default: \code{parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function (see details).}
}
\description{
Write a SeqArray \href{http://zhengxwen.github.io/SeqArray/}{SeqArray}
file (Zheng et al. 2017) from a vcf file and generate a connection object.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\details{
A vcf file of 35 GB with ~4 millions SNPs take about ~7 min with 8 CPU.
A vcf file of 21 GB with ~2 millions SNPs take about ~5 min with 7 CPU.

After the file is generated, it's a matter of sec to open a connection.

\strong{Advance mode, using \emph{dots-dots-dots}}
\enumerate{
\item \code{whitelist.markers}: detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{blacklist.id}: detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{pop.select}: detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{pop.levels}: detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{pop.labels}: detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{keep.both.strands}: (logical) Radiator removes
duplicate SNPs found on different strands,
by default (\code{keep.both.strands = FALSE}).
\item \code{common.markers}: (logical) Default: code{common.markers = TRUE}.
Detailed in \code{\link[radiator]{tidy_genomic_data}}.
\item \code{filter.mac}: (integer) To blacklist markers below a specific MAC
(calculated overall).
\item \code{filter.coverage.outliers}: (logical) To blacklist markers with low and
high coverage based on outlier statistics.
\item \code{filter.markers.missing}: (integer) To blacklist markers with too
many missing data. e.g. \code{filter.markers.missing = 10}, will only keep
markers with missing rate <= to 10%.
\item \code{filter.snp.read.position}: 3 options available, \code{"outliers", "q75", "iqr"}.
This argument will blacklist markers based on it's position on the read.
\code{filter.snp.read.position = "outliers"}, will remove markers based
on outlier statistics of the position on the reads. e.g. if majority of SNPs
are found between 10 and 90 pb, and very few above and below, those markers are
discarded. Use this function argument with dataset with problematic assembly,
or \emph{de novo} assembly with undocumented or poorly selected
mismatch threshold.
\item \code{filter.short.ld}: this is the \code{snp.ld} argument
in \code{\link[radiator]{snp_ld}}
\item \code{filter.long.ld}: this is the \code{ld.threshold} argument
in \code{\link[radiator]{snp_ld}}
\item \code{long.ld.missing}: this is the \code{long.ld.missing} argument
in \code{\link[radiator]{snp_ld}}
\item \code{filter.individuals.missing}: (double) Use this argument to
blacklist individuals with too many missing data.
e.g. \code{filter.individuals.missing = 0.7}, will remove individuals with >
0.3 or 30% missing genotypes. This can help discover more polymorphic markers
with some dataset.
\item \code{path.folder}: to write ouput in a specific path
(used internally in radiator).
}
}
\references{
Zheng X, Gogarten S, Lawrence M, Stilp A, Conomos M, Weir BS,
Laurie C, Levine D (2017). SeqArray -- A storage-efficient high-performance
data format for WGS variant calls.
Bioinformatics.

Danecek P, Auton A, Abecasis G et al. (2011)
The variant call format and VCFtools.
Bioinformatics, 27, 2156-2158.
}
\seealso{
\code{\link{snp_ld}}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
