% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stockr.R
\name{write_stockr}
\alias{write_stockr}
\title{Write a stockR dataset from a tidy data frame}
\usage{
write_stockr(data)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{radiator} \code{\link{tidy_genomic_data}}.}
}
\value{
The object generated is a matrix with
dimension: MARKERS x INDIVIDUALS. The genotypes are coded like PLINK:
0, 1 or 2 alternate allele. 0: homozygote for the reference allele,
1: heterozygote, 2: homozygote for the alternate allele.
Missing genotypes have NA. The object also as 2 attributes.
\code{attributes(data)$grps} with \code{STRATA/POP_ID} of the individuals and
\code{attributes(data)$sample.grps} filled with \code{INDIVIDUALS}.
Both attributes can be used inside \emph{stockR}.
}
\description{
Write a stockR dataset (Fost et al. submitted) from a tidy data frame.
Used internally in \href{https://github.com/thierrygosselin/radiator}{radiator}
and might be of interest for users.
}
\references{
Foster et al. submitted
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
